Equations to Multivariate Normal Sample Data

Generates data from a multivariate normal distribution from model equations using the MASS::mvrnorm() function.

Eq2Data(eq, n, ...)

Arguments

eq
Character string. Equations. See Details.

n
Integer. Sample size.

...
Additional arguments to pass to MASS::mvrnorm().

Value

\(n\) variates from \( \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right), \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right) . \)

Details

The argument eq is a character string that specifies the associations between the variables. See Syntax, Operations, par.label, Line breaks, and Comments below.

The multivariate normal distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbf{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{g} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} = \mathbf{F} \mathbf{E} \mathbf{u} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{M} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{C} \mathbf{F}^{\mathsf{T}} $$

Syntax

Each line should follow the syntax below

lhs <space> op <space> rhs <space> par.label <\n> or <;>

lhs

is the variable on the left-hand side,

rhs

is the variable on the right-hand side,

op

is the operation between lhs and rhs,

par.label

is the column of parameter label,

\n or ;

are line breaks. Each line should end with a line break.

Operations

The associations are defined by the following operations

by

left-hand side measured by right-hand side,

on

left-hand side regressed on right-hand side,

with

left-hand side covarying with right-hand side,

on 1

left-hand side regressed on 1 for mean structure.

par.label

Each parameter should have a numeric value.

Line breaks

The characters \n and ; can be used as line breaks. Each line should end with a line break.

Comments

Comments can be written after a hash (#) sign.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

eq <- "
  # lhs op   rhs value
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2Data(eq, n = 100)

#>                  y            x
#>   [1,]  2.02009291  1.032714988
#>   [2,]  0.13683836  0.766950402
#>   [3,]  3.10783050  1.590129689
#>   [4,]  0.42632682  0.839828479
#>   [5,] -0.36844920  1.083603755
#>   [6,] -0.84810906  0.491825455
#>   [7,]  2.47254297  1.216455212
#>   [8,]  0.72151671  0.793269849
#>   [9,]  0.79472512  0.467588726
#>  [10,]  0.91303960  0.158151791
#>  [11,]  1.09558031  0.734391594
#>  [12,] -0.34089059  0.929867105
#>  [13,] -1.77348039 -0.152786394
#>  [14,]  2.27177420  1.116639467
#>  [15,] -0.01350512  0.123460639
#>  [16,]  2.78804171  0.085118021
#>  [17,]  1.12392769  0.456516022
#>  [18,]  0.38946161  1.230194954
#>  [19,] -0.07959395  0.251188050
#>  [20,]  1.62555636  0.283973425
#>  [21,]  0.18346671 -0.490457228
#>  [22,]  0.14097556  0.213428506
#>  [23,]  2.09237666  0.250945693
#>  [24,] -0.27869629  0.124634428
#>  [25,] -1.59207910 -0.042181103
#>  [26,]  0.61074524  0.516132445
#>  [27,]  1.60268733  0.489270938
#>  [28,]  0.49138932  0.615904748
#>  [29,]  1.36760230  0.942018472
#>  [30,]  2.20337466  0.815831745
#>  [31,] -0.65662721  0.741633072
#>  [32,]  0.38843881  0.093703389
#>  [33,]  0.23685736  0.401186050
#>  [34,] -1.43888945  0.628324340
#>  [35,]  0.28036920  0.872444151
#>  [36,]  0.69482636  0.341980177
#>  [37,]  2.54104644  1.365835554
#>  [38,]  0.78212932  0.695425678
#>  [39,]  0.93888648  0.201744236
#>  [40,] -0.68170704  0.190815233
#>  [41,]  0.40568142  0.550780629
#>  [42,]  1.12860762  1.019879540
#>  [43,]  0.74404861 -0.284991461
#>  [44,]  0.85698590  0.240119848
#>  [45,]  0.10310438 -0.033722319
#>  [46,] -2.43548820 -0.786491408
#>  [47,]  0.33576119  0.964009964
#>  [48,] -0.09161974  0.129267808
#>  [49,]  0.21686584  1.110789136
#>  [50,]  2.78262804  0.350611580
#>  [51,] -0.41685856  0.090859879
#>  [52,]  0.78310797  0.512004877
#>  [53,]  0.61521852  1.038198809
#>  [54,]  0.42261753  0.732296882
#>  [55,]  0.57276449  0.044558117
#>  [56,] -0.04419866  0.067233208
#>  [57,]  3.03612968  1.081411631
#>  [58,] -2.52285403 -0.053573913
#>  [59,]  0.98082780  0.261989142
#>  [60,]  0.34319240  0.101779426
#>  [61,] -0.88707867  1.025735610
#>  [62,]  1.33454473  0.276815963
#>  [63,] -0.77143478  0.338627734
#>  [64,]  0.80022307  0.452748756
#>  [65,]  1.63145037  1.571125039
#>  [66,]  1.50747994  1.091784027
#>  [67,] -0.79972060  0.701936015
#>  [68,]  0.38989539  0.310717650
#>  [69,] -0.07280095  0.493151709
#>  [70,]  2.44434295  0.726549185
#>  [71,]  2.12735633  0.930375589
#>  [72,]  0.08579180  0.839603205
#>  [73,] -0.86154541 -0.391984684
#>  [74,]  0.33291112  0.029100496
#>  [75,] -0.89075724 -0.173490651
#>  [76,]  0.02063124 -0.078325512
#>  [77,] -0.22320485  0.542063049
#>  [78,]  0.49436427  1.071330623
#>  [79,]  0.27758678  0.584727594
#>  [80,]  0.82265519 -0.416977818
#>  [81,] -1.74737725  0.437204443
#>  [82,] -1.06089912  0.646342985
#>  [83,] -0.50200375  1.015727830
#>  [84,]  0.03384806 -0.295232693
#>  [85,]  1.16613682  0.408408772
#>  [86,] -0.53647913 -0.639641801
#>  [87,]  3.39340243  1.252244183
#>  [88,]  0.93084880 -0.549781482
#>  [89,]  0.12348739 -0.065749896
#>  [90,]  1.04004190  0.114384630
#>  [91,]  0.14050003 -0.089501513
#>  [92,] -0.72364269  0.389401923
#>  [93,] -0.75041878  0.574483456
#>  [94,]  0.48430488  0.756230276
#>  [95,]  0.13575974 -0.377976236
#>  [96,] -1.34942899  0.003691844
#>  [97,]  0.93238964  0.433014615
#>  [98,] -0.43944111 -0.233083556
#>  [99,] -0.84167780 -0.374089012
#> [100,]  0.89177095  0.317638616