Equations to Expectations

Converts equations to expectations.

Eq2Expectations(
  eq,
  par = FALSE,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE
)

Arguments

eq
Character string. Equations. See Details.

par
Logical. If par = TRUE, use par.index as labels. If par = FALSE, use label as labels.

str
Logical. If TRUE, returns symbolic result as string. If FALSE, returns symbolic result as an R expression.

ysym
Logical. If TRUE, returns symbolic result as yac_symbol if str = TRUE. If FALSE, returns symbolic result as string if str = TRUE.

simplify
Logical. Simplify symbolic results.

tex
Logical. Return results as latex if str = TRUE.

Details

The input is a character string that specifies the associations between the variables.

Syntax

Each line should follow the syntax below

lhs operation rhs label

The associations are defined by the following operations

by

left-hand side measured by right-hand side

on

left-hand side regressed on right-hand side

with

left-hand side covarying with right-hand side

on 1

left-hand side regressed on 1 for mean structure

label

Each parameter should be labeled. The label should be a number for fixed parameters and a character string for free parameters. Equality contraints can be imposed by using the same label.

Comments

Comments can be written after a hash (#) sign.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

Author

Ivan Jacob Agaloos Pesigan

Examples

# Numerical ---------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"

# Symbolic ----------------------------------------------------------
eq <- "
  # lhs op   rhs label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2Expectations(eq, par = FALSE)

#> $variables
#> [1] "y" "x" "e"
#> 
#> $A
#> {{   0, beta,    1},
#>  {   0,    0,    0},
#>  {   0,    0,    0}} 
#> 
#> $S
#> {{      0,       0,       0},
#>  {      0, sigmax2,       0},
#>  {      0,       0, sigmae2}} 
#> 
#> $u
#> {{alpha},
#>  {  mux},
#>  {    0}} 
#> 
#> $Filter
#> {{1, 0, 0},
#>  {0, 1, 0}} 
#> 
#> $v
#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 
#> 
#> $g
#> {{alpha+beta*mux},
#>  {           mux}} 
#> 
#> $C
#> {{sigmax2*beta^2+sigmae2,           beta*sigmax2,                sigmae2},
#>  {          sigmax2*beta,                sigmax2,                      0},
#>  {               sigmae2,                      0,                sigmae2}} 
#> 
#> $M
#> {{sigmax2*beta^2+sigmae2,           beta*sigmax2},
#>  {          sigmax2*beta,                sigmax2}} 
#> 

# Expressions using `label`

beta <- 1
sigmae2 <- 1
sigmax2 <- 0.25
alpha <- 0
mux <- 0.50
Exp <- Eq2Expectations(eq, par = FALSE, str = FALSE)
eval(Exp$M)

#>      [,1] [,2]
#> [1,] 1.25 0.25
#> [2,] 0.25 0.25

eval(Exp$g)

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5

# Expressions using `par.index`

p <- c(beta, sigmae2, sigmax2, alpha, mux)
p1 <- p[1]
p2 <- p[2]
p3 <- p[3]
p4 <- p[4]
p5 <- p[5]
Exp <- Eq2Expectations(eq, par = TRUE, str = FALSE)
eval(Exp$M)

#>      [,1] [,2]
#> [1,] 1.25 0.25
#> [2,] 0.25 0.25

eval(Exp$g)

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5