\(\mathbf{I} - \mathbf{A}\)

Derives \(\mathbf{I} - \mathbf{A}\)

IminusA(A, ...)

# S3 method for default
IminusA(A, ...)

# S3 method for yac_symbol
IminusA(
  A,
  exe = TRUE,
  str = TRUE,
  ysym = TRUE,
  simplify = FALSE,
  tex = FALSE,
  ...
)

Arguments

A
t by t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

...
...

exe
Logical. If exe = TRUE, executes the resulting yacas expression. If exe = FALSE, returns the resulting yacas expression as a character string. If exe = FALSE, the arguments str, ysym, simplify, and tex, are ignored.

str
Logical. If str = TRUE, returns symbolic result as string. If str = FALSE, returns symbolic result as an R expression.

ysym
Logical. If ysym = TRUE, returns symbolic result as yac_symbol if str = TRUE. If ysym = FALSE, returns symbolic result as string if str = TRUE.

simplify
Logical. Simplify symbolic results.

tex
Logical. If tex = TRUE returns results as latex math if str = TRUE. If tex = FALSE, returns symbolic result as string if str = TRUE.

Value

\(\mathbf{I} - \mathbf{A}\)

Details

$$ \mathbf{I} - \mathbf{A} $$

where

\(\mathbf{A}_{t \times t}\) represents asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings, and

\(\mathbf{I}_{t \times t}\) represents an identity matrix.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

# Numeric -----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
IminusA(A)

#>   y  x  e
#> y 1 -1 -1
#> x 0  1  0
#> e 0  0  1

# Symbolic ----------------------------------------------------------
A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
IminusA(Ryacas::ysym(A))

#> {{    1, -beta,    -1},
#>  {    0,     1,     0},
#>  {    0,     0,     1}} 

IminusA(Ryacas::ysym(A), ysym = FALSE)

#> [1] "{{1,-beta,-1},{0,1,0},{0,0,1}}"

IminusA(Ryacas::ysym(A), tex = TRUE)

#> [1] "\\left( \\begin{array}{ccc} 1 &  - \\beta  & -1 \\\\ 0 & 1 & 0 \\\\ 0 & 0 & 1 \\end{array} \\right)"

IminusA(Ryacas::ysym(A), str = FALSE)

#> expression(rbind(c(1, -beta, -1), c(0, 1, 0), c(0, 0, 1)))

beta <- 1
IminusA(Ryacas::ysym(A))

#> {{    1, -beta,    -1},
#>  {    0,     1,     0},
#>  {    0,     0,     1}} 

IminusA(Ryacas::ysym(A), ysym = FALSE)

#> [1] "{{1,-beta,-1},{0,1,0},{0,0,1}}"

IminusA(Ryacas::ysym(A), tex = TRUE)

#> [1] "\\left( \\begin{array}{ccc} 1 &  - \\beta  & -1 \\\\ 0 & 1 & 0 \\\\ 0 & 0 & 1 \\end{array} \\right)"

IminusA(Ryacas::ysym(A), str = FALSE)

#> expression(rbind(c(1, -beta, -1), c(0, 1, 0), c(0, 0, 1)))

eval(IminusA(Ryacas::ysym(A), str = FALSE))

#>      [,1] [,2] [,3]
#> [1,]    1   -1   -1
#> [2,]    0    1    0
#> [3,]    0    0    1