Generate Data from a Multivariate Normal Distribution Using the Reticular Action Model (RAM) Notation

Generates data from a multivariate normal distribution using the Reticular Action Model (RAM) notation with the MASS::mvrnorm() function.

RAM2Data(n, A, S, u, Filter, ...)

Arguments

n
Integer. Sample size.

A
t by t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
t by t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

u
vector of length t or t by 1 matrix. Mean structure parameters.

Filter
p by t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

...
Additional arguments to pass to MASS::mvrnorm().

Value

\(n\) variates from \( \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu} \left( \boldsymbol{\theta} \right), \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) \right) . \)

Details

The multivariate normal distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbf{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{g} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} = \mathbf{F} \mathbf{E} \mathbf{u} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{M} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{C} \mathbf{F}^{\mathsf{T}} $$

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# This is an example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

n <- 50
A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
u <- c(0.00, 0.50, 0.00)
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(Filter) <- c("y", "x", "e")
rownames(Filter) <- c("y", "x")
RAM2Data(n, A, S, u, Filter)

#>                 y            x
#>  [1,]  0.89476350  0.965980451
#>  [2,] -1.06545955  0.356572435
#>  [3,]  0.45905252  1.026311896
#>  [4,]  0.72757027  0.891266856
#>  [5,]  3.37059304  0.523941573
#>  [6,]  0.34388303  0.835211950
#>  [7,]  1.75230265  0.665511714
#>  [8,]  1.15685415  0.870619675
#>  [9,]  2.73718468  1.299712964
#> [10,] -2.30302577 -0.817334943
#> [11,]  1.79935664  0.739526276
#> [12,]  0.73539223  1.199169841
#> [13,] -0.37666211  0.297674627
#> [14,]  1.89342743  0.924243512
#> [15,] -0.19468521  0.742979163
#> [16,]  0.20530024  1.374356099
#> [17,]  0.06818925 -0.452108508
#> [18,] -0.05976718  0.802572479
#> [19,]  2.10685763  0.925399264
#> [20,] -0.52395511  0.150528427
#> [21,]  0.88352524  0.563192720
#> [22,]  0.36551210  1.446044271
#> [23,]  0.44339792  0.539560563
#> [24,]  0.37286377  0.419563705
#> [25,]  1.61256280  0.759099327
#> [26,] -0.68486382 -0.622867936
#> [27,]  0.33768324 -0.507659876
#> [28,]  1.09762134  0.322397585
#> [29,] -0.21219862 -0.012524283
#> [30,]  2.17157436  1.033000730
#> [31,]  0.33723375  0.007158959
#> [32,]  1.86956212  1.235989639
#> [33,] -0.27526198  0.064012056
#> [34,]  2.27226896  0.773572171
#> [35,]  0.45769712  0.325366597
#> [36,]  1.32557433  0.187564070
#> [37,]  1.34660751  1.122615897
#> [38,] -0.23002388  0.229849543
#> [39,] -0.64218735 -0.405193575
#> [40,]  2.33942164  1.106535421
#> [41,]  1.48781995  0.811362311
#> [42,] -0.98947110  0.247953062
#> [43,] -0.82644173  0.640176223
#> [44,] -0.79214763 -0.020893771
#> [45,]  2.25483509  0.193836593
#> [46,] -1.19201098  0.780858788
#> [47,]  0.35434233  1.101449242
#> [48,]  0.67476170  0.147558093
#> [49,]  0.07112890  0.118504353
#> [50,]  0.56852694 -0.432294580