Scaled/Standardized RAM Matrices

Derives the scaled/standardized RAM matrices.

RAMScaled(A, S, Filter, C = NULL, C.scaled = NULL, check = TRUE, ...)

# S3 method for default
RAMScaled(A, S, Filter, C = NULL, C.scaled = NULL, check = TRUE, ...)

# S3 method for yac_symbol
RAMScaled(
  A,
  S,
  Filter,
  C = NULL,
  C.scaled = NULL,
  check = TRUE,
  exe = TRUE,
  R = FALSE,
  format = "ysym",
  simplify = FALSE,
  ...
)

Arguments

A
t by t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
t by t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

Filter
p by t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

C
t by t numeric matrix \(\mathbf{C}\). Model-implied variance-covariance matrix.

C.scaled
t by t numeric matrix \(\mathbf{C}_{\mathrm{scaled}}\). Scaled/standardized model-implied variance-covariance matrix.

check
Logical. If check = TRUE do some preprocessing with input matrices using CheckRAMMatrices().

...
...

exe
Logical. If exe = TRUE, executes the resulting yacas expression. If exe = FALSE, returns the resulting yacas expression as a character string. If exe = FALSE, the arguments str, ysym, simplify, and tex, are ignored.

R
Logical. If R = TRUE, returns symbolic result as an R expression. If R = FALSE, returns symbolic result as "ysym", "str", or "tex" depending of format.

format
Character string. Only used when R = FALSE. If format = "ysym", returns symbolic result as yac_symbol. If format = "str", returns symbolic result as a characetr string. If format = "tex", returns symbolic result as LaTeX math.

simplify
Logical. Simplify symbolic results.

Value

Returns a list with the following elements

A.scaled

t by t matrix \(\mathbf{A}_{\mathrm{scaled}}\). Scaled/standardized asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S.scaled

t by t numeric matrix \(\mathbf{S}_{\mathrm{scaled}}\). Scaled/standardized symmetric paths (double-headed arrows), such as variances and covariances.

Details

The scaled/standardized \(\mathbf{A}\) and \(\mathbf{S}\) are given by

$$ \mathbf{A}_{\mathrm{scaled}} = \mathbf{D} \mathbf{A} \mathbf{D}^{-1} $$

$$ \mathbf{S}_{\mathrm{scaled}} = \mathbf{D} \mathbf{S} \mathbf{D} $$

where \(\mathbf{D}\) is a diagonal matrix whose diagonal elements are the diagonal elements of \(\mathbf{C}\) raised to \(-\frac{1}{2}\), that is, the inverse of the standard deviations of the variables.

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
colnames(A) <- rownames(A) <- c("y", "x", "e")
Filter <- diag(2)
Filter <- cbind(Filter, 0)
colnames(Filter) <- c("y", "x", "e")
(RAM <- RAMScaled(A, S, Filter))

#> $A.scaled
#>   y         x         e
#> y 0 0.4472136 0.8944272
#> x 0 0.0000000 0.0000000
#> e 0 0.0000000 0.0000000
#> 
#> $S.scaled
#>   y x e
#> y 0 0 0
#> x 0 1 0
#> e 0 0 1
#> 

C(A = RAM$A.scaled, S = RAM$S.scaled)

#>           y         x         e
#> y 1.0000000 0.4472136 0.8944272
#> x 0.4472136 1.0000000 0.0000000
#> e 0.8944272 0.0000000 1.0000000

M(A = RAM$A.scaled, S = RAM$S.scaled, Filter = Filter)

#>           y         x
#> y 1.0000000 0.4472136
#> x 0.4472136 1.0000000

# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e
#--------------------------------------------------------------------

A <- S <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
diag(S) <- c(0, "sigmax2", "sigmae2")
(RAM <- RAMScaled(Ryacas::ysym(A), S, Filter))

#> $A.scaled
#> {{                                                0, (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),        Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {                                                0,                                                 0,                                                 0},
#>  {                                                0,                                                 0,                                                 0}} 
#> 
#> $S.scaled
#> {{0, 0, 0},
#>  {0, 1, 0},
#>  {0, 0, 1}} 
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, format = "str")

#> $A.scaled
#> [1] "{{0,(Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},{0,0,0},{0,0,0}}"
#> 
#> $S.scaled
#> [1] "{{0,0,0},{0,1,0},{0,0,1}}"
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, format = "tex")

#> $A.scaled
#> [1] "\\left( \\begin{array}{ccc} 0 & \\frac{\\sqrt{\\mathrm{ sigmax2 }} \\beta }{\\sqrt{\\mathrm{ sigmax2 } \\beta  ^{2} + \\mathrm{ sigmae2 }}}  & \\frac{\\sqrt{\\mathrm{ sigmae2 }}}{\\sqrt{\\mathrm{ sigmax2 } \\beta  ^{2} + \\mathrm{ sigmae2 }}}  \\\\ 0 & 0 & 0 \\\\ 0 & 0 & 0 \\end{array} \\right)"
#> 
#> $S.scaled
#> [1] "\\left( \\begin{array}{ccc} 0 & 0 & 0 \\\\ 0 & 1 & 0 \\\\ 0 & 0 & 1 \\end{array} \\right)"
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, R = TRUE)

#> $A.scaled
#> expression(rbind(c(0, (sqrt(sigmax2) * beta)/sqrt(sigmax2 * beta^2 + 
#>     sigmae2), sqrt(sigmae2)/sqrt(sigmax2 * beta^2 + sigmae2)), 
#>     c(0, 0, 0), c(0, 0, 0)))
#> 
#> $S.scaled
#> expression(rbind(c(0, 0, 0), c(0, 1, 0), c(0, 0, 1)))
#> 

C(A = RAM$A.scaled, S = RAM$S.scaled)

#> {{((Sqrt(sigmax2)*beta)^2+sigmae2)/(sigmax2*beta^2+sigmae2),         (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {        (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                                                         1,                                                         0},
#>  {               Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2),                                                         0,                                                         1}} 

M(A = RAM$A.scaled, S = RAM$S.scaled, Filter = Filter)

#> {{((Sqrt(sigmax2)*beta)^2+sigmae2)/(sigmax2*beta^2+sigmae2),         (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {        (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                                                         1}} 

# Assigning values to symbols

beta <- 1
sigmax2 <- 0.25
sigmae2 <- 1

RAMScaled(Ryacas::ysym(A), S, Filter)

#> $A.scaled
#> {{                                                0, (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),        Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {                                                0,                                                 0,                                                 0},
#>  {                                                0,                                                 0,                                                 0}} 
#> 
#> $S.scaled
#> {{0, 0, 0},
#>  {0, 1, 0},
#>  {0, 0, 1}} 
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, format = "str")

#> $A.scaled
#> [1] "{{0,(Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},{0,0,0},{0,0,0}}"
#> 
#> $S.scaled
#> [1] "{{0,0,0},{0,1,0},{0,0,1}}"
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, format = "tex")

#> $A.scaled
#> [1] "\\left( \\begin{array}{ccc} 0 & \\frac{\\sqrt{\\mathrm{ sigmax2 }} \\beta }{\\sqrt{\\mathrm{ sigmax2 } \\beta  ^{2} + \\mathrm{ sigmae2 }}}  & \\frac{\\sqrt{\\mathrm{ sigmae2 }}}{\\sqrt{\\mathrm{ sigmax2 } \\beta  ^{2} + \\mathrm{ sigmae2 }}}  \\\\ 0 & 0 & 0 \\\\ 0 & 0 & 0 \\end{array} \\right)"
#> 
#> $S.scaled
#> [1] "\\left( \\begin{array}{ccc} 0 & 0 & 0 \\\\ 0 & 1 & 0 \\\\ 0 & 0 & 1 \\end{array} \\right)"
#> 

RAMScaled(Ryacas::ysym(A), S, Filter, R = TRUE)

#> $A.scaled
#> expression(rbind(c(0, (sqrt(sigmax2) * beta)/sqrt(sigmax2 * beta^2 + 
#>     sigmae2), sqrt(sigmae2)/sqrt(sigmax2 * beta^2 + sigmae2)), 
#>     c(0, 0, 0), c(0, 0, 0)))
#> 
#> $S.scaled
#> expression(rbind(c(0, 0, 0), c(0, 1, 0), c(0, 0, 1)))
#> 

eval(RAMScaled(Ryacas::ysym(A), S, Filter, R = TRUE))

#> $A.scaled
#> expression(rbind(c(0, (sqrt(sigmax2) * beta)/sqrt(sigmax2 * beta^2 + 
#>     sigmae2), sqrt(sigmae2)/sqrt(sigmax2 * beta^2 + sigmae2)), 
#>     c(0, 0, 0), c(0, 0, 0)))
#> 
#> $S.scaled
#> expression(rbind(c(0, 0, 0), c(0, 1, 0), c(0, 0, 1)))
#> 

C(A = RAM$A.scaled, S = RAM$S.scaled)

#> {{((Sqrt(sigmax2)*beta)^2+sigmae2)/(sigmax2*beta^2+sigmae2),         (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {        (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                                                         1,                                                         0},
#>  {               Sqrt(sigmae2)/Sqrt(sigmax2*beta^2+sigmae2),                                                         0,                                                         1}} 

M(A = RAM$A.scaled, S = RAM$S.scaled, Filter = Filter)

#> {{((Sqrt(sigmax2)*beta)^2+sigmae2)/(sigmax2*beta^2+sigmae2),         (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2)},
#>  {        (Sqrt(sigmax2)*beta)/Sqrt(sigmax2*beta^2+sigmae2),                                                         1}}