Equations to Sample Data

Generates data from a multivariate normal distribution from model equations.

eq2dat(eq, n, ...)

Arguments

eq
Character string. Equations. See Details.

n
Integer. Sample size.

...
Additional arguments to pass to MASS::mvrnorm().

Details

The input is a character string that specifies the associations between the variables.

Syntax

Each line should follow the syntax below

VARIABLE1 OPERATION VARIABLE2 VALUE

The associations are defined by the following operations

by

VARIABLE1 measured by VARIABLE2

on

VARIABLE1 regressed on VARIABLE2

with

VARIABLE1 covarying with VARIABLE2

on 1

VARIABLE1 regressed on 1 for mean structure

Each line should end with a semicolon (;).

Value

Each parameter should have a numeric value.

Comments

Comments can be written after a hash (#) sign.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

Author

Ivan Jacob Agaloos Pesigan

Examples

eq <- "
  # VARIABLE1 OPERATION VARIABLE2 VALUE
  e           by        y         1.00;
  y           on        x         1.00;
  e           with      e         0.25;
  x           with      x         0.25;
  y           on        1         0.00;
  x           on        1         0.50
"
eq2dat(eq, n = 100)

#>                   y           x
#>   [1,]  1.400590634  1.19725919
#>   [2,]  0.196691841  0.59787036
#>   [3,]  2.005625474  1.81440985
#>   [4,]  0.374590569  0.71149005
#>   [5,] -0.213071203  0.69730215
#>   [6,] -0.402511168  0.19307000
#>   [7,]  1.663017798  1.42168678
#>   [8,]  0.583020795  0.74626404
#>   [9,]  0.704955001  0.54416040
#>  [10,]  0.853514972  0.36303971
#>  [11,]  0.847115867  0.79041047
#>  [12,] -0.160207766  0.60033477
#>  [13,] -0.879129537 -0.44626810
#>  [14,]  1.550657491  1.30983509
#>  [15,]  0.239727052  0.13258648
#>  [16,]  2.127635470  0.73331039
#>  [17,]  0.928266732  0.61034453
#>  [18,]  0.262572037  0.96511253
#>  [19,]  0.166832601  0.20349072
#>  [20,]  1.303351173  0.60676702
#>  [21,]  0.509135980 -0.23519790
#>  [22,]  0.323238419  0.22748169
#>  [23,]  1.623888973  0.68899554
#>  [24,]  0.061568941  0.07407534
#>  [25,] -0.782173942 -0.33150925
#>  [26,]  0.570682839  0.53558536
#>  [27,]  1.242104242  0.73937111
#>  [28,]  0.468306688  0.57582586
#>  [29,]  0.983166981  0.99051678
#>  [30,]  1.572036334  1.09275253
#>  [31,] -0.329919916  0.40346269
#>  [32,]  0.516013166  0.20250211
#>  [33,]  0.345574025  0.37487311
#>  [34,] -0.829340776  0.15253372
#>  [35,]  0.269386134  0.70073219
#>  [36,]  0.666027799  0.43756158
#>  [37,]  1.675568860  1.53721201
#>  [38,]  0.645559592  0.69418156
#>  [39,]  0.861106019  0.39806195
#>  [40,] -0.223577365  0.02835483
#>  [41,]  0.425374276  0.51297440
#>  [42,]  0.805434210  0.98930673
#>  [43,]  0.839242210  0.02798265
#>  [44,]  0.797584391  0.40549157
#>  [45,]  0.353098262  0.05321963
#>  [46,] -1.181517128 -1.01939979
#>  [47,]  0.286069473  0.77454243
#>  [48,]  0.186027649  0.11901509
#>  [49,]  0.173491226  0.84641908
#>  [50,]  2.064637709  0.91019835
#>  [51,] -0.023560907  0.02052466
#>  [52,]  0.687230228  0.57135798
#>  [53,]  0.456946023  0.88679838
#>  [54,]  0.396147101  0.63852624
#>  [55,]  0.650651807  0.21075093
#>  [56,]  0.231710038  0.08800469
#>  [57,]  2.071279918  1.45711862
#>  [58,] -1.404009238 -0.54727938
#>  [59,]  0.875769937  0.44785383
#>  [60,]  0.483855091  0.19780227
#>  [61,] -0.548038717  0.54251395
#>  [62,]  1.109735093  0.53689347
#>  [63,] -0.316820445  0.10744673
#>  [64,]  0.711961460  0.53543482
#>  [65,]  1.019489143  1.47153251
#>  [66,]  1.043511160  1.12226016
#>  [67,] -0.417033348  0.34483644
#>  [68,]  0.468464400  0.34840540
#>  [69,]  0.117284763  0.36732384
#>  [70,]  1.753646967  1.08674211
#>  [71,]  1.495428904  1.15259266
#>  [72,]  0.146203104  0.63519299
#>  [73,] -0.213898590 -0.40281233
#>  [74,]  0.493209705  0.14674880
#>  [75,] -0.282351233 -0.26277404
#>  [76,]  0.307747176  0.00485734
#>  [77,]  0.005453857  0.36650323
#>  [78,]  0.368466026  0.88200005
#>  [79,]  0.331855012  0.50710389
#>  [80,]  0.921486157 -0.04297953
#>  [81,] -0.993544979 -0.04465337
#>  [82,] -0.579806250  0.24914220
#>  [83,] -0.287484801  0.62190589
#>  [84,]  0.365115207 -0.13769305
#>  [85,]  0.967338587  0.58751146
#>  [86,]  0.059540317 -0.49625874
#>  [87,]  2.272746441  1.65160519
#>  [88,]  1.023760438 -0.10787411
#>  [89,]  0.373933189  0.03629265
#>  [90,]  0.948497339  0.36207839
#>  [91,]  0.390656634  0.02416373
#>  [92,] -0.296113986  0.15219373
#>  [93,] -0.355461426  0.27036288
#>  [94,]  0.432176560  0.66837496
#>  [95,]  0.451981663 -0.17041096
#>  [96,] -0.629656806 -0.24647853
#>  [97,]  0.805034173  0.55175010
#>  [98,]  0.033726212 -0.20183282
#>  [99,] -0.204572584 -0.38636501
#> [100,]  0.803585185  0.46527091