Equations to RAM Matrices

Converts model equations to RAM matrices.

eq2ram(eq)

Arguments

eq
Character string. Equations. See Details.

Details

The input is a character string that specifies the associations between the variables.

Syntax

Each line should follow the syntax below

VARIABLE1 OPERATION VARIABLE2 LABEL

The associations are defined by the following operations

by

VARIABLE1 measured by VARIABLE2

on

VARIABLE1 regressed on VARIABLE2

with

VARIABLE1 covarying with VARIABLE2

on 1

VARIABLE1 regressed on 1 for mean structure

Label

Each parameter should be labeled. The LABEL should be a number for fixed parameters and a variable name for free parameters. Equality contraints can be imposed by using the same variable name. Variable names can be written following the R mathematical notation grDevices::plotmath().

Comments

Comments can be written after a hash (#) sign.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

Author

Ivan Jacob Agaloos Pesigan

Examples

eq <- "
  # VARIABLE1 OPERATION VARIABLE2 LABEL
  e           by        y         1
  y           on        x         beta
  e           with      e         sigma[varepsilon]^2
  x           with      x         sigma[x]^2
  y           on        1         alpha
  x           on        1         mu[x]
"
eq2ram(eq)

#> $eq
#>   var1   op var2               label
#> 1    e   by    y                   1
#> 2    y   on    x                beta
#> 3    e with    e sigma[varepsilon]^2
#> 4    x with    x          sigma[x]^2
#> 5    y   on    1               alpha
#> 6    x   on    1               mu[x]
#> 
#> $variables
#> [1] "y" "x" "e"
#> 
#> $A
#>   y   x      e  
#> y "0" "beta" "1"
#> x "0" "0"    "0"
#> e "0" "0"    "0"
#> 
#> $S
#>   y   x            e                    
#> y "0" "0"          "0"                  
#> x "0" "sigma[x]^2" "0"                  
#> e "0" "0"          "sigma[varepsilon]^2"
#> 
#> $filter
#>   y x e
#> y 1 0 0
#> x 0 1 0
#> 
#> $u
#>   u      
#> y "alpha"
#> x "mu[x]"
#> e "0"    
#>