Generate Data from a Multivariate Normal Distribution Using the Reticular Action Model (RAM) Notation

Generates data from a multivariate normal distribution using the Reticular Action Model (RAM) notation with the MASS::mvrnorm() function.

mvn(n, A, S, u, filter, ...)

Arguments

n
Integer. Sample size.

A
t x t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
t x t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

u
vector of length t or t by 1 matrix. Mean structure parameters.

filter
p x t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

...
Additional arguments to pass to MASS::mvrnorm().

Details

The Multivariate Normal Distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbf{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{g} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} = \mathbf{F} \mathbf{E} \mathbf{u} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{M} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{C} \mathbf{F}^{\mathsf{T}} $$

Author

Ivan Jacob Agaloos Pesigan

Examples

# This is an example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

n <- 1000
A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- 1
A[1, 3] <- 1
diag(S) <- c(0, 0.25, 0.25)
u <- c(-0.50, 0.50, 0.00)
filter <- diag(2)
filter <- cbind(filter, 0)
colnames(filter) <- c("y", "x", "e")
rownames(filter) <- c("y", "x")
mvn(n, A, S, u, filter)

#>                    y             x
#>    [1,]  0.147681546  0.9217942519
#>    [2,] -0.431139840 -0.8959357990
#>    [3,] -0.124995938  0.8112860502
#>    [4,]  0.073554019  0.7997885536
#>    [5,]  1.651543820  1.5927993882
#>    [6,]  0.075741097  0.2766747511
#>    [7,]  0.747834948  0.9804090177
#>    [8,]  0.649080598  0.4241265547
#>    [9,]  1.434162081  1.3551146728
#>   [10,] -1.925296635 -0.4610918545
#>   [11,]  0.900180969  0.8183742085
#>   [12,]  0.237574580  0.6397840462
#>   [13,] -0.428051704 -0.0174853711
#>   [14,]  1.200716335  0.5123188002
#>   [15,] -0.382353597  0.2843910545
#>   [16,] -0.012629335  0.4048585974
#>   [17,] -0.387920849  0.2682037983
#>   [18,] -0.420851896  0.5417079544
#>   [19,]  1.227305337  0.7490377952
#>   [20,] -0.765329210  0.2899613819
#>   [21,]  0.449418487  0.2943942890
#>   [22,] -0.079799525  0.7454151525
#>   [23,]  0.263952994  0.0110468868
#>   [24,]  0.026006339  0.2666411879
#>   [25,]  0.701632592  0.9011643301
#>   [26,] -0.796081917 -0.1099045492
#>   [27,] -0.222690723  0.3364617387
#>   [28,]  0.363170579  0.6397919832
#>   [29,] -0.165051012 -0.3236006654
#>   [30,]  0.897607274  1.4004664545
#>   [31,] -0.254150923  0.5458647930
#>   [32,]  0.945026489  0.9911284179
#>   [33,] -0.546277746  0.2343371675
#>   [34,]  1.144972081  1.0518648880
#>   [35,] -0.037400978  0.4511761764
#>   [36,]  0.190477504  1.1759437412
#>   [37,]  0.655410299  0.7401463652
#>   [38,] -0.322642639 -0.0170479898
#>   [39,] -1.004573043  0.3505704816
#>   [40,]  0.985348790  1.5009361081
#>   [41,]  0.436941538  1.1823030939
#>   [42,] -0.903109683 -0.0663676076
#>   [43,] -0.760627224  0.0377027690
#>   [44,] -0.873711474  0.0612866304
#>   [45,]  0.981055041  1.1151188521
#>   [46,] -0.817674080 -0.3050565412
#>   [47,] -0.124025824  0.6958676718
#>   [48,]  0.140713679  0.3921759829
#>   [49,] -0.559290358  0.7256626172
#>   [50,] -0.248978420  0.7044636084
#>   [51,]  0.893587060  0.3317141460
#>   [52,]  0.045793427  1.2008544819
#>   [53,]  0.673649021  1.2468378947
#>   [54,]  0.200298547  1.3326933911
#>   [55,] -1.131909685  0.0911587832
#>   [56,]  0.266541259  1.3438169086
#>   [57,] -0.321357893  0.5740509508
#>   [58,]  0.234364708  0.8595162895
#>   [59,]  0.295676759  0.9523216266
#>   [60,] -0.892166760 -0.3033223905
#>   [61,] -0.112668487  0.4519685988
#>   [62,]  0.683590510  1.5995816978
#>   [63,] -0.199965875  0.8394067592
#>   [64,] -0.143407294  1.0586355516
#>   [65,]  0.541109113  1.0191811869
#>   [66,]  1.164509350  1.8506643293
#>   [67,] -1.296416969 -0.3159354448
#>   [68,]  0.646081274  0.9444110151
#>   [69,] -0.028937347  0.7047114495
#>   [70,] -0.139036807  0.3557064665
#>   [71,] -0.149516913  0.6482096866
#>   [72,]  1.444188188  1.5141993725
#>   [73,]  0.182324653  0.3863606262
#>   [74,]  0.089760031  0.1819613097
#>   [75,] -0.189750619  0.7948873398
#>   [76,] -1.143285128 -0.5396765051
#>   [77,] -1.288430832 -0.7372598752
#>   [78,] -0.599855601  0.3784484007
#>   [79,] -0.705654410  0.4615067612
#>   [80,]  0.271385704  0.5351773141
#>   [81,] -0.860212826  0.3345428758
#>   [82,]  0.722647638  0.7450065452
#>   [83,] -0.316869873  0.1457534301
#>   [84,] -0.204810783  0.3351408842
#>   [85,] -0.293564195  0.4107430200
#>   [86,] -0.936182956  0.2761383174
#>   [87,]  0.711975023  0.7968167774
#>   [88,]  0.014967095  0.1405650681
#>   [89,] -0.871489680 -0.2679956339
#>   [90,]  0.185548808  0.7902801648
#>   [91,] -0.128163311  0.9752639759
#>   [92,]  0.252814675  0.4321669872
#>   [93,]  0.894458294  0.6062410433
#>   [94,] -0.305250988  0.2541463544
#>   [95,] -0.875664503 -0.5520740819
#>   [96,]  1.107090402  1.0400731263
#>   [97,]  1.027525483  1.0164655898
#>   [98,] -0.840474923  0.5106954195
#>   [99,] -0.559375714  0.4446478926
#>  [100,] -1.377291928 -0.5219553661
#>  [101,]  0.058223451  0.9538116031
#>  [102,]  0.688431085  1.0241370916
#>  [103,] -0.666948796 -0.0783660990
#>  [104,] -0.815904607  0.0019035518
#>  [105,] -0.224518984  0.5579853049
#>  [106,]  0.417839788  0.4400813436
#>  [107,] -0.303675352  0.0431930980
#>  [108,] -1.385372224 -0.2963355168
#>  [109,] -1.419009118 -0.1041921208
#>  [110,]  0.889652602  1.5051475032
#>  [111,]  0.203223455  1.0020338590
#>  [112,]  0.800030222  1.0051400556
#>  [113,] -0.429439012  0.3344712685
#>  [114,]  1.330924701  1.1462061325
#>  [115,] -0.121723004  0.0916578237
#>  [116,] -0.189075134  0.7410922709
#>  [117,] -1.054092682  0.3902412790
#>  [118,]  0.309961696  0.3938324782
#>  [119,]  0.891598500  0.6828974316
#>  [120,]  0.034371873  0.1384984747
#>  [121,] -0.459295595  0.2420876831
#>  [122,] -0.081944044  0.1033600398
#>  [123,] -1.079215058 -0.0274817919
#>  [124,] -0.117436659  0.5791803532
#>  [125,] -1.539799601 -0.2808135766
#>  [126,]  0.741318933  1.5721542266
#>  [127,] -0.434310584  0.5750534259
#>  [128,] -1.015048590 -0.0027281407
#>  [129,] -0.189704518  0.2525370678
#>  [130,] -0.442626811  0.2089359581
#>  [131,] -0.540476233 -0.2445917943
#>  [132,]  1.276592877  1.4806514448
#>  [133,] -0.753688315 -0.1400209304
#>  [134,]  0.346788618  0.1993794930
#>  [135,] -0.216048343  0.3955694568
#>  [136,] -0.935523705  0.0650113639
#>  [137,]  0.866666242  0.8448181888
#>  [138,]  0.649073505  1.1902898355
#>  [139,] -0.035878121  0.3887896052
#>  [140,] -0.713569975  0.3420940037
#>  [141,]  0.257920440  0.4432848516
#>  [142,] -1.255889962 -0.6790108804
#>  [143,] -0.076860897  0.7950509048
#>  [144,]  1.072322531  0.9083169252
#>  [145,]  0.717121573  1.0974776668
#>  [146,] -1.358810285  0.0754781014
#>  [147,]  0.093030615  0.4726939565
#>  [148,]  0.270189651  0.7357305980
#>  [149,] -0.188194790  0.9879587330
#>  [150,] -0.230013803 -0.3380717553
#>  [151,]  0.265712339  0.9609735309
#>  [152,]  0.162409269  0.4610052487
#>  [153,] -0.436812956  0.3966426641
#>  [154,] -1.309305943 -0.0492008861
#>  [155,] -0.927168412 -0.2290642337
#>  [156,] -1.234860901  0.1617878773
#>  [157,]  0.479056587  0.3157965622
#>  [158,] -1.035496623 -0.6361878082
#>  [159,]  0.338961561  0.8002072133
#>  [160,]  0.526803595  0.9798715885
#>  [161,]  0.415074004  0.3574983968
#>  [162,] -0.622463915 -0.1284309175
#>  [163,] -0.554510417  0.1460487780
#>  [164,] -0.856352852 -0.8896675030
#>  [165,]  0.492394070 -0.1350262454
#>  [166,] -0.479168543 -0.2365292806
#>  [167,]  0.975810128  1.5576076835
#>  [168,]  0.637615477  0.7486628857
#>  [169,] -0.605496105 -0.2137695954
#>  [170,]  0.042831554  0.6981787287
#>  [171,]  0.134119273  0.0078270554
#>  [172,]  0.577763893  0.6399307923
#>  [173,]  0.736692460  0.7859858428
#>  [174,]  0.550347483  1.1105407864
#>  [175,]  0.388813260  0.3919025942
#>  [176,] -0.768237443 -0.0302455162
#>  [177,] -0.403778494  0.5299407058
#>  [178,]  0.550404291  0.3341023585
#>  [179,]  0.240948935  0.3151886675
#>  [180,] -0.679118358 -0.2888334598
#>  [181,] -0.142272256  0.2178532187
#>  [182,]  0.078413159  0.9072390728
#>  [183,]  0.260790246  0.2296864358
#>  [184,] -0.108005517  0.6212571328
#>  [185,] -0.307897245  0.4039998180
#>  [186,] -0.067633200  0.5773768517
#>  [187,]  0.006708664  0.4774703037
#>  [188,] -0.680866236  0.1692430372
#>  [189,]  0.248041266  1.1523826114
#>  [190,] -0.238767493  0.3671263777
#>  [191,]  0.352112480  0.5644847519
#>  [192,] -0.610198564  0.0496419668
#>  [193,] -1.542214886  0.1633998356
#>  [194,]  0.589396480  0.6309423993
#>  [195,] -0.167543196  0.7579945623
#>  [196,] -1.243330031 -0.6185461123
#>  [197,]  0.654296667  1.5059683168
#>  [198,] -0.623627066 -0.2744357854
#>  [199,]  0.122134162  0.6150892068
#>  [200,]  0.568697367  0.1615544607
#>  [201,] -1.379497766  0.0604740850
#>  [202,]  0.536248677  0.9330340586
#>  [203,]  0.516484740  1.1440165141
#>  [204,] -0.678019583  0.0313010000
#>  [205,]  0.907861361  1.3332394468
#>  [206,]  0.899568083  0.8677168167
#>  [207,] -0.464968673  0.2825041393
#>  [208,] -1.441382548 -0.4012891819
#>  [209,]  0.480624745  0.6667766641
#>  [210,]  1.120232753  1.2028749796
#>  [211,]  0.193533763  0.2029230201
#>  [212,]  0.512494792  0.6810250756
#>  [213,]  0.513296501  0.6997748400
#>  [214,]  0.194548322  0.9216869886
#>  [215,] -0.681493730 -0.4277544730
#>  [216,] -0.469194878  0.2796038952
#>  [217,] -0.309721647  0.2096710779
#>  [218,]  1.029554363  1.5292352246
#>  [219,] -0.993399601  0.7322491327
#>  [220,] -0.242785141  0.5496138474
#>  [221,] -0.411504063  0.2720427274
#>  [222,]  0.290080512  0.3042789025
#>  [223,] -0.474975218  0.1282955273
#>  [224,]  0.606120889  1.0541828723
#>  [225,]  2.268239545  1.3525327489
#>  [226,] -0.588959592  0.3465053065
#>  [227,]  0.752446829  0.3566347679
#>  [228,]  1.140901182  1.5687430722
#>  [229,]  0.146131664  0.1461415304
#>  [230,] -0.427379572 -0.1554466082
#>  [231,] -0.411736476  0.4678350642
#>  [232,]  0.667801215  0.7285936444
#>  [233,] -0.314912193  0.1382439416
#>  [234,] -0.551787775 -0.3802605072
#>  [235,]  0.663691811  0.5897922012
#>  [236,]  0.249510598  0.8491745313
#>  [237,]  1.011317537  0.6580238914
#>  [238,]  0.425605619  0.5494902587
#>  [239,] -0.291957392  0.7948943280
#>  [240,]  1.268210242  1.1687394618
#>  [241,]  1.162727156  0.7848015055
#>  [242,] -0.339715080 -0.0416428904
#>  [243,]  0.726121123  1.2345722881
#>  [244,]  0.186719975  0.7646782422
#>  [245,] -0.240607747  0.1782433647
#>  [246,]  0.252374685  0.5884382411
#>  [247,]  0.670636395  1.3957074337
#>  [248,]  0.899327610  0.6299263150
#>  [249,] -1.004374300 -0.2048321267
#>  [250,] -0.160256124  0.2254701273
#>  [251,] -1.311563521 -0.1161246137
#>  [252,] -0.042184171 -0.0267083925
#>  [253,]  0.420605917  1.2331719369
#>  [254,]  1.193464908  1.0069763782
#>  [255,]  0.220252396  0.2729774726
#>  [256,]  1.380056294  1.4756161850
#>  [257,] -0.016432466  0.5210977144
#>  [258,] -0.034400580  1.1031363086
#>  [259,] -0.939859368  0.2566581342
#>  [260,]  0.284409576  0.3801896586
#>  [261,] -0.615180533 -0.0714333930
#>  [262,]  0.345891679  0.3461577737
#>  [263,] -1.006327984 -0.4239612919
#>  [264,]  0.600162148  0.7201143010
#>  [265,]  0.528417975  1.0667959573
#>  [266,]  0.278779378  0.4729277613
#>  [267,]  0.320746690  0.8941691787
#>  [268,]  0.056323679  0.5970450095
#>  [269,] -0.936480441  0.1997559070
#>  [270,] -0.539700183  0.3867981540
#>  [271,]  0.700899311  0.3347581861
#>  [272,]  0.692515904  0.6224620412
#>  [273,]  0.647121266  0.7770936419
#>  [274,]  1.427073216  1.5290621388
#>  [275,]  1.093053586  0.8485968467
#>  [276,]  0.532804833  0.3967565156
#>  [277,]  0.392564475  0.7598877710
#>  [278,]  0.242929480  0.3642712818
#>  [279,] -0.144916754 -0.0133862917
#>  [280,]  0.873470468  1.1385802389
#>  [281,]  0.129318024  0.6938826866
#>  [282,] -0.495478635  0.2977819804
#>  [283,] -0.645882004  0.3926897284
#>  [284,] -0.632366149  0.1435255066
#>  [285,] -1.036952673 -0.1139730286
#>  [286,]  0.569871563  0.5926460614
#>  [287,] -0.349491700  0.2386956522
#>  [288,] -0.303298600 -0.1584643081
#>  [289,] -1.486088627 -0.0135900276
#>  [290,]  1.280921303  1.5975486716
#>  [291,] -0.288805917  0.8679749858
#>  [292,]  0.136329151  0.6874346983
#>  [293,]  0.314822960  0.6789981993
#>  [294,]  0.969391189  1.1023376166
#>  [295,] -0.130594807  1.4906053647
#>  [296,]  0.188955057  0.6093850791
#>  [297,]  0.730800123  0.9874222740
#>  [298,] -0.153528063  0.1911080325
#>  [299,]  0.161838989  0.7546853284
#>  [300,] -0.835451324 -0.2468906861
#>  [301,]  0.395677411  0.9547035610
#>  [302,] -0.483178892  0.2628114796
#>  [303,] -0.512151411  0.8806750439
#>  [304,] -0.385489591  0.8191092204
#>  [305,]  0.647179376  1.3049395689
#>  [306,]  0.286981965  0.0969255562
#>  [307,] -0.576717758  0.3759916230
#>  [308,] -0.230834069 -0.2117969579
#>  [309,] -0.917730819  0.4592477888
#>  [310,] -0.916049411  0.2213809720
#>  [311,]  0.770648966  1.1595506081
#>  [312,] -1.761796328 -0.7348221054
#>  [313,]  0.265801914  0.3113941939
#>  [314,]  0.197270104  0.8324945473
#>  [315,]  0.397142985  0.1626954435
#>  [316,]  0.652506522  0.8211815886
#>  [317,] -0.705953443  0.2507295752
#>  [318,] -0.188646378  0.6826687213
#>  [319,]  0.920642566  0.9475049468
#>  [320,] -0.721055713  0.0883255695
#>  [321,]  0.552313107  0.7308912575
#>  [322,]  0.053443078  0.7061311327
#>  [323,] -0.403163366  0.3384929492
#>  [324,] -0.244757685  0.0493393006
#>  [325,] -0.631017475  0.6751305046
#>  [326,]  0.469081302  0.7559418592
#>  [327,] -0.099131646  0.5720354531
#>  [328,]  1.911423714  1.7283353202
#>  [329,]  0.694059314  0.7808118520
#>  [330,]  0.443554760  0.9862710573
#>  [331,]  0.425603166  0.8333120082
#>  [332,] -0.474811233  0.3044690793
#>  [333,]  0.361050135  0.6963833361
#>  [334,] -0.128023698  0.2911033800
#>  [335,] -0.232290835  0.1564815413
#>  [336,] -0.492008035  0.1819720433
#>  [337,] -0.683928390  0.6620214602
#>  [338,]  0.254676254  1.0387281012
#>  [339,]  0.181472245 -0.1332889597
#>  [340,] -0.747565056 -0.0261148915
#>  [341,] -1.064698239 -0.5679013396
#>  [342,]  0.128358396  0.4213062180
#>  [343,] -0.852079958 -0.2259898115
#>  [344,]  0.536974313  0.5966799110
#>  [345,]  0.291797262  0.3612200964
#>  [346,]  0.527516241  0.6985881816
#>  [347,]  0.314143853  0.6760317097
#>  [348,] -0.281605074  0.0223547352
#>  [349,] -0.475326183  0.3085749326
#>  [350,]  0.349973253  1.0038897643
#>  [351,]  0.539561994  0.8328668996
#>  [352,]  0.463005260  1.2175485576
#>  [353,] -1.174880236 -0.3577717691
#>  [354,] -0.329423786  0.4959989662
#>  [355,] -0.065736345  0.2076439407
#>  [356,]  0.544764116  0.8587855862
#>  [357,] -0.266619970  0.7687989185
#>  [358,]  0.210100040  0.6734081182
#>  [359,] -1.244944041 -0.0121824615
#>  [360,]  0.180487208  1.1988347941
#>  [361,] -0.323426882  0.1199839284
#>  [362,]  0.314663390  0.2223104926
#>  [363,]  0.914839370  1.6524429226
#>  [364,] -1.187958257  0.2200241640
#>  [365,]  2.077444145  1.2091878887
#>  [366,]  0.914013544  0.6098836001
#>  [367,]  0.452853608  0.8559463629
#>  [368,]  0.378301753  0.9659459166
#>  [369,] -0.305362186  0.1663658160
#>  [370,]  0.435713754  0.2817854998
#>  [371,]  0.571711094  0.8669985831
#>  [372,]  1.020773808  0.9333731785
#>  [373,]  0.845358710  0.3902890798
#>  [374,]  0.383877285  0.8551147287
#>  [375,] -0.634231647  0.2705912932
#>  [376,] -0.140784564  0.2618395292
#>  [377,] -1.454146770  0.0720385214
#>  [378,]  0.726161600  0.7008115643
#>  [379,]  0.112679940  0.3891256981
#>  [380,]  0.558148275  0.3219672583
#>  [381,]  0.558612395  0.4066427772
#>  [382,]  0.104220044  0.4777659953
#>  [383,]  1.668738348  1.6398962310
#>  [384,]  0.110093675  0.0650485942
#>  [385,] -0.354938797  0.5349507180
#>  [386,] -0.684333424 -0.5989642570
#>  [387,] -0.234024763 -0.2999493578
#>  [388,] -1.023779791  0.3598809258
#>  [389,]  0.043108818  0.6409827346
#>  [390,]  0.188761265  0.9869306284
#>  [391,] -1.145340391  0.0145534159
#>  [392,]  0.179102938  0.7157020729
#>  [393,]  0.281760522  0.1266938541
#>  [394,]  0.447552733  0.6430197127
#>  [395,]  0.508696550  0.8216698762
#>  [396,] -0.084156494  0.8039659152
#>  [397,]  0.158365599  1.1067865170
#>  [398,]  0.105375936  0.0401935174
#>  [399,] -0.584151209  0.2928134020
#>  [400,]  1.199739496  1.5080299298
#>  [401,] -0.277579353  0.5858186772
#>  [402,] -0.193196758 -0.1552608650
#>  [403,] -0.150333421  0.1001230981
#>  [404,] -1.150165175 -0.6809121638
#>  [405,]  0.396860868  0.6369601300
#>  [406,]  0.697963415  1.0787379042
#>  [407,]  0.630340044  0.9399963762
#>  [408,] -0.364321944  0.3556981986
#>  [409,]  0.831092291  0.3789538012
#>  [410,] -0.174535398  0.4218121500
#>  [411,]  0.861447818  0.9873069600
#>  [412,]  1.752300553  1.4404931811
#>  [413,]  1.086373880  1.0339468817
#>  [414,]  0.324397039  0.6401215490
#>  [415,]  0.836036283  0.5977063514
#>  [416,]  0.085977349  0.5477792897
#>  [417,] -0.916247329 -0.0740608341
#>  [418,]  0.457840299  1.0831508635
#>  [419,] -0.451004846  0.2040449331
#>  [420,]  0.930669182  1.1816626727
#>  [421,] -0.609030337 -0.3150995172
#>  [422,]  0.971910790  1.0890487190
#>  [423,] -0.728878194 -0.0160047436
#>  [424,] -0.467812340  0.1828458520
#>  [425,]  0.685554543  0.7208470283
#>  [426,]  0.724032894  1.0080903342
#>  [427,] -0.011539214  0.5756871814
#>  [428,] -0.518048207  0.0916749955
#>  [429,]  0.192810678  0.9564547342
#>  [430,] -0.406234030 -0.3015701374
#>  [431,] -0.011937729  0.7874526085
#>  [432,]  0.974326637  0.6258284746
#>  [433,]  0.783515187  0.6879670354
#>  [434,] -0.179448354  0.3454903988
#>  [435,] -0.460228254 -0.1047656165
#>  [436,]  0.498441772  1.4614856422
#>  [437,]  0.627412442  1.2354313691
#>  [438,]  0.912084722  1.2359112275
#>  [439,]  0.037814816  1.1992229635
#>  [440,] -0.537417753  0.0756992491
#>  [441,] -0.579558298  0.2196992543
#>  [442,] -0.268858999  1.1948548121
#>  [443,] -0.421140506  0.2385372870
#>  [444,]  0.639511364  0.6519404400
#>  [445,] -0.574697160  0.0624201276
#>  [446,] -0.105248949  0.6499613876
#>  [447,] -0.220940194  0.5107268860
#>  [448,]  0.732075773  0.4387005287
#>  [449,]  0.880559045  0.9564397003
#>  [450,] -0.513266195  0.7044733597
#>  [451,]  0.465737632  0.9021096595
#>  [452,]  0.033498226  0.6952704362
#>  [453,] -0.106392449  0.1304587129
#>  [454,]  0.190422652  0.6367046627
#>  [455,]  0.082627293  0.2052476256
#>  [456,] -0.252827538 -0.1991695649
#>  [457,]  0.364595317  0.8503756181
#>  [458,]  0.866680832  0.8010465633
#>  [459,] -0.617341541  0.6759018997
#>  [460,] -0.409132616  0.0292983452
#>  [461,]  0.759376463  1.1525395568
#>  [462,] -0.769035259  0.1725286454
#>  [463,] -0.395005921  0.4233174302
#>  [464,] -0.184162483 -0.4184312577
#>  [465,]  0.331219624  0.1645323061
#>  [466,]  0.582524650  0.9890172443
#>  [467,]  0.393435089  0.4248566182
#>  [468,] -0.140773243  0.4909904730
#>  [469,]  0.043969184 -0.2069235253
#>  [470,] -2.005988888 -0.0718364043
#>  [471,]  1.528321869  1.3817317585
#>  [472,] -0.413638753  0.5216221820
#>  [473,]  0.064337572  0.9460791411
#>  [474,]  0.639787483  0.4602196232
#>  [475,]  0.193911748  0.2634044985
#>  [476,] -0.093324903  0.0093937976
#>  [477,]  0.538151073  0.6027054042
#>  [478,] -0.704957721 -0.1292481877
#>  [479,]  0.028815512  0.8157356847
#>  [480,]  1.104887625  1.2407365455
#>  [481,]  1.031765737  1.1440352099
#>  [482,]  1.347726708  1.4748579923
#>  [483,]  0.452909429  0.9262300447
#>  [484,]  0.299728903  0.2214516969
#>  [485,] -0.558778821 -0.1436175302
#>  [486,] -1.423240133 -0.5321066567
#>  [487,] -1.320286193 -0.9377790774
#>  [488,]  0.342294799  1.5164255166
#>  [489,] -0.063628937  0.5278360904
#>  [490,]  0.640244667  0.6518768293
#>  [491,]  0.411213313  1.0429954047
#>  [492,]  0.245587627  1.2207579645
#>  [493,] -0.272079502 -0.1090685149
#>  [494,]  0.402484829  0.2024583176
#>  [495,]  0.849597840  1.4498988474
#>  [496,]  0.501365279  0.5870384842
#>  [497,]  1.633233451  0.9703241774
#>  [498,] -0.349783684  0.4450320403
#>  [499,] -0.334880680  0.1954614226
#>  [500,] -0.219051034  0.8107948955
#>  [501,] -0.536926383 -0.0056745023
#>  [502,]  0.530427221  0.2212100532
#>  [503,] -0.422887010  0.2518847062
#>  [504,] -0.068279697  0.6179803027
#>  [505,]  0.477719313  0.5284571265
#>  [506,]  0.346302135  0.9230198841
#>  [507,]  0.449784442  0.7507739439
#>  [508,] -0.063718025  0.4390774407
#>  [509,] -0.672132792 -0.2235511011
#>  [510,]  0.110507991 -0.3672698790
#>  [511,] -1.548913464 -0.4915231023
#>  [512,] -0.178104949  0.5788078114
#>  [513,]  1.501534217  1.1466715172
#>  [514,]  0.488932584  0.3559747634
#>  [515,]  0.565939658  0.1665995038
#>  [516,] -0.860472925  0.0136952048
#>  [517,]  0.937200696  1.3549791994
#>  [518,]  0.130811235  0.7658611530
#>  [519,] -0.011511910  0.5495699759
#>  [520,]  1.666453588  1.4817615667
#>  [521,] -0.770646838  0.3775707945
#>  [522,]  0.817229634  1.4595097554
#>  [523,] -0.314289107  0.3225912024
#>  [524,] -0.757887517 -0.3815805385
#>  [525,]  0.219261876  0.1761729817
#>  [526,] -0.013985998  0.6906984054
#>  [527,] -0.071276612  0.2080554336
#>  [528,] -0.360156468  0.6151710066
#>  [529,] -0.032216213  0.8343327839
#>  [530,] -0.389551282  0.2690949229
#>  [531,]  0.200950914  0.4619314345
#>  [532,]  0.505533042  0.9319936250
#>  [533,]  1.147399413  1.1680643537
#>  [534,] -0.178765090  0.0076356675
#>  [535,] -1.154198139 -0.3321151731
#>  [536,] -0.279822480  0.0411521094
#>  [537,] -0.627957186 -0.0485048049
#>  [538,] -0.249294244  0.7162117231
#>  [539,]  0.716134040  1.0002457188
#>  [540,]  0.520930887  1.4168828128
#>  [541,]  0.825762752 -0.0210743856
#>  [542,]  0.625103022  0.5369336521
#>  [543,]  0.221922366  0.4524568291
#>  [544,] -1.117092629 -0.2846785315
#>  [545,]  0.585786280  1.1429913252
#>  [546,] -0.185983992  1.0409518399
#>  [547,]  0.060912612  0.4728445624
#>  [548,]  0.570860967  1.0438112682
#>  [549,] -0.221452710  0.2177585093
#>  [550,] -0.098389707  0.4837243644
#>  [551,] -0.055579011  0.4915678821
#>  [552,]  0.973148822  0.3401189766
#>  [553,] -0.405505458 -0.2308907782
#>  [554,]  0.744032035  0.5237085027
#>  [555,] -0.586887015  0.4213164389
#>  [556,] -0.178539272  0.5500323699
#>  [557,] -0.128531674  0.5099380063
#>  [558,]  1.114468330  1.0558975162
#>  [559,] -0.144808009  0.4229032237
#>  [560,]  0.541928608  0.7273195414
#>  [561,] -0.284828402  0.4042710294
#>  [562,] -0.932287394 -0.1443735505
#>  [563,] -0.327655571  0.4566331196
#>  [564,]  1.063046606  1.1892332939
#>  [565,]  0.138615878  0.3552645141
#>  [566,] -0.213171446  0.0538276319
#>  [567,] -0.448389969  0.3793018552
#>  [568,] -0.595106758  0.1274700475
#>  [569,] -0.462644414  0.1954525098
#>  [570,]  0.199461962  0.4276025097
#>  [571,]  1.312409607  1.1223316170
#>  [572,]  0.320128405  1.5780242768
#>  [573,] -0.702289951  0.3588630198
#>  [574,] -0.496711715  0.3690855908
#>  [575,] -0.022011287  0.7236404867
#>  [576,] -0.689893181  0.1803386750
#>  [577,]  0.616542330  0.9816662609
#>  [578,] -0.153374096  0.3556341156
#>  [579,]  0.548718837  0.4514409931
#>  [580,] -0.639968855  0.1038133986
#>  [581,]  0.392765620  0.6918510130
#>  [582,]  0.226779171  0.8292757244
#>  [583,]  0.686665475  0.0649512190
#>  [584,]  0.148470199  1.2080400342
#>  [585,] -0.193036156  0.1327333666
#>  [586,] -0.110920053 -0.0631687026
#>  [587,] -0.063457580 -0.2315978984
#>  [588,]  1.467188043  1.6511949861
#>  [589,] -0.386261770  0.6474022771
#>  [590,] -0.700866543 -0.5266294908
#>  [591,] -0.859856692 -0.2278887895
#>  [592,] -0.493073777 -0.3339709899
#>  [593,] -0.749217963  0.7394460797
#>  [594,] -0.889667446  0.2699795901
#>  [595,] -1.063102245  0.1327591232
#>  [596,] -0.393377652  0.5787851600
#>  [597,] -1.540820928 -0.3451772416
#>  [598,] -0.239793660  0.6738944490
#>  [599,] -0.931078192 -0.1114325834
#>  [600,]  0.455271378  0.5196192043
#>  [601,] -0.893779076 -0.4158275635
#>  [602,]  0.372202916  0.5382270024
#>  [603,]  0.547272777  0.4155285800
#>  [604,] -0.557236316  0.0927133153
#>  [605,] -0.136309170  0.2058155691
#>  [606,]  0.551905637  0.8832390719
#>  [607,]  0.297452939  0.3552494868
#>  [608,]  1.060617973  1.1615234315
#>  [609,] -0.310342925  0.6432416694
#>  [610,] -0.193607896  0.7654528927
#>  [611,] -0.536628084  0.8175838769
#>  [612,] -1.280312489  0.0961709786
#>  [613,] -1.167564653  0.1888398312
#>  [614,]  0.715554236  0.8015535769
#>  [615,] -0.492632578 -0.2659138616
#>  [616,] -0.293396667  0.9166581917
#>  [617,]  1.180289089  1.2743605799
#>  [618,]  0.509743435  1.1680078969
#>  [619,]  0.331217100  0.3165163430
#>  [620,]  1.047178074  0.3057635325
#>  [621,] -2.126265717 -0.2027649377
#>  [622,] -0.936158565 -0.1279374506
#>  [623,] -0.855015989 -0.2094970597
#>  [624,]  0.468110093  0.3048991701
#>  [625,]  0.594265528  0.8756150300
#>  [626,]  0.317492926  0.7657963077
#>  [627,] -0.907532815  0.0991766035
#>  [628,] -0.664284902  0.7947063575
#>  [629,]  1.671735815  1.0189039848
#>  [630,]  0.073437973  0.3287323061
#>  [631,] -0.692913839  0.3089436012
#>  [632,] -0.536525206  0.2244055172
#>  [633,] -0.455502683  0.3796475798
#>  [634,]  0.775071508  1.1672799750
#>  [635,]  0.176541343  0.5342357767
#>  [636,] -0.025426556 -0.1112808812
#>  [637,]  0.094498034  1.1268261539
#>  [638,]  0.503957637  0.6256170324
#>  [639,] -1.909815387 -0.1210160886
#>  [640,]  0.172462412  0.4747944293
#>  [641,]  0.428297106  0.4847159479
#>  [642,] -0.568627789  0.6168203542
#>  [643,]  1.204306582  1.5033854210
#>  [644,]  1.347808592  1.3701844624
#>  [645,] -0.463896142  0.4485551556
#>  [646,] -0.595902129  0.2673046229
#>  [647,]  0.572433440  0.8235056750
#>  [648,] -0.477757944 -0.1284689629
#>  [649,] -0.694993409  0.1645407447
#>  [650,]  1.037704943  0.8953929313
#>  [651,] -0.311990125  0.4597900234
#>  [652,] -0.194794011 -0.0008089577
#>  [653,] -0.037246421  1.0388071909
#>  [654,] -0.001908294  0.8788044154
#>  [655,] -0.168349650  1.2215498095
#>  [656,]  0.724895474  1.0630405913
#>  [657,]  0.532935249  1.3656297391
#>  [658,] -1.429810996 -0.3472037802
#>  [659,]  0.504691147  1.0829975138
#>  [660,] -0.282368112  0.2515082121
#>  [661,]  0.112898107  0.3890654460
#>  [662,]  0.303845948  1.0562483148
#>  [663,]  1.210738988  0.8320560366
#>  [664,]  0.235289464  0.7233090618
#>  [665,]  1.476219240  0.8039108641
#>  [666,]  0.045514567  0.4858592512
#>  [667,]  0.780353285  0.4640545693
#>  [668,]  0.591391916  0.9525645585
#>  [669,] -0.230017130  0.7594285019
#>  [670,] -0.799640392 -0.1115044073
#>  [671,] -0.490843678  0.1887419363
#>  [672,] -0.319654685  0.1870585700
#>  [673,]  0.572794190  1.2062707291
#>  [674,]  0.605174318  0.3529173977
#>  [675,]  0.331556334  1.0641174703
#>  [676,]  0.187522081  0.4719126003
#>  [677,] -0.029042604  0.2931287686
#>  [678,]  0.616993751  0.6210492092
#>  [679,]  0.360205786  0.7448210344
#>  [680,]  0.782178839  0.1571974537
#>  [681,] -0.959332110  0.2331859508
#>  [682,] -0.029808261  0.2570621039
#>  [683,] -0.158723572  0.3737268399
#>  [684,] -0.604375322 -0.1327187850
#>  [685,]  0.972773218  0.9100500035
#>  [686,] -0.334597558  0.4641467132
#>  [687,]  0.482176396  0.5768314843
#>  [688,] -0.082506046  0.1681108568
#>  [689,] -0.286137941  0.6209479923
#>  [690,]  0.641291067  0.9429919683
#>  [691,] -0.095109754  0.5738212318
#>  [692,] -0.631549932  0.1368495229
#>  [693,]  0.093550731  0.9563217057
#>  [694,] -0.499582548 -0.1599381694
#>  [695,]  0.557504186  0.5031080854
#>  [696,]  0.756207855  1.5486599319
#>  [697,]  0.982051652  1.0311859507
#>  [698,]  1.108239307  0.7776078958
#>  [699,]  0.401024513  0.9846979718
#>  [700,]  0.146423271  0.9865093756
#>  [701,] -1.065823858  0.0999999139
#>  [702,] -1.374104375  0.1457708028
#>  [703,] -0.891278591  0.1310459689
#>  [704,]  0.336164571  0.1852052765
#>  [705,]  0.039255850  0.7101247215
#>  [706,] -0.553020875  0.1589291120
#>  [707,]  0.575834360  0.2081236839
#>  [708,] -0.666866375 -0.2868171577
#>  [709,] -0.947945757  0.1250906287
#>  [710,] -0.730688287  0.6271317990
#>  [711,] -0.061007169  0.6399505983
#>  [712,] -0.335334057  0.5661864838
#>  [713,] -0.063857033  0.2188211846
#>  [714,]  1.053019952  0.8835565701
#>  [715,] -0.293762187  0.0529665821
#>  [716,] -0.091848478  0.6353063778
#>  [717,] -0.259382882  0.7800918098
#>  [718,]  0.351472268  0.9424133117
#>  [719,]  0.458422679  0.4996431745
#>  [720,]  0.003412652  0.4671516476
#>  [721,]  0.337868010  1.2760075983
#>  [722,] -0.820902522 -0.2610449454
#>  [723,] -0.829988841 -0.0268367344
#>  [724,] -0.962446791 -0.1800703042
#>  [725,]  0.225048718  0.2680684709
#>  [726,]  0.664653582  0.3749085007
#>  [727,]  0.142341936  0.6055538971
#>  [728,]  0.815934434  1.2203025572
#>  [729,]  1.548643852  1.6293073270
#>  [730,]  0.917307696  1.1854850103
#>  [731,] -0.367696225  0.7034680509
#>  [732,] -0.362385345  0.6310542302
#>  [733,] -0.149989895  0.6552982415
#>  [734,]  0.065361701  0.5522121276
#>  [735,] -1.102614588 -0.7196328113
#>  [736,] -0.600030911  0.1188964185
#>  [737,]  0.787422123  1.2366009449
#>  [738,] -0.207531992  0.6967682773
#>  [739,] -0.707613459  0.2563563859
#>  [740,] -0.283253247  0.1128408977
#>  [741,]  0.393451431  1.2019271042
#>  [742,]  0.172095831  0.2818882341
#>  [743,]  0.211037465  0.9563493124
#>  [744,]  0.717628964  0.7420442224
#>  [745,] -0.292548286  0.7420441625
#>  [746,] -0.300017679  0.3078014449
#>  [747,] -0.854841840 -0.1477726054
#>  [748,] -0.217598072  0.4726625160
#>  [749,] -0.543922329 -0.0692372944
#>  [750,]  0.386013390  0.4019339761
#>  [751,]  0.477940735  0.1516218924
#>  [752,]  0.005994196 -0.1000858904
#>  [753,] -0.991213211 -0.1227492233
#>  [754,] -0.930801416 -0.6621298621
#>  [755,] -0.533028857  0.6600862196
#>  [756,] -0.455440921  0.1481829685
#>  [757,] -1.351803401 -0.4885812105
#>  [758,] -0.097706612  0.6212980108
#>  [759,] -0.428290403  0.8145111260
#>  [760,]  0.042989322  0.0117858411
#>  [761,]  0.853675508  0.6459607253
#>  [762,] -0.075967822  0.6653566217
#>  [763,] -1.615311966 -0.3069773191
#>  [764,] -0.126875586  0.4659931156
#>  [765,]  0.502863190  1.0237827597
#>  [766,]  0.768004607  1.0646958639
#>  [767,]  1.341166131  1.0378125553
#>  [768,]  0.370064871 -0.1885867852
#>  [769,] -0.757542517 -0.0662822670
#>  [770,] -0.376349892  0.5893910143
#>  [771,]  0.600693523  1.1507949151
#>  [772,] -0.380969292  0.0996021613
#>  [773,]  0.451859204  1.0359726178
#>  [774,] -0.455001257  0.5636818484
#>  [775,] -0.429521411  0.6219338623
#>  [776,]  0.989265138  1.4334217233
#>  [777,]  1.432390688  1.1423857590
#>  [778,] -0.597982033  0.8815134877
#>  [779,] -0.671517217 -0.5304551711
#>  [780,] -0.599171692  0.1986296477
#>  [781,]  0.354951769  0.5660111204
#>  [782,] -0.783636148  0.2565771158
#>  [783,] -0.274339362  0.5712125945
#>  [784,]  0.663878217  0.8691434989
#>  [785,] -0.883126519 -0.3883492340
#>  [786,]  0.535978546  1.2729448006
#>  [787,] -0.119351408  0.4434059697
#>  [788,] -0.210452964  0.4396973302
#>  [789,]  1.126865919  0.9668000296
#>  [790,] -0.998595201 -0.0619830711
#>  [791,] -0.384643135  0.2567556096
#>  [792,] -0.297475937 -0.0066960159
#>  [793,]  0.614535530  0.8760494534
#>  [794,]  0.136237718  0.4588955405
#>  [795,]  0.499903681  1.1406889454
#>  [796,] -0.918080262  0.2684024616
#>  [797,] -0.383990353  0.3978458617
#>  [798,] -0.640646083  0.2122403483
#>  [799,]  0.304238049  0.1160617955
#>  [800,]  0.100207083  1.2814488586
#>  [801,]  0.204715862  0.6869138220
#>  [802,]  0.168069948  0.5602010017
#>  [803,] -0.534962881  0.4097638170
#>  [804,]  0.797450579  1.1856114952
#>  [805,]  0.845743428  1.1321949385
#>  [806,]  0.161439510  0.8185739704
#>  [807,] -0.123094918  0.5395190099
#>  [808,]  0.534254678  0.7182287834
#>  [809,] -0.900758166 -0.3468257887
#>  [810,]  0.492800782  0.1686342334
#>  [811,]  0.990628393  1.0158551565
#>  [812,] -0.556264836  0.0409962576
#>  [813,] -0.486229056  0.2535255688
#>  [814,] -0.794142637 -0.1078600285
#>  [815,] -1.185067086 -0.1146358765
#>  [816,] -0.001376383  0.7837212627
#>  [817,]  0.946254329  1.0913878684
#>  [818,] -0.888140030  0.0125643146
#>  [819,]  1.264445064  1.3413164261
#>  [820,]  0.369731260  0.4692679683
#>  [821,]  1.503900103  1.3432901531
#>  [822,] -0.609708226  0.9669448243
#>  [823,]  1.096449697  1.1489286661
#>  [824,]  0.065441432  0.6962180544
#>  [825,] -0.690287217  0.2087754641
#>  [826,] -0.749605386  0.0530360352
#>  [827,] -0.772143816 -0.4862787750
#>  [828,]  1.088893463  1.2822713624
#>  [829,]  1.743795945  1.9285559871
#>  [830,]  0.091518950 -0.0549030835
#>  [831,] -0.831259198  0.2683520713
#>  [832,]  0.493822078  0.7153613483
#>  [833,]  0.108728634  0.9185346882
#>  [834,] -0.001724086  0.8759571417
#>  [835,] -0.322977967  0.6831312362
#>  [836,]  0.985886063  1.1839660420
#>  [837,] -0.286185150 -0.1521277948
#>  [838,]  1.719643306  1.3761776094
#>  [839,] -0.180050574  0.8919823713
#>  [840,]  0.160585501  0.9629685861
#>  [841,]  0.242002098  0.5401025612
#>  [842,] -0.324491582  0.1404558606
#>  [843,]  0.035901345  0.7584145255
#>  [844,] -0.554743314  1.0361496050
#>  [845,] -0.031368238  0.3805153939
#>  [846,] -0.224013559  0.4405077037
#>  [847,] -1.015480679 -0.2804698558
#>  [848,] -0.307148307  0.1232715468
#>  [849,] -0.117555737  0.7819429976
#>  [850,]  0.236376885  0.3477259893
#>  [851,]  0.058477153 -0.0042096701
#>  [852,] -0.277616305  0.1360636371
#>  [853,]  0.704995992  0.2508791215
#>  [854,] -0.530886400 -0.0909035121
#>  [855,]  0.467378949  0.8307433391
#>  [856,]  1.004658144  0.8370709996
#>  [857,] -0.211717948 -0.2858041511
#>  [858,]  0.226361251  0.7803720297
#>  [859,]  0.616913079 -0.1052044811
#>  [860,]  0.457143947  0.3285247459
#>  [861,] -1.152288578  0.4986759516
#>  [862,]  0.414815888  0.5141569893
#>  [863,] -0.870850636  0.1568367856
#>  [864,]  0.051128785  0.1240863953
#>  [865,]  1.166834546  0.7146856080
#>  [866,]  0.911572012  1.0226200036
#>  [867,]  1.513478718  1.3397961385
#>  [868,]  0.229764494  0.2762783198
#>  [869,] -0.485460646  0.2167522834
#>  [870,]  0.220631856  0.8733309653
#>  [871,] -1.372134298  0.0069809019
#>  [872,] -1.436398297 -0.2933152736
#>  [873,] -0.546621880  0.7264150978
#>  [874,]  0.382494202  0.7528334636
#>  [875,]  0.683881022  0.6765978674
#>  [876,] -0.186979562  0.1896814647
#>  [877,]  1.114980848  0.9650036492
#>  [878,]  0.164682180  0.8420600725
#>  [879,]  0.307414850  0.2037227403
#>  [880,]  1.194433138  1.6442713915
#>  [881,]  0.796632322  0.7436018345
#>  [882,] -0.388309546  0.4818090765
#>  [883,] -0.013487529  0.6185573808
#>  [884,]  0.239370495  0.2360006391
#>  [885,]  0.271391014  0.1105591802
#>  [886,]  0.674585497  0.5148956153
#>  [887,]  0.824303628  0.8844599126
#>  [888,]  0.205787337  1.3677410105
#>  [889,]  1.142719744  1.3854986198
#>  [890,]  0.284555678  0.6979494153
#>  [891,]  1.468927876  1.2486539691
#>  [892,] -2.090422482 -0.3506170714
#>  [893,] -0.960683723 -0.2748792744
#>  [894,] -0.199313768  0.7701928551
#>  [895,] -0.441221800  0.0739489908
#>  [896,]  0.838348234  1.0990967697
#>  [897,]  0.031280752  0.7010861520
#>  [898,]  0.240184023  1.0515225228
#>  [899,] -0.713825427 -0.2714578805
#>  [900,] -0.210277558  0.4550341857
#>  [901,]  1.310629178  1.0039777718
#>  [902,]  0.794052132  0.5315421145
#>  [903,] -0.015682917  0.7483561744
#>  [904,]  0.165798357  0.7310051534
#>  [905,] -0.057266176  0.8582086970
#>  [906,]  1.007722891  0.7715753567
#>  [907,]  1.231081600  1.4354132968
#>  [908,] -0.104089764  0.8138475162
#>  [909,]  0.010027253  0.4336773570
#>  [910,] -0.039177111 -0.1464626003
#>  [911,] -0.860869735 -0.2328399932
#>  [912,]  0.145480520  0.9054986600
#>  [913,] -0.452594014  0.1844494709
#>  [914,] -0.012925685  1.1585687446
#>  [915,]  0.585983206  0.3495109769
#>  [916,]  0.532443485  0.6910864772
#>  [917,]  0.606949314  0.8808084412
#>  [918,] -0.289690770  0.8929614547
#>  [919,] -0.184975597  0.5137474543
#>  [920,]  0.519892017  0.5954250748
#>  [921,]  0.570304680  0.7022792510
#>  [922,] -1.884652951 -0.6287447910
#>  [923,]  0.889908129  1.1212031223
#>  [924,]  0.545495797  0.6114845090
#>  [925,]  0.633990467  0.9090819843
#>  [926,]  0.351274874  1.1151284030
#>  [927,] -0.158593548  0.1848072610
#>  [928,] -0.376218901  0.5617801573
#>  [929,]  0.810252580  0.7032017643
#>  [930,]  0.031694531  0.1182823041
#>  [931,]  0.285804928  0.1607961787
#>  [932,]  1.060386477  0.9730603113
#>  [933,] -0.972575633  0.3683945197
#>  [934,] -0.609373374 -0.1736196050
#>  [935,]  0.189112704  0.8714848307
#>  [936,]  0.476897516  0.9258425904
#>  [937,]  1.140687320  1.4528514171
#>  [938,]  0.588924710  1.2768824496
#>  [939,] -0.821506565  0.1972471572
#>  [940,]  0.798460711  1.4835489992
#>  [941,]  0.962027799  1.3310119171
#>  [942,] -0.952529388  0.8447651780
#>  [943,] -0.681422353 -0.0415957832
#>  [944,]  0.208991379  0.4443572689
#>  [945,] -1.153128398 -0.0320883660
#>  [946,] -0.006673942  0.8389449529
#>  [947,] -0.834437697  0.4182020232
#>  [948,] -0.063130984 -0.0300262727
#>  [949,]  0.781045809  1.2057191964
#>  [950,]  0.430105388  1.0080133091
#>  [951,]  1.356392980  1.8078423382
#>  [952,]  0.220643876  0.3374940124
#>  [953,] -0.507907184 -0.9066575535
#>  [954,]  1.174724427  0.8194312559
#>  [955,] -1.468134002  0.6177218045
#>  [956,]  0.774276559  0.3910151457
#>  [957,]  0.292839257  0.4943284795
#>  [958,] -0.150690151  0.2242194562
#>  [959,]  0.298622086  0.9516688994
#>  [960,]  0.023312211  0.9171005392
#>  [961,]  0.209080246  0.3687324270
#>  [962,] -0.494791786  0.0467657107
#>  [963,] -0.055734532  1.0101637291
#>  [964,] -1.684920983 -0.0968475237
#>  [965,] -0.519762110 -0.3390186108
#>  [966,] -0.424596083  0.0430592434
#>  [967,]  0.753399460  1.1396784895
#>  [968,] -0.110457097  0.1742632262
#>  [969,]  0.365830171  0.7300699252
#>  [970,] -1.109843579  0.3199737922
#>  [971,]  0.081618596  0.3234175825
#>  [972,]  0.382890978  0.4893537845
#>  [973,]  0.843921964  0.2037979839
#>  [974,]  0.727565755  1.6188499837
#>  [975,] -1.008372657 -0.0909021251
#>  [976,]  0.132535300  0.8195582322
#>  [977,] -0.032581182  0.6146699456
#>  [978,] -0.738977160 -0.2226793993
#>  [979,]  0.952210914  1.0310353947
#>  [980,]  0.558302690  0.4797905333
#>  [981,]  0.437525492  0.8784284647
#>  [982,]  1.240000254  1.3518891324
#>  [983,]  0.127208156  0.6742213706
#>  [984,]  0.178867463  0.7269615383
#>  [985,]  0.485273515  0.0985625037
#>  [986,] -0.671664787  0.8796439433
#>  [987,]  0.430499917  0.5114198681
#>  [988,] -0.101690903  0.5738510857
#>  [989,]  0.565497927  0.7277353079
#>  [990,]  1.441265937  1.4973807733
#>  [991,]  0.162087417  0.5961278997
#>  [992,] -0.407473256 -0.1870375944
#>  [993,]  0.593961884  1.0370027539
#>  [994,]  0.501743178  0.4771987755
#>  [995,]  0.520829630  1.0783818425
#>  [996,]  1.156295703  0.9287589779
#>  [997,]  0.179156859  0.4654586821
#>  [998,] -1.192646079 -0.0050725236
#>  [999,] -0.753983559  0.1586586811
#> [1000,]  0.607320690  1.2815832427
#> attr(,"n")
#> [1] 1000
#> attr(,"A")
#>      [,1] [,2] [,3]
#> [1,]    0    1    1
#> [2,]    0    0    0
#> [3,]    0    0    0
#> attr(,"S")
#>      [,1] [,2] [,3]
#> [1,]    0 0.00 0.00
#> [2,]    0 0.25 0.00
#> [3,]    0 0.00 0.25
#> attr(,"u")
#> [1] -0.5  0.5  0.0
#> attr(,"filter")
#>   y x e
#> y 1 0 0
#> x 0 1 0
#> attr(,"v")
#>        v
#> [1,] 0.0
#> [2,] 0.5
#> [3,] 0.0
#> attr(,"g")
#>     g
#> y 0.0
#> x 0.5
#> attr(,"C")
#>      [,1] [,2] [,3]
#> [1,] 0.50 0.25 0.25
#> [2,] 0.25 0.25 0.00
#> [3,] 0.25 0.00 0.25
#> attr(,"M")
#>      y    x
#> y 0.50 0.25
#> x 0.25 0.25