Generate Data from a Multivariate Normal Distribution Using the Reticular Action Model (RAM) Notation

Generates data from a multivariate normal distribution using the Reticular Action Model (RAM) notation with the MASS::mvrnorm() function.

ram2dat(n, A, S, u, filter, ...)

Arguments

n
Integer. Sample size.

A
t x t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
t x t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

u
vector of length t or t by 1 matrix. Mean structure parameters.

filter
p x t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

...
Additional arguments to pass to MASS::mvrnorm().

Details

The Multivariate Normal Distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbf{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{g} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} = \mathbf{F} \mathbf{E} \mathbf{u} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{M} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{C} \mathbf{F}^{\mathsf{T}} $$

Author

Ivan Jacob Agaloos Pesigan

Examples

# This is an example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

n <- 1000
A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, ] <- c(0, 1, 1)
diag(S) <- c(0, 0.25, 1)
u <- c(0.00, 0.50, 0.00)
filter <- diag(2)
filter <- cbind(filter, 0)
colnames(filter) <- c("y", "x", "e")
rownames(filter) <- c("y", "x")
ram2dat(n, A, S, u, filter)

#>                    y             x
#>    [1,]  0.886767709  0.9998511768
#>    [2,] -0.702716262 -1.1800328053
#>    [3,]  0.471511705  0.9735339441
#>    [4,]  0.732752589  0.8693141934
#>    [5,]  3.204493294  1.2275513806
#>    [6,]  0.501743258  0.1665052889
#>    [7,]  1.718171312  0.8100943764
#>    [8,]  1.336901368  0.1079274232
#>    [9,]  2.806549249  1.0058799415
#>   [10,] -2.512869835  0.0715787989
#>   [11,]  1.850100778  0.5245706557
#>   [12,]  0.881253072  0.5812933955
#>   [13,] -0.305718119 -0.0028489385
#>   [14,]  2.116439143 -0.0204492647
#>   [15,] -0.109404432  0.3817239715
#>   [16,]  0.440507440  0.3780024242
#>   [17,] -0.124112846  0.3624962394
#>   [18,] -0.045979705  0.7441677824
#>   [19,]  2.257975985  0.2852516310
#>   [20,] -0.620729009  0.5604692511
#>   [21,]  1.011008503  0.0231649348
#>   [22,]  0.502690695  0.8649464107
#>   [23,]  0.635463622 -0.2740428068
#>   [24,]  0.430529931  0.1752859117
#>   [25,]  1.620745045  0.7244387931
#>   [26,] -0.840813576  0.0377458373
#>   [27,]  0.128092509  0.3801807176
#>   [28,]  1.049761337  0.5251358060
#>   [29,] -0.089764585 -0.5311631878
#>   [30,]  2.106967370  1.3066803458
#>   [31,]  0.179532115  0.6751938248
#>   [32,]  1.987525387  0.7362892230
#>   [33,] -0.351716535  0.3878787432
#>   [34,]  2.282942724  0.7283573613
#>   [35,]  0.427986645  0.4512221929
#>   [36,]  1.057843619  1.3216895631
#>   [37,]  1.486721908  0.5290817620
#>   [38,] -0.164101414 -0.0494024972
#>   [39,] -0.914603080  0.7487779860
#>   [40,]  2.269616361  1.4022353323
#>   [41,]  1.391353808  1.2199994634
#>   [42,] -0.964936079  0.1440210459
#>   [43,] -0.727234114  0.2199260151
#>   [44,] -0.868405533  0.3021398995
#>   [45,]  2.091312971  0.8865273881
#>   [46,] -0.957057126 -0.2144217258
#>   [47,]  0.421196581  0.8182500921
#>   [48,]  0.640566845  0.2924098304
#>   [49,] -0.149447366  1.0528804232
#>   [50,]  0.257399302  0.8856632509
#>   [51,]  1.623613124 -0.1254034474
#>   [52,]  0.874869983  1.4198155150
#>   [53,]  1.737791063  1.2007231795
#>   [54,]  1.141120505  1.5275991829
#>   [55,] -1.201455551  0.4576873484
#>   [56,]  1.234969019  1.5128983060
#>   [57,]  0.101969735  0.7430653948
#>   [58,]  0.975213824  0.8775306337
#>   [59,]  1.098976381  0.9746224761
#>   [60,] -1.056224009 -0.1787809791
#>   [61,]  0.327359067  0.4859459921
#>   [62,]  1.908880794  1.6695326842
#>   [63,]  0.383504943  1.0447894093
#>   [64,]  0.557428339  1.3136219325
#>   [65,]  1.458158941  0.9545633039
#>   [66,]  2.666388692  1.7913219394
#>   [67,] -1.604223306 -0.0149169779
#>   [68,]  1.565555636  0.8073036145
#>   [69,]  0.552726218  0.7875904052
#>   [70,]  0.248932519  0.3685890605
#>   [71,]  0.365683398  0.7657101568
#>   [72,]  2.891145524  1.2148306517
#>   [73,]  0.693793118  0.2659987176
#>   [74,]  0.478194492  0.0331644153
#>   [75,]  0.377300458  0.9804919389
#>   [76,] -1.498835426 -0.3835797515
#>   [77,] -1.781930792 -0.5837545273
#>   [78,] -0.359150269  0.6051849970
#>   [79,] -0.463949438  0.7639341469
#>   [80,]  0.879833903  0.4258279427
#>   [81,] -0.728091288  0.6627148447
#>   [82,]  1.579103803  0.5055329256
#>   [83,] -0.083549428  0.1664368686
#>   [84,]  0.151490262  0.3704124530
#>   [85,]  0.066225367  0.5115350943
#>   [86,] -0.856135226  0.6183317820
#>   [87,]  1.587955228  0.5798165846
#>   [88,]  0.359336277  0.0110739471
#>   [89,] -1.012684188 -0.1406322332
#>   [90,]  0.878757084  0.8064717592
#>   [91,]  0.540595386  1.1949499095
#>   [92,]  0.808850634  0.2959303034
#>   [93,]  1.747554168  0.2425234072
#>   [94,] -0.019486268  0.3066651133
#>   [95,] -1.145329045 -0.5198964278
#>   [96,]  2.226845682  0.7294782547
#>   [97,]  2.109540530  0.7333879345
#>   [98,] -0.622832299  0.8902212230
#>   [99,] -0.275235488  0.6758977795
#>  [100,] -1.804863359 -0.2551533262
#>  [101,]  0.781065675  1.0828138373
#>  [102,]  1.658028455  0.8953279592
#>  [103,] -0.653458889  0.0223090210
#>  [104,] -0.817406403  0.1966171225
#>  [105,]  0.224707991  0.6782033920
#>  [106,]  1.033794464  0.2327470528
#>  [107,] -0.111713495  0.0229369747
#>  [108,] -1.714803949  0.0511610842
#>  [109,] -1.674003440  0.3239913776
#>  [110,]  2.143109846  1.4506821478
#>  [111,]  0.997169243  1.0826648093
#>  [112,]  1.799258677  0.8199321093
#>  [113,] -0.150179592  0.4699708669
#>  [114,]  2.574615030  0.7717688831
#>  [115,]  0.154075280  0.0065876638
#>  [116,]  0.354148830  0.9080161943
#>  [117,] -0.963299328  0.8241478017
#>  [118,]  0.868377694  0.2189422381
#>  [119,]  1.777999131  0.3466476831
#>  [120,]  0.384446310 -0.0003767599
#>  [121,] -0.231551598  0.3593775509
#>  [122,]  0.212677746  0.0044981421
#>  [123,] -1.183815967  0.2749485601
#>  [124,]  0.377852718  0.6587508621
#>  [125,] -1.915048395  0.1410479751
#>  [126,]  1.974065663  1.6069179342
#>  [127,] -0.049123881  0.7949243352
#>  [128,] -1.086657449  0.2794629516
#>  [129,]  0.134815893  0.2528320767
#>  [130,] -0.224013980  0.3074453376
#>  [131,] -0.558116532 -0.2572664010
#>  [132,]  2.651289731  1.2447722850
#>  [133,] -0.797405046 -0.0216185230
#>  [134,]  0.830824176 -0.0584133182
#>  [135,]  0.163437972  0.4565114803
#>  [136,] -0.949669591  0.3347804239
#>  [137,]  1.816962160  0.5750372199
#>  [138,]  1.679530486  1.1358465021
#>  [139,]  0.402129649  0.3668407819
#>  [140,] -0.527972291  0.6072650691
#>  [141,]  0.820672796  0.3085631165
#>  [142,] -1.712222892 -0.5201580193
#>  [143,]  0.528831062  0.9302255659
#>  [144,]  2.121276525  0.5682574259
#>  [145,]  1.729319687  0.9808938897
#>  [146,] -1.512887264  0.5381225393
#>  [147,]  0.612602033  0.4217605427
#>  [148,]  0.967919384  0.6954332997
#>  [149,]  0.465731974  1.2388286070
#>  [150,] -0.183392982 -0.5215259493
#>  [151,]  1.062644141  0.9996307202
#>  [152,]  0.700455916  0.3750514182
#>  [153,] -0.132268883  0.5566802751
#>  [154,] -1.502228359  0.3487091098
#>  [155,] -1.069974395 -0.0634999926
#>  [156,] -1.307992809  0.5984874050
#>  [157,]  1.060343574  0.0386245362
#>  [158,] -1.397383138 -0.5612677367
#>  [159,]  1.089021402  0.7511820234
#>  [160,]  1.421386288  0.9082215656
#>  [161,]  0.993151559  0.1231872991
#>  [162,] -0.616165827 -0.0647542248
#>  [163,] -0.402245591  0.2731092017
#>  [164,] -1.270396115 -0.9814619637
#>  [165,]  0.876623596 -0.5721824336
#>  [166,] -0.472257969 -0.2738670363
#>  [167,]  2.282163202  1.4825340485
#>  [168,]  1.466656354  0.5484659593
#>  [169,] -0.631565738 -0.1868363111
#>  [170,]  0.646092782  0.7467298120
#>  [171,]  0.459833455 -0.2202992587
#>  [172,]  1.337730557  0.4293529902
#>  [173,]  1.616273409  0.5542313872
#>  [174,]  1.511410770  1.0730035441
#>  [175,]  0.973305099  0.1810895353
#>  [176,] -0.767831690  0.1321672175
#>  [177,] -0.028335806  0.7207449042
#>  [178,]  1.164253165  0.0312766759
#>  [179,]  0.740616778  0.1442939391
#>  [180,] -0.763909936 -0.2546201807
#>  [181,]  0.182941863  0.1850864574
#>  [182,]  0.787325142  1.0113011192
#>  [183,]  0.728998930  0.0207073533
#>  [184,]  0.409323231  0.7109850508
#>  [185,]  0.043979912  0.5088980742
#>  [186,]  0.443864521  0.6340584265
#>  [187,]  0.498925031  0.4667730346
#>  [188,] -0.561396923  0.3607356056
#>  [189,]  1.124536653  1.2643356934
#>  [190,]  0.120236903  0.4285113977
#>  [191,]  1.001246923  0.4290458850
#>  [192,] -0.520073517  0.1686703880
#>  [193,] -1.719630563  0.7381029551
#>  [194,]  1.349317579  0.4120915444
#>  [195,]  0.390595890  0.9210636228
#>  [196,] -1.668331310 -0.4446529921
#>  [197,]  1.827713802  1.5570377590
#>  [198,] -0.683021720 -0.2601201338
#>  [199,]  0.715329631  0.5997883163
#>  [200,]  1.111630136 -0.2084015338
#>  [201,] -1.547352432  0.5272442615
#>  [202,]  1.413111815  0.8411584562
#>  [203,]  1.480949933  1.1330598251
#>  [204,] -0.619267292  0.1743938806
#>  [205,]  2.090659840  1.2119000831
#>  [206,]  1.871345190  0.5910448025
#>  [207,] -0.221088042  0.4161389944
#>  [208,] -1.836886390 -0.0646004553
#>  [209,]  1.219410553  0.5088124544
#>  [210,]  2.317285199  0.9420224231
#>  [211,]  0.626795926  0.0148784770
#>  [212,]  1.268540790  0.5136759864
#>  [213,]  1.278001546  0.5384728997
#>  [214,]  0.949598118  0.9787478086
#>  [215,] -0.829224198 -0.4399399795
#>  [216,] -0.228055120  0.4141379252
#>  [217,] -0.045374235  0.2489946077
#>  [218,]  2.341580110  1.4204332479
#>  [219,] -0.728920452  1.2558567163
#>  [220,]  0.196457539  0.6751407772
#>  [221,] -0.154036227  0.3781934373
#>  [222,]  0.801654710  0.1076846442
#>  [223,] -0.303477419  0.2137214939
#>  [224,]  1.561034621  0.9724488428
#>  [225,]  3.924426924  0.6294051452
#>  [226,] -0.358817124  0.5574559955
#>  [227,]  1.445398475 -0.0288490947
#>  [228,]  2.508635991  1.4236427761
#>  [229,]  0.537805882 -0.0401030221
#>  [230,] -0.366514499 -0.1882439479
#>  [231,] -0.066787046  0.6409803599
#>  [232,]  1.498179533  0.5080407744
#>  [233,] -0.084281270  0.1554863315
#>  [234,] -0.633965451 -0.4342264045
#>  [235,]  1.430592296  0.3236568785
#>  [236,]  0.990909193  0.8568822609
#>  [237,]  1.927495289  0.2597363460
#>  [238,]  1.093142574  0.3760615306
#>  [239,]  0.240178808  1.0262095730
#>  [240,]  2.500551965  0.8300472755
#>  [241,]  2.187329281  0.3621139240
#>  [242,] -0.198005672 -0.0747649909
#>  [243,]  1.802704429  1.1608009041
#>  [244,]  0.868878842  0.7715994138
#>  [245,]  0.033296891  0.1759212305
#>  [246,]  0.878146955  0.5057869610
#>  [247,]  1.800325682  1.4017999125
#>  [248,]  1.764679417  0.2721229094
#>  [249,] -1.162720035  0.0035383130
#>  [250,]  0.162220356  0.2033482431
#>  [251,] -1.535186417  0.2599311265
#>  [252,]  0.207852841 -0.1877881271
#>  [253,]  1.392186511  1.2955526898
#>  [254,]  2.327927927  0.6464464542
#>  [255,]  0.693972036  0.0969174520
#>  [256,]  2.787848633  1.1917465286
#>  [257,]  0.487388718  0.5356543765
#>  [258,]  0.723577536  1.3245766785
#>  [259,] -0.869779452  0.5938405148
#>  [260,]  0.827994673  0.2120657304
#>  [261,] -0.580904077  0.0084587514
#>  [262,]  0.895262049  0.1389115331
#>  [263,] -1.263338719 -0.2895805986
#>  [264,]  1.403640125  0.5269136516
#>  [265,]  1.462425966  1.0241208713
#>  [266,]  0.861914711  0.3390048524
#>  [267,]  1.106604657  0.8853911663
#>  [268,]  0.618965993  0.6050108277
#>  [269,] -0.890693595  0.5159870640
#>  [270,] -0.274709183  0.5894850456
#>  [271,]  1.366456717 -0.0351468581
#>  [272,]  1.483874272  0.3545975419
#>  [273,]  1.492124328  0.5823587038
#>  [274,]  2.874830211  1.2424251932
#>  [275,]  2.122382523  0.4788633231
#>  [276,]  1.168660806  0.1232067651
#>  [277,]  1.142905855  0.6731158633
#>  [278,]  0.765224371  0.2092594495
#>  [279,]  0.075980443 -0.1239712455
#>  [280,]  1.957465371  0.9661174253
#>  [281,]  0.760205309  0.7022881415
#>  [282,] -0.255188993  0.4502808389
#>  [283,] -0.414532253  0.6448753762
#>  [284,] -0.507828457  0.3045420193
#>  [285,] -1.165795085  0.1400080762
#>  [286,]  1.305995510  0.3694434241
#>  [287,] -0.085751176  0.3057208691
#>  [288,] -0.201391964 -0.2477833146
#>  [289,] -1.723481557  0.4755457094
#>  [290,]  2.709374953  1.3996706414
#>  [291,]  0.277089619  1.1228481819
#>  [292,]  0.766728095  0.6905017863
#>  [293,]  1.002429752  0.5993581769
#>  [294,]  2.069948329  0.8745957944
#>  [295,]  0.767800868  1.8874403336
#>  [296,]  0.802428306  0.5622522129
#>  [297,]  1.698453120  0.8271218523
#>  [298,]  0.155879802  0.1542377747
#>  [299,]  0.831028530  0.7693196280
#>  [300,] -0.954895241 -0.1284338169
#>  [301,]  1.234206567  0.9330965256
#>  [302,] -0.254326440  0.3978623766
#>  [303,] -0.016880167  1.2397702394
#>  [304,]  0.125521424  1.1005261312
#>  [305,]  1.728262166  1.2905123407
#>  [306,]  0.704766338 -0.1691233502
#>  [307,] -0.329206299  0.5915413256
#>  [308,] -0.128021679 -0.3517436964
#>  [309,] -0.749490040  0.8557468704
#>  [310,] -0.853611470  0.5358630990
#>  [311,]  1.828894084  1.0402353015
#>  [312,] -2.415927045 -0.3687884301
#>  [313,]  0.772263608  0.1280885285
#>  [314,]  0.913361700  0.8578663734
#>  [315,]  0.881976043 -0.1301492928
#>  [316,]  1.519066136  0.6391005314
#>  [317,] -0.558613056  0.4812806089
#>  [318,]  0.328596260  0.8294410321
#>  [319,]  1.935301913  0.6886670167
#>  [320,] -0.651504157  0.2701467114
#>  [321,]  1.344263501  0.5627711977
#>  [322,]  0.663886058  0.7526534640
#>  [323,] -0.113128564  0.4636156919
#>  [324,] -0.029918484  0.0048341890
#>  [325,] -0.268278038  1.0171618216
#>  [326,]  1.243799286  0.6336025015
#>  [327,]  0.399216174  0.6409787220
#>  [328,]  3.613772270  1.2931700934
#>  [329,]  1.556761162  0.5663558727
#>  [330,]  1.312558185  0.9540373649
#>  [331,]  1.220068228  0.7568494628
#>  [332,] -0.224470202  0.4500097803
#>  [333,]  1.072224885  0.6020093643
#>  [334,]  0.234816777  0.2769897128
#>  [335,]  0.034723016  0.1430052446
#>  [336,] -0.302324474  0.2933534046
#>  [337,] -0.345128023  1.0232366734
#>  [338,]  1.082610580  1.1088849102
#>  [339,]  0.460255133 -0.4308031533
#>  [340,] -0.738249502  0.1284640598
#>  [341,] -1.406022580 -0.4565924656
#>  [342,]  0.637017930  0.3370175927
#>  [343,] -0.967857739 -0.0929558001
#>  [344,]  1.263663210  0.3895674986
#>  [345,]  0.829422849  0.1833115182
#>  [346,]  1.296548640  0.5305215691
#>  [347,]  1.000191980  0.5956819193
#>  [348,] -0.091422310 -0.0148907510
#>  [349,] -0.223324886  0.4557486531
#>  [350,]  1.194884744  1.0195262626
#>  [351,]  1.372760981  0.7052883498
#>  [352,]  1.442084211  1.2556301642
#>  [353,] -1.459874441 -0.1253991764
#>  [354,]  0.056242296  0.6419548457
#>  [355,]  0.281059834  0.1371613739
#>  [356,]  1.391331550  0.7377354569
#>  [357,]  0.262502304  0.9798676680
#>  [358,]  0.859429251  0.6386918096
#>  [359,] -1.399322663  0.3695910205
#>  [360,]  1.054677420  1.3568689296
#>  [361,] -0.103871050  0.1347958375
#>  [362,]  0.797978716 -0.0132813152
#>  [363,]  2.242773955  1.6370358250
#>  [364,] -1.219022250  0.6556440823
#>  [365,]  3.604342262  0.5224141312
#>  [366,]  1.775419289  0.2386650365
#>  [367,]  1.266750924  0.7750298679
#>  [368,]  1.215922410  0.9559503589
#>  [369,] -0.058892107  0.1889448921
#>  [370,]  0.986982853  0.0123774117
#>  [371,]  1.431157677  0.7367032939
#>  [372,]  2.063322152  0.6249272073
#>  [373,]  1.585103510 -0.0252485535
#>  [374,]  1.173837557  0.8047612631
#>  [375,] -0.453505738  0.4758529373
#>  [376,]  0.204609086  0.2434349698
#>  [377,] -1.642332808  0.5761436309
#>  [378,]  1.564053681  0.4446676452
#>  [379,]  0.601591508  0.3008545134
#>  [380,]  1.169215828  0.0115325156
#>  [381,]  1.207706554  0.1249290845
#>  [382,]  0.629882513  0.4235613317
#>  [383,]  3.248624521  1.2830487994
#>  [384,]  0.453189983 -0.1327839346
#>  [385,]  0.039430069  0.7056247644
#>  [386,] -0.909601390 -0.6683720590
#>  [387,] -0.171725396 -0.4685856299
#>  [388,] -0.936207878  0.7698587763
#>  [389,]  0.620885945  0.6698693373
#>  [390,]  0.971011809  1.0688693871
#>  [391,] -1.253733695  0.3609168575
#>  [392,]  0.836756706  0.7092974298
#>  [393,]  0.711073825 -0.1268498873
#>  [394,]  1.164415360  0.4917294132
#>  [395,]  1.326343181  0.7040694126
#>  [396,]  0.523029905  0.9454489977
#>  [397,]  0.983832927  1.2432662907
#>  [398,]  0.435744944 -0.1640206823
#>  [399,] -0.376377751  0.4832703721
#>  [400,]  2.560424131  1.3158742540
#>  [401,]  0.165967498  0.7392750980
#>  [402,] -0.052242218 -0.2927244856
#>  [403,]  0.119476164  0.0307399884
#>  [404,] -1.571228485 -0.5699904206
#>  [405,]  1.093695159  0.5062697210
#>  [406,]  1.695235644  0.9643196487
#>  [407,]  1.542462274  0.8084204466
#>  [408,] -0.053322907  0.4693285442
#>  [409,]  1.560893809 -0.0340762890
#>  [410,]  0.230869522  0.4731545942
#>  [411,]  1.873683825  0.7685397172
#>  [412,]  3.271559229  0.9781513806
#>  [413,]  2.196311812  0.7305237452
#>  [414,]  0.997888560  0.5429180191
#>  [415,]  1.665355985  0.2572000339
#>  [416,]  0.636718266  0.5256524043
#>  [417,] -0.986002596  0.1395748333
#>  [418,]  1.375050213  1.0776265769
#>  [419,] -0.237441626  0.3046301102
#>  [420,]  2.053472549  0.9983385330
#>  [421,] -0.681623526 -0.3212041106
#>  [422,]  2.067385752  0.8556400679
#>  [423,] -0.708657050  0.1336712279
#>  [424,] -0.269471763  0.2837050982
#>  [425,]  1.518533729  0.4897080688
#>  [426,]  1.698616966  0.8578773619
#>  [427,]  0.518366857  0.6067055028
#>  [428,] -0.377643097  0.1838527209
#>  [429,]  0.962815433  1.0261707320
#>  [430,] -0.403493207 -0.3937458083
#>  [431,]  0.612536570  0.8909968585
#>  [432,]  1.863468560  0.2330844951
#>  [433,]  1.635257345  0.4017855973
#>  [434,]  0.190145974  0.3729553609
#>  [435,] -0.387920403 -0.1055578850
#>  [436,]  1.598719262  1.5670584163
#>  [437,]  1.670657029  1.2060973462
#>  [438,]  2.052799569  1.0794318302
#>  [439,]  0.863435915  1.4211947468
#>  [440,] -0.410774641  0.1710813324
#>  [441,] -0.402913355  0.3831234375
#>  [442,]  0.450036320  1.5524829561
#>  [443,] -0.181948960  0.3375507201
#>  [444,]  1.425944360  0.4178513150
#>  [445,] -0.466728818  0.1699373791
#>  [446,]  0.425858489  0.7482630749
#>  [447,]  0.208374833  0.6131990864
#>  [448,]  1.454768759  0.0903638904
#>  [449,]  1.885519970  0.7185801417
#>  [450,] -0.097175595  1.0038694196
#>  [451,]  1.304681509  0.8312025189
#>  [452,]  0.632270179  0.7470019198
#>  [453,]  0.191995664  0.0517884865
#>  [454,]  0.816614980  0.5982489524
#>  [455,]  0.479038877  0.0675961439
#>  [456,] -0.151881872 -0.3249664927
#>  [457,]  1.145848688  0.8070262371
#>  [458,]  1.797406470  0.5163048968
#>  [459,] -0.249584869  1.0120806932
#>  [460,] -0.259413184  0.0514571298
#>  [461,]  1.810634997  1.0358702057
#>  [462,] -0.678218711  0.4045740973
#>  [463,] -0.064249542  0.5737715548
#>  [464,] -0.157814844 -0.6498448687
#>  [465,]  0.794352044 -0.0982030599
#>  [466,]  1.500233990  0.8955725369
#>  [467,]  0.994243420  0.2232350534
#>  [468,]  0.307103692  0.5508681594
#>  [469,]  0.242845040 -0.4681010515
#>  [470,] -2.447049524  0.6299064599
#>  [471,]  2.944781384  0.9994809718
#>  [472,] -0.045284888  0.7139940752
#>  [473,]  0.785810567  1.0697053327
#>  [474,]  1.340574656  0.1605073635
#>  [475,]  0.654351185  0.0958538551
#>  [476,]  0.155385739 -0.1164811257
#>  [477,]  1.267936682  0.3971252828
#>  [478,] -0.727208376 -0.0289583562
#>  [479,]  0.679861356  0.9107171839
#>  [480,]  2.313629756  0.9996815942
#>  [481,]  2.172280296  0.9026442407
#>  [482,]  2.744134868  1.2051875368
#>  [483,]  1.298257634  0.8693003580
#>  [484,]  0.777557932 -0.0077546046
#>  [485,] -0.537514966 -0.1136100440
#>  [486,] -1.871049140 -0.2482240496
#>  [487,] -1.914344157 -0.8385331167
#>  [488,]  1.413795973  1.7105990419
#>  [489,]  0.427081501  0.5658017600
#>  [490,]  1.426899742  0.4174380291
#>  [491,]  1.294535480  1.0446045977
#>  [492,]  1.151823138  1.3571681564
#>  [493,] -0.137416678 -0.1954735089
#>  [494,]  0.906925377 -0.0791907441
#>  [495,]  2.064662793  1.3944713320
#>  [496,]  1.211576901  0.3925570114
#>  [497,]  2.901547978  0.4006018955
#>  [498,]  0.006133524  0.5826807624
#>  [499,] -0.085483372  0.2411818166
#>  [500,]  0.345103901  1.0149377087
#>  [501,] -0.446506847  0.0616872411
#>  [502,]  1.086964139 -0.1112502281
#>  [503,] -0.178322994  0.3562392225
#>  [504,]  0.461155568  0.6888227951
#>  [505,]  1.153654128  0.3245366701
#>  [506,]  1.153793313  0.9126697069
#>  [507,]  1.215598670  0.6352988336
#>  [508,]  0.387267902  0.4467593767
#>  [509,] -0.725342657 -0.1701587459
#>  [510,]  0.260407147 -0.7129853185
#>  [511,] -2.021508089 -0.1375726756
#>  [512,]  0.296291060  0.6853827288
#>  [513,]  2.803719842  0.6960943670
#>  [514,]  1.091561853  0.0881125694
#>  [515,]  1.110186493 -0.2003996131
#>  [516,] -0.871927688  0.2323688429
#>  [517,]  2.139745000  1.2279460727
#>  [518,]  0.794398410  0.7981896037
#>  [519,]  0.506723519  0.5716533843
#>  [520,]  3.174839223  1.0719106596
#>  [521,] -0.588683235  0.6803877278
#>  [522,]  2.025534416  1.4218412201
#>  [523,] -0.001002710  0.4025354826
#>  [524,] -0.911067607 -0.3438268056
#>  [525,]  0.649350830 -0.0325164274
#>  [526,]  0.566518734  0.7621034869
#>  [527,]  0.273810811  0.1401911323
#>  [528,]  0.068305433  0.8155849887
#>  [529,]  0.606295581  0.9629620275
#>  [530,] -0.125901778  0.3644209601
#>  [531,]  0.752579163  0.3590576789
#>  [532,]  1.371437170  0.8534990174
#>  [533,]  2.338165312  0.8831697681
#>  [534,]  0.039969441 -0.0806298917
#>  [535,] -1.420652517 -0.1002265557
#>  [536,] -0.080624267  0.0095313727
#>  [537,] -0.587791778  0.0449345729
#>  [538,]  0.262229496  0.9015662415
#>  [539,]  1.684511323  0.8508851808
#>  [540,]  1.608944585  1.4971600029
#>  [541,]  1.374845438 -0.5683869777
#>  [542,]  1.355180941  0.2699971241
#>  [543,]  0.776478145  0.3369674438
#>  [544,] -1.349655941 -0.0531777110
#>  [545,]  1.573469227  1.1006917990
#>  [546,]  0.492397308  1.3089376216
#>  [547,]  0.569578564  0.4363262091
#>  [548,]  1.509090150  0.9743025919
#>  [549,]  0.076667780  0.2203699671
#>  [550,]  0.360717671  0.5221651580
#>  [551,]  0.421661974  0.5135428161
#>  [552,]  1.734115184 -0.1497082858
#>  [553,] -0.370906954 -0.2992453246
#>  [554,]  1.508826489  0.1990670527
#>  [555,] -0.322579915  0.6568987773
#>  [556,]  0.282839586  0.6469706580
#>  [557,]  0.332001075  0.5708143467
#>  [558,]  2.243821094  0.7474093919
#>  [559,]  0.271240942  0.4613239308
#>  [560,]  1.328733914  0.5626233269
#>  [561,]  0.075051301  0.4989452424
#>  [562,] -1.038967405  0.0524137605
#>  [563,]  0.041009664  0.5883492503
#>  [564,]  2.234461185  0.9492946034
#>  [565,]  0.621245038  0.2438260638
#>  [566,]  0.014466145 -0.0032698782
#>  [567,] -0.155556122  0.5385925363
#>  [568,] -0.465019958  0.2663384543
#>  [569,] -0.256900394  0.2983075376
#>  [570,]  0.735229162  0.3136664730
#>  [571,]  2.539097417  0.7480180612
#>  [572,]  1.411604438  1.8031555634
#>  [573,] -0.505339218  0.6247184848
#>  [574,] -0.224955400  0.5464961211
#>  [575,]  0.570483826  0.8098889454
#>  [576,] -0.568545720  0.3796589384
#>  [577,]  1.542586036  0.8704969891
#>  [578,]  0.229664671  0.3749038221
#>  [579,]  1.214467324  0.1894567319
#>  [580,] -0.535788352  0.2546626691
#>  [581,]  1.112748756  0.5817450190
#>  [582,]  0.951512765  0.8403510136
#>  [583,]  1.226698679 -0.3907653247
#>  [584,]  1.015838804  1.3835374968
#>  [585,]  0.076768188  0.0935884984
#>  [586,]  0.099328433 -0.2059651455
#>  [587,]  0.087682195 -0.4531624871
#>  [588,]  2.983269369  1.3883437066
#>  [589,]  0.047695758  0.8705024218
#>  [590,] -0.899433806 -0.5639309509
#>  [591,] -0.979140572 -0.0920256850
#>  [592,] -0.534491055 -0.3983801981
#>  [593,] -0.398097836  1.1563110860
#>  [594,] -0.796482387  0.5892666139
#>  [595,] -1.090536445  0.4827284386
#>  [596,]  0.007462293  0.7816252180
#>  [597,] -1.945202954  0.0551518079
#>  [598,]  0.256051007  0.8405428704
#>  [599,] -1.022613442  0.0960679340
#>  [600,]  1.119584625  0.3227183748
#>  [601,] -1.108701000 -0.3290014585
#>  [602,]  1.016458249  0.3848327010
#>  [603,]  1.196466712  0.1419218712
#>  [604,] -0.429755112  0.2027712283
#>  [605,]  0.185558781  0.1662694826
#>  [606,]  1.411848836  0.7673494654
#>  [607,]  0.834340598  0.1727718092
#>  [608,]  2.218810604  0.9132040397
#>  [609,]  0.147690896  0.8309684415
#>  [610,]  0.358961893  0.9427265108
#>  [611,] -0.077934297  1.1660708573
#>  [612,] -1.398317283  0.5307796656
#>  [613,] -1.205607418  0.6046856130
#>  [614,]  1.594875608  0.5845709954
#>  [615,] -0.503463051 -0.3072692891
#>  [616,]  0.292702273  1.1902166025
#>  [617,]  2.429828561  1.0110724101
#>  [618,]  1.482634796  1.1682624460
#>  [619,]  0.862317984  0.1057060520
#>  [620,]  1.818071626 -0.2289078385
#>  [621,] -2.666970843  0.5080368324
#>  [622,] -1.036810678  0.0761963427
#>  [623,] -0.964421055 -0.0695154187
#>  [624,]  1.040783850  0.0288995712
#>  [625,]  1.465271019  0.7381768206
#>  [626,]  1.044829181  0.7146160144
#>  [627,] -0.896836712  0.3680999963
#>  [628,] -0.259435028  1.1924673086
#>  [629,]  2.974929870  0.4485597657
#>  [630,]  0.521934139  0.2373778165
#>  [631,] -0.515084487  0.5535514323
#>  [632,] -0.343073698  0.3701925677
#>  [633,] -0.164944216  0.5422372769
#>  [634,]  1.838284224  1.0486275148
#>  [635,]  0.752165771  0.4669804255
#>  [636,]  0.192513575 -0.3087482607
#>  [637,]  0.907107594  1.2987147288
#>  [638,]  1.232307765  0.4431562274
#>  [639,] -2.340013056  0.5209148926
#>  [640,]  0.720110332  0.3890556429
#>  [641,]  1.067785630  0.2879540038
#>  [642,] -0.210650584  0.9110290301
#>  [643,]  2.564474432  1.3076005287
#>  [644,]  2.697433302  1.0647166399
#>  [645,] -0.145388819  0.6384401603
#>  [646,] -0.403551137  0.4543019250
#>  [647,]  1.412676187  0.6780283912
#>  [648,] -0.422039409 -0.1295197456
#>  [649,] -0.582453443  0.3607446821
#>  [650,]  2.069052371  0.5663995246
#>  [651,]  0.063438875  0.5855788812
#>  [652,]  0.014687836 -0.0847911940
#>  [653,]  0.690990584  1.2395428090
#>  [654,]  0.666846240  1.0090728687
#>  [655,]  0.596822128  1.5433489061
#>  [656,]  1.724348742  0.9312151103
#>  [657,]  1.602129054  1.4230282751
#>  [658,] -1.797173798  0.0027877703
#>  [659,]  1.437838642  1.0564685006
#>  [660,]  0.010034518  0.2928921410
#>  [661,]  0.601857265  0.3006761092
#>  [662,]  1.156413926  1.1104013875
#>  [663,]  2.272876782  0.4040409861
#>  [664,]  0.915540790  0.6943758981
#>  [665,]  2.616469013  0.2475538965
#>  [666,]  0.554740264  0.4606734617
#>  [667,]  1.530878510  0.1027895457
#>  [668,]  1.495828539  0.8427005651
#>  [669,]  0.307419589  0.9509266473
#>  [670,] -0.846303250  0.0371908010
#>  [671,] -0.297734736  0.3019154527
#>  [672,] -0.068813425  0.2230989348
#>  [673,]  1.584337920  1.1914002672
#>  [674,]  1.246149209  0.0320257989
#>  [675,]  1.197110503  1.1085665063
#>  [676,]  0.739026205  0.3784543750
#>  [677,]  0.368519630  0.2354413659
#>  [678,]  1.381918834  0.3864765624
#>  [679,]  1.092754069  0.6673729606
#>  [680,]  1.396097074 -0.3097188889
#>  [681,] -0.906401957  0.5710577515
#>  [682,]  0.351362891  0.1873952696
#>  [683,]  0.230578908  0.4015701175
#>  [684,] -0.593815025 -0.0785964676
#>  [685,]  1.988492161  0.6151024010
#>  [686,]  0.035056192  0.6015343449
#>  [687,]  1.181267604  0.3874444132
#>  [688,]  0.240881186  0.0916218144
#>  [689,]  0.170195254  0.7902335370
#>  [690,]  1.558494282  0.8075420090
#>  [691,]  0.405410733  0.6415759505
#>  [692,] -0.509718978  0.2952202242
#>  [693,]  0.829584747  1.0703826533
#>  [694,] -0.465393644 -0.1619798572
#>  [695,]  1.249360332  0.2548464827
#>  [696,]  1.983534280  1.5687385019
#>  [697,]  2.055114130  0.7734738864
#>  [698,]  2.111009081  0.3768303903
#>  [699,]  1.254794366  0.9709469537
#>  [700,]  0.914021040  1.0872383436
#>  [701,] -1.108838236  0.4399944897
#>  [702,] -1.501970587  0.6392698192
#>  [703,] -0.860776969  0.4035881269
#>  [704,]  0.810231619 -0.0726788069
#>  [705,]  0.646637881  0.7643561469
#>  [706,] -0.394486900  0.2897238395
#>  [707,]  1.142031807 -0.1491141248
#>  [708,] -0.746570458 -0.2573942808
#>  [709,] -0.939467259  0.4209405267
#>  [710,] -0.423466139  0.9973389426
#>  [711,]  0.480738104  0.7150466663
#>  [712,]  0.079701648  0.7387644355
#>  [713,]  0.288579811  0.1513167662
#>  [714,]  2.084306229  0.5436702997
#>  [715,] -0.094042754  0.0316161776
#>  [716,]  0.437283188  0.7226084432
#>  [717,]  0.277262212  0.9917821472
#>  [718,]  1.169402778  0.9363765667
#>  [719,]  1.114878986  0.2945084140
#>  [720,]  0.489888323  0.4544031262
#>  [721,]  1.300338651  1.3900242768
#>  [722,] -0.941705952 -0.1539301705
#>  [723,] -0.849155432  0.1643566433
#>  [724,] -1.095394556  0.0180092322
#>  [725,]  0.698211606  0.0881863546
#>  [726,]  1.335783736  0.0349300242
#>  [727,]  0.738176885  0.5779581644
#>  [728,]  1.916820013  1.1014903174
#>  [729,]  3.082765346  1.3225501851
#>  [730,]  2.037255636  1.0094421752
#>  [731,]  0.097677426  0.9374197867
#>  [732,]  0.072418274  0.8378913535
#>  [733,]  0.368218924  0.7754319752
#>  [734,]  0.611041897  0.5408192786
#>  [735,] -1.524749077 -0.6432048900
#>  [736,] -0.475460646  0.2570378722
#>  [737,]  1.885855594  1.1361079920
#>  [738,]  0.309564064  0.8568034745
#>  [739,] -0.558323815  0.4895721492
#>  [740,] -0.053166923  0.1072462603
#>  [741,]  1.341781832  1.2657772028
#>  [742,]  0.633348239  0.1304087635
#>  [743,]  0.987222088  1.0178780268
#>  [744,]  1.571045755  0.5038029718
#>  [745,]  0.215750728  0.9555678914
#>  [746,]  0.011530238  0.3763105645
#>  [747,] -0.936583394  0.0132187451
#>  [748,]  0.195835858  0.5606357325
#>  [749,] -0.484319038 -0.0204625111
#>  [750,]  0.974034848  0.1958001867
#>  [751,]  0.985425358 -0.1811398728
#>  [752,]  0.239675489 -0.3077803919
#>  [753,] -1.108353992  0.1077782663
#>  [754,] -1.268521418 -0.6428937743
#>  [755,] -0.143540521  0.9531559533
#>  [756,] -0.268375475  0.2316672929
#>  [757,] -1.755741535 -0.2217760135
#>  [758,]  0.423158943  0.7064340841
#>  [759,]  0.066041778  1.1134982454
#>  [760,]  0.339340218 -0.1742334372
#>  [761,]  1.710601500  0.3140515692
#>  [762,]  0.472028201  0.7558230308
#>  [763,] -2.028059645  0.1397157871
#>  [764,]  0.314570197  0.5111154640
#>  [765,]  1.408904537  0.9778410585
#>  [766,]  1.782925973  0.9141569012
#>  [767,]  2.539880269  0.6217635321
#>  [768,]  0.688548753 -0.5893341578
#>  [769,] -0.769599067  0.0790359268
#>  [770,]  0.035050512  0.7882394189
#>  [771,]  1.596959265  1.1044946914
#>  [772,] -0.190187297  0.1331845754
#>  [773,]  1.345926980  1.0170051450
#>  [774,] -0.081968855  0.7889209130
#>  [775,] -0.021732963  0.8556792576
#>  [776,]  2.244677343  1.3099038355
#>  [777,]  2.709037415  0.7212663450
#>  [778,] -0.131659067  1.2792797508
#>  [779,] -0.861768449 -0.5821890572
#>  [780,] -0.438650099  0.3636269702
#>  [781,]  1.005738842  0.4298239545
#>  [782,] -0.660220241  0.5238666693
#>  [783,]  0.163782363  0.7182300138
#>  [784,]  1.555772085  0.6983626096
#>  [785,] -1.082120412 -0.2968993791
#>  [786,]  1.564762102  1.2973173774
#>  [787,]  0.314563664  0.4774467128
#>  [788,]  0.190679546  0.5132129055
#>  [789,]  2.220608597  0.6223281997
#>  [790,] -1.091082521  0.1926059409
#>  [791,] -0.124835117  0.3456710441
#>  [792,] -0.125707197 -0.0467687581
#>  [793,]  1.492660361  0.7296946403
#>  [794,]  0.664399606  0.3839252208
#>  [795,]  1.457215974  1.1360106981
#>  [796,] -0.835307493  0.5998572714
#>  [797,] -0.060861839  0.5346715480
#>  [798,] -0.488206943  0.4004355527
#>  [799,]  0.736475790 -0.1511665463
#>  [800,]  0.983916463  1.5036096916
#>  [801,]  0.858245552  0.6592194905
#>  [802,]  0.752412204  0.5056049534
#>  [803,] -0.258082868  0.6181781298
#>  [804,]  1.876507004  1.0632136050
#>  [805,]  1.917410050  0.9699505549
#>  [806,]  0.859064442  0.8552138885
#>  [807,]  0.352524277  0.6080700362
#>  [808,]  1.314372746  0.5538586809
#>  [809,] -1.087205900 -0.2333047192
#>  [810,]  1.012970352 -0.1649610371
#>  [811,]  2.059764895  0.7490698326
#>  [812,] -0.451580306  0.1329510552
#>  [813,] -0.262571450  0.3867680947
#>  [814,] -0.837297421  0.0396215772
#>  [815,] -1.364807857  0.2053575516
#>  [816,]  0.625037395  0.8812675550
#>  [817,]  2.034010098  0.8702522840
#>  [818,] -0.909552753  0.2432246997
#>  [819,]  2.572679214  1.0632674081
#>  [820,]  0.982302756  0.2934198066
#>  [821,]  2.894824539  0.9588278918
#>  [822,] -0.109185351  1.3991420295
#>  [823,]  2.261251356  0.8802819546
#>  [824,]  0.675550276  0.7339878468
#>  [825,] -0.556357056  0.4179871128
#>  [826,] -0.705589521  0.2355686346
#>  [827,] -0.977016913 -0.4779186193
#>  [828,]  2.310746270  1.0625592054
#>  [829,]  3.478417422  1.6367597237
#>  [830,]  0.374625353 -0.2854091283
#>  [831,] -0.718847347  0.5609621053
#>  [832,]  1.258844364  0.5680936217
#>  [833,]  0.833049172  1.0128982851
#>  [834,]  0.665820041  1.0051704699
#>  [835,]  0.148578365  0.8901364736
#>  [836,]  2.128583866  0.9767350874
#>  [837,] -0.175598086 -0.2469353576
#>  [838,]  3.198982137  0.9064677509
#>  [839,]  0.433736652  1.1064206013
#>  [840,]  0.922493904  1.0493215185
#>  [841,]  0.842614295  0.4455766034
#>  [842,] -0.096144168  0.1627379650
#>  [843,]  0.663733197  0.8306438978
#>  [844,] -0.004492864  1.4674089297
#>  [845,]  0.404479953  0.3537228814
#>  [846,]  0.172848509  0.5203646179
#>  [847,] -1.211447027 -0.0929734257
#>  [848,] -0.080560783  0.1319266407
#>  [849,]  0.468371171  0.9308387487
#>  [850,]  0.749033862  0.1899920199
#>  [851,]  0.352965914 -0.2026200362
#>  [852,] -0.035218647  0.1358819178
#>  [853,]  1.334441133 -0.1495145236
#>  [854,] -0.476518931 -0.0553606372
#>  [855,]  1.274967579  0.7347205332
#>  [856,]  1.998633076  0.5029314201
#>  [857,] -0.135471734 -0.4595837546
#>  [858,]  0.929081671  0.7749267213
#>  [859,]  1.057020076 -0.5878589322
#>  [860,]  1.036636900  0.0655008169
#>  [861,] -1.046549486  1.0135427206
#>  [862,]  1.062865113  0.3334822894
#>  [863,] -0.821835987  0.4290544792
#>  [864,]  0.400482788 -0.0272065119
#>  [865,]  2.161483141  0.2662066953
#>  [866,]  1.956724962  0.7935009155
#>  [867,]  2.906113028  0.9498564925
#>  [868,]  0.708210039  0.0970920636
#>  [869,] -0.277986033  0.3370879116
#>  [870,]  0.962967381  0.9022064842
#>  [871,] -1.561396182  0.4521825823
#>  [872,] -1.781911917  0.0780327194
#>  [873,] -0.132114300  1.0482244664
#>  [874,]  1.126240385  0.6681550794
#>  [875,]  1.496499640  0.4310900110
#>  [876,]  0.110361926  0.1672837984
#>  [877,]  2.203859736  0.6252332676
#>  [878,]  0.873918244  0.8852736350
#>  [879,]  0.779941083 -0.0349777563
#>  [880,]  2.614233937  1.5010344316
#>  [881,]  1.677736467  0.4705613519
#>  [882,] -0.029107153  0.6492516235
#>  [883,]  0.534925053  0.6650932242
#>  [884,]  0.703085115  0.0387579502
#>  [885,]  0.689946025 -0.1438594391
#>  [886,]  1.411712939  0.2183007593
#>  [887,]  1.777855068  0.6471673100
#>  [888,]  1.164158262  1.5721658367
#>  [889,]  2.429126437  1.1769812595
#>  [890,]  0.970297174  0.6383198413
#>  [891,]  2.805581784  0.8475001831
#>  [892,] -2.685003581  0.2936427974
#>  [893,] -1.135429012 -0.1099788173
#>  [894,]  0.353426438  0.9516375814
#>  [895,] -0.282496967  0.1257129766
#>  [896,]  1.892686405  0.9288519331
#>  [897,]  0.631895994  0.7557962055
#>  [898,]  1.068889053  1.1325315506
#>  [899,] -0.802703760 -0.2157869218
#>  [900,]  0.197773729  0.5337110121
#>  [901,]  2.483779273  0.5900259469
#>  [902,]  1.579438789  0.1872072785
#>  [903,]  0.590027417  0.8402183868
#>  [904,]  0.825750484  0.7357786561
#>  [905,]  0.583365162  1.0061976103
#>  [906,]  1.973454323  0.4136891980
#>  [907,]  2.569998829  1.2044322027
#>  [908,]  0.500705803  0.9676209864
#>  [909,]  0.483792583  0.4065345134
#>  [910,]  0.158331527 -0.3498000551
#>  [911,] -0.982713957 -0.0982153754
#>  [912,]  0.876527113  0.9789726748
#>  [913,] -0.248337076  0.2790507364
#>  [914,]  0.777179270  1.3893432305
#>  [915,]  1.218878232  0.0360381322
#>  [916,]  1.299804366  0.5182534464
#>  [917,]  1.484610667  0.7394721517
#>  [918,]  0.287076753  1.1567667658
#>  [919,]  0.257977243  0.6011677471
#>  [920,]  1.240183726  0.3955233941
#>  [921,]  1.355606050  0.5163380870
#>  [922,] -2.533317283 -0.1715276269
#>  [923,]  1.971747524  0.9354524314
#>  [924,]  1.281716798  0.4056190277
#>  [925,]  1.533534494  0.7653119188
#>  [926,]  1.246378483  1.1681864548
#>  [927,]  0.146265949  0.1480497566
#>  [928,]  0.022878304  0.7511369863
#>  [929,]  1.677942500  0.4102678054
#>  [930,]  0.371813332 -0.0263022550
#>  [931,]  0.731750988 -0.0829055309
#>  [932,]  2.134216749  0.6604577592
#>  [933,] -0.863702897  0.7583817654
#>  [934,] -0.618812019 -0.1312354789
#>  [935,]  0.919854384  0.9138254280
#>  [936,]  1.330267755  0.8580527349
#>  [937,]  2.456520741  1.2682534471
#>  [938,]  1.637557804  1.2789220437
#>  [939,] -0.737561901  0.4612033424
#>  [940,]  2.011103941  1.4624869673
#>  [941,]  2.162335561  1.1846875709
#>  [942,] -0.623768603  1.3885351170
#>  [943,] -0.656433018  0.0781143479
#>  [944,]  0.755507172  0.3318836566
#>  [945,] -1.285041242  0.3018232431
#>  [946,]  0.642626758  0.9577270509
#>  [947,] -0.656096815  0.7634283808
#>  [948,]  0.178265942 -0.1828718309
#>  [949,]  1.863490133  1.0975273530
#>  [950,]  1.304237390  0.9892223985
#>  [951,]  2.904677018  1.6480572419
#>  [952,]  0.723349936  0.1833003979
#>  [953,] -0.810505375 -1.1600861487
#>  [954,]  2.218912204  0.4032092667
#>  [955,] -1.417061665  1.3145098599
#>  [956,]  1.490061503  0.0075144706
#>  [957,]  0.890348710  0.3614291599
#>  [958,]  0.174495138  0.1973922586
#>  [959,]  1.102636042  0.9724295608
#>  [960,]  0.717809645  1.0491735578
#>  [961,]  0.721805942  0.2303825416
#>  [962,] -0.366525377  0.1132000051
#>  [963,]  0.653376434  1.2093817072
#>  [964,] -2.027477040  0.4527645898
#>  [965,] -0.572554576 -0.3932169067
#>  [966,] -0.274005541  0.0768347847
#>  [967,]  1.796864362  1.0212882703
#>  [968,]  0.206132340  0.1123761739
#>  [969,]  1.093703077  0.6450669693
#>  [970,] -1.069521579  0.7548066342
#>  [971,]  0.530532780  0.2265888805
#>  [972,]  1.008941020  0.3144825522
#>  [973,]  1.499774559 -0.2748100816
#>  [974,]  1.976496816  1.6757174748
#>  [975,] -1.117133349  0.1581795698
#>  [976,]  0.820725550  0.8694607702
#>  [977,]  0.507569717  0.6684166234
#>  [978,] -0.814633951 -0.1390955249
#>  [979,]  2.015011247  0.7866170783
#>  [980,]  1.240003712  0.2232056019
#>  [981,]  1.256240399  0.8120477147
#>  [982,]  2.544611317  1.0883842338
#>  [983,]  0.748581816  0.6768532797
#>  [984,]  0.841476169  0.7245088955
#>  [985,]  0.971534434 -0.2556058315
#>  [986,] -0.231350940  1.3097234230
#>  [987,]  1.082683368  0.3227959450
#>  [988,]  0.396594547  0.6445591829
#>  [989,]  1.360541409  0.5526406166
#>  [990,]  2.879703407  1.1935730033
#>  [991,]  0.760452794  0.5564814142
#>  [992,] -0.353935292 -0.2395300787
#>  [993,]  1.537038421  0.9548369675
#>  [994,]  1.162962094  0.2450225766
#>  [995,]  1.457426498  1.0430585786
#>  [996,]  2.243080320  0.5581293736
#>  [997,]  0.724916802  0.3735365760
#>  [998,] -1.325977923  0.3557416434
#>  [999,] -0.664227574  0.3792342629
#> [1000,]  1.664340860  1.2770018873