Vector of Expected Values \(\mathbf{v}\)

Derives the vector of expected values \(\mathbf{v}\) using the Reticular Action Model (RAM) notation.

v(A, u, ...)

# S3 method for default
v(A, u, ...)

# S3 method for yac_symbol
v(A, u, exe = TRUE, R = FALSE, format = "ysym", simplify = FALSE, ...)

Arguments

A
t by t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

u
vector of length t or t by 1 matrix. Mean structure parameters.

...
...

exe
Logical. If exe = TRUE, executes the resulting yacas expression. If exe = FALSE, returns the resulting yacas expression as a character string. If exe = FALSE, the arguments str, ysym, simplify, and tex, are ignored.

R
Logical. If R = TRUE, returns symbolic result as an R expression. If R = FALSE, returns symbolic result as "ysym", "str", or "tex" depending of format.

format
Character string. Only used when R = FALSE. If format = "ysym", returns symbolic result as yac_symbol. If format = "str", returns symbolic result as a characetr string. If format = "tex", returns symbolic result as LaTeX math.

simplify
Logical. Simplify symbolic results.

Value

\(\mathbf{v} = \mathbf{E} \mathbf{u}\)

Details

The vector of expected values \(\mathbf{v}\) as a function of Reticular Action Model (RAM) matrices is given by

$$ \mathbf{v} = \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} \\ = \mathbf{E} \mathbf{u} $$

where

\(\mathbf{A}_{t \times t}\) represents asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings,

\(\mathbf{I}_{t \times t}\) represents an identity matrix, and

\(\mathbf{u}_{t \times 1}\) vector of parameters for the mean structure.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Numeric -----------------------------------------------------------
# This is a numerical example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, 1, 1)
u <- c(0.00, 0.50, 0.00)
colnames(A) <- rownames(A) <- c("y", "x", "e")
v(A, u)

#>     v
#> y 0.5
#> x 0.5
#> e 0.0

# Symbolic ----------------------------------------------------------
# This is a symbolic example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

A <- matrixR::ZeroMatrix(3)
A[1, ] <- c(0, "beta", 1)
u <- c("alpha", "mux", 0)
v(Ryacas::ysym(A), u)

#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 

v(Ryacas::ysym(A), u, format = "tex")

#> [1] "\\left( \\begin{array}{c} \\alpha  + \\beta  \\mathrm{ mux } \\\\ \\mathrm{ mux } \\\\ 0 \\end{array} \\right)"

v(Ryacas::ysym(A), u, format = "ysym")

#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 

v(Ryacas::ysym(A), u, R = TRUE)

#> expression(rbind(c(alpha + beta * mux), c(mux), c(0)))

# Assigning values to symbols

alpha <- 0
beta <- 1
mux <- 0.50
v(Ryacas::ysym(A), u)

#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 

v(Ryacas::ysym(A), u, format = "tex")

#> [1] "\\left( \\begin{array}{c} \\alpha  + \\beta  \\mathrm{ mux } \\\\ \\mathrm{ mux } \\\\ 0 \\end{array} \\right)"

v(Ryacas::ysym(A), u, format = "ysym")

#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 

v(Ryacas::ysym(A), u, R = TRUE)

#> expression(rbind(c(alpha + beta * mux), c(mux), c(0)))

eval(v(Ryacas::ysym(A), u, R = TRUE))

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5
#> [3,]  0.0