Equations to Expectations

Converts equations to expectations.

Eq2Expectations(eq, par = FALSE, R = FALSE, format = "ysym", simplify = FALSE)

Arguments

eq
Character string. Equations. See Details.

par
Logical. If par = TRUE, use par.index as labels. If par = FALSE, use par.label as labels.

R
Logical. If R = TRUE, returns symbolic result as an R expression. If R = FALSE, returns symbolic result as "ysym", "str", or "tex" depending of format.

format
Character string. Only used when R = FALSE. If format = "ysym", returns symbolic result as yac_symbol. If format = "str", returns symbolic result as a characetr string. If format = "tex", returns symbolic result as LaTeX math.

simplify
Logical. Simplify symbolic results.

Value

Returns a list with the following elements

par.table

Parameter table.

variables

Variable names.

g.variables

Variable names of observed variables.

h.variables

Variable names of latent variables.

A

t by t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S

t by t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

u

t by 1 matrix \(\mathbf{u}\) of mean structure parameters.

Filter

p by t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

v

t by 1 matrix \(\mathbf{v}\) of expected values.

g

p by 1 matrix \(\mathbf{g}\) of expected values of observed variables.

C

t by t matrix \(\mathbf{C}\) of expected covariances.

M

p by p matrix \(\mathbf{M}\) of expected covariances of observed variables.

Details

The argument eq is a character string that specifies the associations between the variables. See Syntax, Operations, par.label, Line breaks, and Comments below.

Syntax

Each line should follow the syntax below

lhs <space> op <space> rhs <space> par.label <\n> or <;>

lhs

is the variable on the left-hand side,

rhs

is the variable on the right-hand side,

op

is the operation between lhs and rhs,

par.label

is the column of parameter label,

\n or ;

are line breaks. Each line should end with a line break.

Operations

The associations are defined by the following operations

by

left-hand side measured by right-hand side,

on

left-hand side regressed on right-hand side,

with

left-hand side covarying with right-hand side,

on 1

left-hand side regressed on 1 for mean structure.

par.label

Each parameter should be labeled. The par.label should be a number for fixed parameters and a character string for free parameters. Equality contraints can be imposed by using the same par.label.

Line breaks

The characters \n and ; can be used as line breaks. Each line should end with a line break.

Comments

Comments can be written after a hash (#) sign.

References

McArdle, J. J., & McDonald, R. P. (1984). Some algebraic properties of the Reticular Action Model for moment structures. British Journal of Mathematical and Statistical Psychology, 37 (2), 234--251. https://doi.org/10.1111/j.2044-8317.1984.tb00802.x

See also

Author

Ivan Jacob Agaloos Pesigan

Examples

# Numerical ---------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   1
    e   with e   1
    x   with x   0.25
    y   on   1   0
    x   on   1   0.50
"
Eq2Expectations(eq)

#> $par.table
#>   lhs   op rhs par.label par.index par.names
#> 1   e   by   y      1.00      1.00        NA
#> 2   y   on   x      1.00      1.00        NA
#> 3   e with   e      1.00      1.00        NA
#> 4   x with   x      0.25      0.25        NA
#> 5   y   on   1      0.00      0.00        NA
#> 6   x   on   1      0.50      0.50        NA
#> 
#> $variables
#> [1] "y" "x" "e"
#> 
#> $g.variables
#> [1] "y" "x"
#> 
#> $h.variables
#> [1] "e"
#> 
#> $v
#>     v
#> y 0.5
#> x 0.5
#> e 0.0
#> 
#> $g
#>     g
#> y 0.5
#> x 0.5
#> 
#> $C
#>      y    x e
#> y 1.25 0.25 1
#> x 0.25 0.25 0
#> e 1.00 0.00 1
#> 
#> $M
#>      y    x
#> y 1.25 0.25
#> x 0.25 0.25
#> 

# Symbolic ----------------------------------------------------------
eq <- "
  # lhs op   rhs par.label
    e   by   y   1
    y   on   x   beta
    e   with e   sigmae2
    x   with x   sigmax2
    y   on   1   alpha
    x   on   1   mux
"
Eq2Expectations(eq, par = FALSE)

#> $par.table
#>   lhs   op rhs par.label par.index par.names
#> 1   e   by   y         1         1      <NA>
#> 2   y   on   x      beta        p1      beta
#> 3   e with   e   sigmae2        p2   sigmae2
#> 4   x with   x   sigmax2        p3   sigmax2
#> 5   y   on   1     alpha        p4     alpha
#> 6   x   on   1       mux        p5       mux
#> 
#> $variables
#> [1] "y" "x" "e"
#> 
#> $g.variables
#> [1] "y" "x"
#> 
#> $h.variables
#> [1] "e"
#> 
#> $v
#> {{alpha+beta*mux},
#>  {           mux},
#>  {             0}} 
#> 
#> $g
#> {{alpha+beta*mux},
#>  {           mux}} 
#> 
#> $C
#> {{sigmax2*beta^2+sigmae2,           beta*sigmax2,                sigmae2},
#>  {          sigmax2*beta,                sigmax2,                      0},
#>  {               sigmae2,                      0,                sigmae2}} 
#> 
#> $M
#> {{sigmax2*beta^2+sigmae2,           beta*sigmax2},
#>  {          sigmax2*beta,                sigmax2}} 
#> 

Eq2Expectations(eq, par = TRUE)

#> $par.table
#>   lhs   op rhs par.label par.index par.names
#> 1   e   by   y         1         1      <NA>
#> 2   y   on   x        p1        p1      beta
#> 3   e with   e        p2        p2   sigmae2
#> 4   x with   x        p3        p3   sigmax2
#> 5   y   on   1        p4        p4     alpha
#> 6   x   on   1        p5        p5       mux
#> 
#> $variables
#> [1] "y" "x" "e"
#> 
#> $g.variables
#> [1] "y" "x"
#> 
#> $h.variables
#> [1] "e"
#> 
#> $v
#> {{p4+p1*p5},
#>  {      p5},
#>  {       0}} 
#> 
#> $g
#> {{p4+p1*p5},
#>  {      p5}} 
#> 
#> $C
#> {{p3*p1^2+p2,      p1*p3,         p2},
#>  {     p3*p1,         p3,          0},
#>  {        p2,          0,         p2}} 
#> 
#> $M
#> {{p3*p1^2+p2,      p1*p3},
#>  {     p3*p1,         p3}} 
#> 

# Expressions using `par.label`

beta <- 1
sigmae2 <- 1
sigmax2 <- 0.25
alpha <- 0
mux <- 0.50
Exp <- Eq2Expectations(eq, par = FALSE, R = TRUE)
eval(Exp$M)

#>      [,1] [,2]
#> [1,] 1.25 0.25
#> [2,] 0.25 0.25

eval(Exp$g)

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5

# Expressions using `par.index`

p <- c(beta, sigmae2, sigmax2, alpha, mux)
p1 <- p[1]
p2 <- p[2]
p3 <- p[3]
p4 <- p[4]
p5 <- p[5]
Exp <- Eq2Expectations(eq, par = TRUE, R = TRUE)
eval(Exp$M)

#>      [,1] [,2]
#> [1,] 1.25 0.25
#> [2,] 0.25 0.25

eval(Exp$g)

#>      [,1]
#> [1,]  0.5
#> [2,]  0.5