Generate Data from a Multivariate Normal Distribution Using the Reticular Action Model (RAM) Notation

Generates data from a multivariate normal distribution using the Reticular Action Model (RAM) notation with the MASS::mvrnorm() function.

mvn(n, A, Omega, M, filter = NULL, ...)

Arguments

n
Integer. Sample size.

A
t x t numeric matrix \(\mathbf{A}_{t \times t}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

Omega
t x t numeric matrix \(\boldsymbol{\Omega}_{t \times t}\). Symmetric paths (double-headed arrows), such as variances and covariances.

M
t x 1 numeric vector \(\mathbf{M}_{t \times 1}\). Mean structure. Vector of means and intercepts.

filter
j x t numeric matrix \(\mathbf{F}_{j \times t}\). Filter matrix used to select variables. If filter = NULL, the filter matrix \(\mathbf{F}\) is assumed to be a \(t \times t\) identity matrix.

...
Additional arguments to pass to MASS::mvrnorm().

Details

The Multivariate Normal Distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbb{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbb{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{M} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \boldsymbol{\Omega} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} $$

Author

Ivan Jacob Agaloos Pesigan