Generate Data from a Multivariate Normal Distribution Using the Reticular Action Model (RAM) Notation

Generates data from a multivariate normal distribution using the Reticular Action Model (RAM) notation with the MASS::mvrnorm() function.

mvn(n, A, S, u, filter, ...)

Arguments

n
Integer. Sample size.

A
t x t matrix \(\mathbf{A}\). Asymmetric paths (single-headed arrows), such as regression coefficients and factor loadings.

S
t x t numeric matrix \(\mathbf{S}\). Symmetric paths (double-headed arrows), such as variances and covariances.

u
vector of length t or t by 1 matrix. Mean structure parameters.

filter
p x t numeric matrix \(\mathbf{F}\). Filter matrix used to select observed variables.

...
Additional arguments to pass to MASS::mvrnorm().

Details

The Multivariate Normal Distribution is given by $$ \mathbf{X} \sim \mathcal{N}_{k} \left( \boldsymbol{\mu}, \boldsymbol{\Sigma} \right) $$

with location parameter

$$ \boldsymbol{\mu} \in \mathbf{R}^{k} $$

and a positive definite covariance matrix

$$ \boldsymbol{\Sigma} \in \mathbf{R}^{k \times k} . $$

The probability density function is given by

$$ f_{\mathbf{X}} \left( x_1, \cdots, x_k \right) = \frac{ \exp \left[ - \frac{1}{2} \left( \mathbf{x} - \boldsymbol{\mu} \right)^{\mathsf{T}} \boldsymbol{\Sigma}^{-1} \left( \mathbf{x} - \boldsymbol{\mu} \right) \right] }{ \sqrt{ \left( 2 \pi \right)^k | \boldsymbol{\Sigma} | } } $$

In this function, the model-implied mean vector and variance-covariance matrix are used as parameters to generate the data.

$$ \boldsymbol{\mu} \left( \boldsymbol{\theta} \right) = \mathbf{g} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{\mathsf{T}} \mathbf{u} = \mathbf{F} \mathbf{E} \mathbf{u} $$ $$ \boldsymbol{\Sigma} \left( \boldsymbol{\theta} \right) = \mathbf{M} = \mathbf{F} \left( \mathbf{I} - \mathbf{A} \right)^{-1} \mathbf{S} \left[ \left( \mathbf{I} - \mathbf{A} \right)^{-1} \right]^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{E} \mathbf{S} \mathbf{E}^{\mathsf{T}} \mathbf{F}^{\mathsf{T}} \\ = \mathbf{F} \mathbf{C} \mathbf{F}^{\mathsf{T}} $$

Author

Ivan Jacob Agaloos Pesigan

Examples

# This is an example for the model
# y = alpha + beta * x + e
# y = 0 + 1 * x + e

n <- 1000
A <- S <- matrix(
  data = 0,
  nrow = 3,
  ncol = 3
)
A[1, 2] <- 1
A[1, 3] <- 1
diag(S) <- c(0, 0.25, 0.25)
u <- c(-0.50, 0.50, 0.00)
filter <- diag(2)
filter <- cbind(filter, 0)
colnames(filter) <- c("y", "x", "e")
rownames(filter) <- c("y", "x")
mvn(n, A, S, u, filter)

#>                     y             x
#>    [1,]  0.9087946309  1.1839848398
#>    [2,] -0.2107722526  0.4481441204
#>    [3,]  1.7782118419  1.3733558466
#>    [4,] -0.2047686437  0.8398959541
#>    [5,] -0.6389597847  0.5773873600
#>    [6,] -0.8515363954  0.1105910836
#>    [7,]  1.2706134348  1.2475933874
#>    [8,]  0.0559064172  0.7901360221
#>    [9,]  0.4113256032  0.2102457483
#>   [10,]  0.1453581257  0.6998445648
#>   [11,]  0.1573711010  1.0974239485
#>   [12,] -0.2893800630  0.0003229397
#>   [13,] -1.3120526290 -0.5548008133
#>   [14,]  1.3222614110  0.8703707111
#>   [15,] -0.0853251330 -0.1504850284
#>   [16,]  1.2523895312  1.3404710707
#>   [17,]  0.1351702786  1.0845845563
#>   [18,]  0.2393725426  0.1936331063
#>   [19,] -0.6784983159  0.7622478842
#>   [20,]  0.5691448608  0.9857207946
#>   [21,] -0.6681049574  0.8606009528
#>   [22,] -0.1949258510  0.2568720998
#>   [23,]  0.6419698844  1.4687570034
#>   [24,] -0.5396250173  0.2378106001
#>   [25,] -1.1509914945 -0.5437669064
#>   [26,]  0.2422959681  0.2579094851
#>   [27,]  0.8800638198  0.5161478238
#>   [28,] -0.2576149346  0.9413747254
#>   [29,]  0.1202961065  1.5776541867
#>   [30,]  1.0837691424  1.0737684458
#>   [31,] -0.4772567462 -0.1671583086
#>   [32,] -0.2285490504  0.5982120860
#>   [33,] -0.2304610846  0.4979004965
#>   [34,] -1.1567080076 -0.1267919639
#>   [35,] -0.0412660516  0.3943609924
#>   [36,] -0.1484021283  0.9463198881
#>   [37,]  1.4322381516  1.1219123678
#>   [38,] -0.1940821318  1.2437334166
#>   [39,]  0.2050843185  0.6505103656
#>   [40,] -0.8108198065  0.1695160620
#>   [41,] -0.0937699844  0.5439504690
#>   [42,]  0.5332267583  0.6207306474
#>   [43,]  0.0009711627  0.5753166994
#>   [44,]  0.2112272835  0.5452203025
#>   [45,] -0.2881326940  0.2817361119
#>   [46,] -1.8517066140 -0.7440274162
#>   [47,]  0.2238387285  0.0662168975
#>   [48,] -0.3053346835  0.1050390498
#>   [49,] -0.0910252412  0.4653987245
#>   [50,]  1.2913267630  1.3524247483
#>   [51,] -0.5500383828  0.0633661121
#>   [52,]  0.2532376360  0.4645557524
#>   [53,]  0.0477000429  0.7399552927
#>   [54,]  0.1398717216  0.2441715205
#>   [55,] -0.1350988032  0.6731051319
#>   [56,] -0.5855879755  0.6014036609
#>   [57,]  1.6841505817  1.2744900450
#>   [58,] -1.9143783848 -0.5305017468
#>   [59,]  0.2071010050  0.7207658904
#>   [60,] -0.2068158828  0.5063143886
#>   [61,] -0.5423809528 -0.2756575010
#>   [62,]  0.3853541141  0.8999495223
#>   [63,] -0.5816242323 -0.2731087340
#>   [64,]  0.2001761038  0.5545039294
#>   [65,]  0.5884043256  1.3600254070
#>   [66,]  0.4606820212  1.2562805198
#>   [67,] -1.0801554869  0.6087736010
#>   [68,] -0.1062440756  0.4692862529
#>   [69,] -0.2234400939  0.1096112454
#>   [70,]  1.0924444990  1.3475731872
#>   [71,]  1.2985853623  0.6620752053
#>   [72,]  0.1335689460 -0.1533815118
#>   [73,] -0.8508609326 -0.1812026012
#>   [74,] -0.2549223915  0.5482349668
#>   [75,] -1.0427023479  0.1584829164
#>   [76,] -0.2958488580  0.1724792438
#>   [77,] -0.6571117292  0.6295398752
#>   [78,]  0.0114134184  0.6507063052
#>   [79,] -0.1397278031  0.4611239192
#>   [80,] -0.2626750610  1.0640165787
#>   [81,] -1.4861144684 -0.0566761898
#>   [82,] -0.8252905032 -0.1626729246
#>   [83,] -0.5255238721  0.1980442085
#>   [84,] -0.3958067631  0.2844875653
#>   [85,]  0.3720609343  0.7416739398
#>   [86,] -0.8810656060  0.2166566186
#>   [87,]  1.6163884901  1.9045976689
#>   [88,] -0.0196154805  0.7713266001
#>   [89,] -0.5872834438  0.7825568415
#>   [90,]  0.1933493137  0.7749165631
#>   [91,] -0.6218823941  0.8534692933
#>   [92,] -0.2847283594 -0.6752455924
#>   [93,] -0.4758039918 -0.3439357518
#>   [94,]  0.0317439053  0.5072716083
#>   [95,] -0.2142142085  0.0984996071
#>   [96,] -1.3577179942  0.1225322263
#>   [97,]  0.2172955291  0.6937142079
#>   [98,] -0.7505221720  0.2580907283
#>   [99,] -0.7028470788 -0.3891569363
#>  [100,]  0.2448118458  0.5603681728
#>  [101,] -0.0106116586  1.1130304472
#>  [102,]  0.4015573665  1.0589231096
#>  [103,]  0.7036960623  0.9875466828
#>  [104,]  0.4947748167  0.9236499102
#>  [105,]  1.1806180056  1.2923499252
#>  [106,]  0.2744113445  1.1186216761
#>  [107,]  0.0753354990  1.2376123522
#>  [108,]  0.3539737406  0.7530983991
#>  [109,] -0.1510178311  0.3948392650
#>  [110,] -0.5984227091 -0.0374190171
#>  [111,]  0.3681732357  0.2257205395
#>  [112,]  0.8954009560  1.2046820868
#>  [113,] -0.0672971812  0.2110411492
#>  [114,]  0.2367234080  0.7968329314
#>  [115,] -0.4755156173  0.3944152149
#>  [116,] -1.5748050013 -0.2247983042
#>  [117,] -0.4362660759  0.5521034238
#>  [118,]  1.1671063716  1.2034249942
#>  [119,] -0.1414898064  0.3451454334
#>  [120,] -0.8372000387  0.2036898052
#>  [121,] -1.5222503494 -0.1752137979
#>  [122,]  0.1317918893 -0.4048820031
#>  [123,] -1.1754304597  0.2601048477
#>  [124,] -0.3985107808  0.4883855688
#>  [125,] -0.2255247660  0.6993519958
#>  [126,] -0.1402887794  0.6926839467
#>  [127,] -0.3550328459  0.1455933768
#>  [128,]  0.2384973059  0.5182809250
#>  [129,]  0.2036863207  0.9833434448
#>  [130,] -0.0973505581  0.3618289554
#>  [131,]  0.7757936409  1.0085594465
#>  [132,] -0.8134145779  0.5139881278
#>  [133,]  0.1237910021  0.1739483722
#>  [134,]  1.0733790416  0.7716047957
#>  [135,]  0.5787569343  1.0176217546
#>  [136,] -0.4390949978  0.5113108082
#>  [137,]  0.2930947731  1.3417840241
#>  [138,]  0.0148206831  0.9175059768
#>  [139,] -0.4436119579 -0.1594251049
#>  [140,] -0.2865477103  0.8600730516
#>  [141,] -0.1401138447  0.9770421055
#>  [142,]  0.6956609962  0.6474963076
#>  [143,] -1.1178155828 -0.5789816225
#>  [144,] -0.5572828465  0.2222723958
#>  [145,] -0.4212454294 -0.3264088722
#>  [146,] -0.9442810552 -0.0061329566
#>  [147,]  0.9217494880  0.7316357910
#>  [148,] -0.2874025213  0.1731393530
#>  [149,]  0.8306255548  1.4782854806
#>  [150,] -1.1819219947  0.0537433492
#>  [151,] -0.6651075130  0.9157766626
#>  [152,] -0.1045647261  0.4812914465
#>  [153,]  1.0184355408  0.6033576949
#>  [154,]  0.1496308242  1.1165892293
#>  [155,] -0.4859885005 -0.3333443518
#>  [156,] -0.5871321469  0.4071518719
#>  [157,] -0.0759191989  0.5635951331
#>  [158,]  0.3001141889  0.5628192980
#>  [159,] -0.6374409307  0.3267278895
#>  [160,] -0.6016054792  0.2355559191
#>  [161,]  1.2857477991  1.3435093179
#>  [162,] -0.5116285420 -0.1121915754
#>  [163,]  0.1683476277  0.7031891786
#>  [164,]  0.1697896255 -0.1795433754
#>  [165,]  1.4096326483  0.9746040017
#>  [166,]  0.6632428003  0.6398562409
#>  [167,]  0.5835066231  1.2994075197
#>  [168,] -0.1970051393  0.2591572499
#>  [169,]  0.1099650963  0.7620090002
#>  [170,] -0.1476654294 -0.0576832020
#>  [171,]  0.0755209976  0.5361646189
#>  [172,]  0.6060642522  1.0183064162
#>  [173,] -0.9873533937  0.1422077392
#>  [174,] -0.9036616645  0.4835364191
#>  [175,] -0.2943932865 -0.2066022084
#>  [176,] -0.7692945843  0.1506112449
#>  [177,]  0.3195843027  0.7121389595
#>  [178,]  1.0802015953  0.7147356782
#>  [179,] -0.0083945560  0.9838845352
#>  [180,] -1.6153556819 -0.2898628717
#>  [181,]  0.6021767307  1.1286203960
#>  [182,]  0.4872736898  1.4758933974
#>  [183,]  1.1128223188  1.2774792952
#>  [184,] -1.1046523658 -0.0608130651
#>  [185,] -0.4218991901  0.3298434116
#>  [186,] -1.2969865015 -0.4685130826
#>  [187,]  0.0313207070  0.6864894629
#>  [188,] -1.7552284232 -0.6061994155
#>  [189,] -0.8514994454  0.2435064023
#>  [190,] -1.1378861897  0.5827119226
#>  [191,] -0.8112520615  0.0832232557
#>  [192,]  0.4153500693  0.4388719504
#>  [193,]  0.4084071900  1.1060519303
#>  [194,]  0.3150823881  0.8809978265
#>  [195,] -1.3638536234 -0.0074312469
#>  [196,] -0.3343081693  0.8362681108
#>  [197,] -0.2860604555  0.3834826474
#>  [198,] -0.5265981161 -0.4002372490
#>  [199,] -0.6893938977 -0.2946447080
#>  [200,] -0.6074921526  0.5410318385
#>  [201,]  0.1476815459  0.9217942519
#>  [202,] -0.4311398401 -0.8959357990
#>  [203,] -0.1249959381  0.8112860502
#>  [204,]  0.0735540193  0.7997885536
#>  [205,]  1.6515438203  1.5927993882
#>  [206,]  0.0757410973  0.2766747511
#>  [207,]  0.7478349479  0.9804090177
#>  [208,]  0.6490805979  0.4241265547
#>  [209,]  1.4341620805  1.3551146728
#>  [210,] -1.9252966346 -0.4610918545
#>  [211,]  0.9001809693  0.8183742085
#>  [212,]  0.2375745801  0.6397840462
#>  [213,] -0.4280517044 -0.0174853711
#>  [214,]  1.2007163353  0.5123188002
#>  [215,] -0.3823535967  0.2843910545
#>  [216,] -0.0126293352  0.4048585974
#>  [217,] -0.3879208491  0.2682037983
#>  [218,] -0.4208518960  0.5417079544
#>  [219,]  1.2273053369  0.7490377952
#>  [220,] -0.7653292103  0.2899613819
#>  [221,]  0.4494184870  0.2943942890
#>  [222,] -0.0797995251  0.7454151525
#>  [223,]  0.2639529938  0.0110468868
#>  [224,]  0.0260063385  0.2666411879
#>  [225,]  0.7016325922  0.9011643301
#>  [226,] -0.7960819170 -0.1099045492
#>  [227,] -0.2226907233  0.3364617387
#>  [228,]  0.3631705789  0.6397919832
#>  [229,] -0.1650510123 -0.3236006654
#>  [230,]  0.8976072741  1.4004664545
#>  [231,] -0.2541509226  0.5458647930
#>  [232,]  0.9450264887  0.9911284179
#>  [233,] -0.5462777456  0.2343371675
#>  [234,]  1.1449720811  1.0518648880
#>  [235,] -0.0374009782  0.4511761764
#>  [236,]  0.1904775041  1.1759437412
#>  [237,]  0.6554102987  0.7401463652
#>  [238,] -0.3226426387 -0.0170479898
#>  [239,] -1.0045730433  0.3505704816
#>  [240,]  0.9853487897  1.5009361081
#>  [241,]  0.4369415378  1.1823030939
#>  [242,] -0.9031096828 -0.0663676076
#>  [243,] -0.7606272241  0.0377027690
#>  [244,] -0.8737114740  0.0612866304
#>  [245,]  0.9810550412  1.1151188521
#>  [246,] -0.8176740801 -0.3050565412
#>  [247,] -0.1240258245  0.6958676718
#>  [248,]  0.1407136793  0.3921759829
#>  [249,] -0.5592903582  0.7256626172
#>  [250,] -0.2489784200  0.7044636084
#>  [251,]  0.8935870598  0.3317141460
#>  [252,]  0.0457934274  1.2008544819
#>  [253,]  0.6736490213  1.2468378947
#>  [254,]  0.2002985468  1.3326933911
#>  [255,] -1.1319096851  0.0911587832
#>  [256,]  0.2665412595  1.3438169086
#>  [257,] -0.3213578933  0.5740509508
#>  [258,]  0.2343647080  0.8595162895
#>  [259,]  0.2956767595  0.9523216266
#>  [260,] -0.8921667604 -0.3033223905
#>  [261,] -0.1126684866  0.4519685988
#>  [262,]  0.6835905097  1.5995816978
#>  [263,] -0.1999658749  0.8394067592
#>  [264,] -0.1434072943  1.0586355516
#>  [265,]  0.5411091130  1.0191811869
#>  [266,]  1.1645093504  1.8506643293
#>  [267,] -1.2964169692 -0.3159354448
#>  [268,]  0.6460812736  0.9444110151
#>  [269,] -0.0289373473  0.7047114495
#>  [270,] -0.1390368067  0.3557064665
#>  [271,] -0.1495169130  0.6482096866
#>  [272,]  1.4441881879  1.5141993725
#>  [273,]  0.1823246527  0.3863606262
#>  [274,]  0.0897600310  0.1819613097
#>  [275,] -0.1897506185  0.7948873398
#>  [276,] -1.1432851275 -0.5396765051
#>  [277,] -1.2884308316 -0.7372598752
#>  [278,] -0.5998556013  0.3784484007
#>  [279,] -0.7056544104  0.4615067612
#>  [280,]  0.2713857044  0.5351773141
#>  [281,] -0.8602128262  0.3345428758
#>  [282,]  0.7226476377  0.7450065452
#>  [283,] -0.3168698730  0.1457534301
#>  [284,] -0.2048107832  0.3351408842
#>  [285,] -0.2935641951  0.4107430200
#>  [286,] -0.9361829564  0.2761383174
#>  [287,]  0.7119750233  0.7968167774
#>  [288,]  0.0149670948  0.1405650681
#>  [289,] -0.8714896796 -0.2679956339
#>  [290,]  0.1855488076  0.7902801648
#>  [291,] -0.1281633108  0.9752639759
#>  [292,]  0.2528146754  0.4321669872
#>  [293,]  0.8944582941  0.6062410433
#>  [294,] -0.3052509878  0.2541463544
#>  [295,] -0.8756645031 -0.5520740819
#>  [296,]  1.1070904020  1.0400731263
#>  [297,]  1.0275254826  1.0164655898
#>  [298,] -0.8404749225  0.5106954195
#>  [299,] -0.5593757143  0.4446478926
#>  [300,] -1.3772919275 -0.5219553661
#>  [301,]  0.0582234510  0.9538116031
#>  [302,]  0.6884310846  1.0241370916
#>  [303,] -0.6669487956 -0.0783660990
#>  [304,] -0.8159046074  0.0019035518
#>  [305,] -0.2245189837  0.5579853049
#>  [306,]  0.4178397879  0.4400813436
#>  [307,] -0.3036753520  0.0431930980
#>  [308,] -1.3853722236 -0.2963355168
#>  [309,] -1.4190091181 -0.1041921208
#>  [310,]  0.8896526021  1.5051475032
#>  [311,]  0.2032234552  1.0020338590
#>  [312,]  0.8000302221  1.0051400556
#>  [313,] -0.4294390116  0.3344712685
#>  [314,]  1.3309247007  1.1462061325
#>  [315,] -0.1217230041  0.0916578237
#>  [316,] -0.1890751338  0.7410922709
#>  [317,] -1.0540926824  0.3902412790
#>  [318,]  0.3099616958  0.3938324782
#>  [319,]  0.8915985000  0.6828974316
#>  [320,]  0.0343718734  0.1384984747
#>  [321,] -0.4592955954  0.2420876831
#>  [322,] -0.0819440435  0.1033600398
#>  [323,] -1.0792150576 -0.0274817919
#>  [324,] -0.1174366594  0.5791803532
#>  [325,] -1.5397996015 -0.2808135766
#>  [326,]  0.7413189333  1.5721542266
#>  [327,] -0.4343105838  0.5750534259
#>  [328,] -1.0150485904 -0.0027281407
#>  [329,] -0.1897045184  0.2525370678
#>  [330,] -0.4426268110  0.2089359581
#>  [331,] -0.5404762333 -0.2445917943
#>  [332,]  1.2765928774  1.4806514448
#>  [333,] -0.7536883154 -0.1400209304
#>  [334,]  0.3467886180  0.1993794930
#>  [335,] -0.2160483434  0.3955694568
#>  [336,] -0.9355237050  0.0650113639
#>  [337,]  0.8666662419  0.8448181888
#>  [338,]  0.6490735046  1.1902898355
#>  [339,] -0.0358781209  0.3887896052
#>  [340,] -0.7135699751  0.3420940037
#>  [341,]  0.2579204398  0.4432848516
#>  [342,] -1.2558899623 -0.6790108804
#>  [343,] -0.0768608968  0.7950509048
#>  [344,]  1.0723225313  0.9083169252
#>  [345,]  0.7171215731  1.0974776668
#>  [346,] -1.3588102852  0.0754781014
#>  [347,]  0.0930306146  0.4726939565
#>  [348,]  0.2701896506  0.7357305980
#>  [349,] -0.1881947898  0.9879587330
#>  [350,] -0.2300138027 -0.3380717553
#>  [351,]  0.2657123388  0.9609735309
#>  [352,]  0.1624092686  0.4610052487
#>  [353,] -0.4368129559  0.3966426641
#>  [354,] -1.3093059434 -0.0492008861
#>  [355,] -0.9271684124 -0.2290642337
#>  [356,] -1.2348609005  0.1617878773
#>  [357,]  0.4790565869  0.3157965622
#>  [358,] -1.0354966229 -0.6361878082
#>  [359,]  0.3389615609  0.8002072133
#>  [360,]  0.5268035947  0.9798715885
#>  [361,]  0.4150740043  0.3574983968
#>  [362,] -0.6224639152 -0.1284309175
#>  [363,] -0.5545104175  0.1460487780
#>  [364,] -0.8563528521 -0.8896675030
#>  [365,]  0.4923940695 -0.1350262454
#>  [366,] -0.4791685430 -0.2365292806
#>  [367,]  0.9758101279  1.5576076835
#>  [368,]  0.6376154773  0.7486628857
#>  [369,] -0.6054961051 -0.2137695954
#>  [370,]  0.0428315542  0.6981787287
#>  [371,]  0.1341192729  0.0078270554
#>  [372,]  0.5777638932  0.6399307923
#>  [373,]  0.7366924601  0.7859858428
#>  [374,]  0.5503474831  1.1105407864
#>  [375,]  0.3888132604  0.3919025942
#>  [376,] -0.7682374426 -0.0302455162
#>  [377,] -0.4037784941  0.5299407058
#>  [378,]  0.5504042908  0.3341023585
#>  [379,]  0.2409489347  0.3151886675
#>  [380,] -0.6791183581 -0.2888334598
#>  [381,] -0.1422722555  0.2178532187
#>  [382,]  0.0784131590  0.9072390728
#>  [383,]  0.2607902465  0.2296864358
#>  [384,] -0.1080055174  0.6212571328
#>  [385,] -0.3078972448  0.4039998180
#>  [386,] -0.0676331996  0.5773768517
#>  [387,]  0.0067086644  0.4774703037
#>  [388,] -0.6808662357  0.1692430372
#>  [389,]  0.2480412661  1.1523826114
#>  [390,] -0.2387674926  0.3671263777
#>  [391,]  0.3521124803  0.5644847519
#>  [392,] -0.6101985642  0.0496419668
#>  [393,] -1.5422148864  0.1633998356
#>  [394,]  0.5893964803  0.6309423993
#>  [395,] -0.1675431964  0.7579945623
#>  [396,] -1.2433300314 -0.6185461123
#>  [397,]  0.6542966667  1.5059683168
#>  [398,] -0.6236270661 -0.2744357854
#>  [399,]  0.1221341619  0.6150892068
#>  [400,]  0.5686973671  0.1615544607
#>  [401,] -1.3794977658  0.0604740850
#>  [402,]  0.5362486767  0.9330340586
#>  [403,]  0.5164847395  1.1440165141
#>  [404,] -0.6780195833  0.0313010000
#>  [405,]  0.9078613613  1.3332394468
#>  [406,]  0.8995680827  0.8677168167
#>  [407,] -0.4649686731  0.2825041393
#>  [408,] -1.4413825475 -0.4012891819
#>  [409,]  0.4806247452  0.6667766641
#>  [410,]  1.1202327531  1.2028749796
#>  [411,]  0.1935337629  0.2029230201
#>  [412,]  0.5124947917  0.6810250756
#>  [413,]  0.5132965011  0.6997748400
#>  [414,]  0.1945483217  0.9216869886
#>  [415,] -0.6814937302 -0.4277544730
#>  [416,] -0.4691948783  0.2796038952
#>  [417,] -0.3097216465  0.2096710779
#>  [418,]  1.0295543628  1.5292352246
#>  [419,] -0.9933996007  0.7322491327
#>  [420,] -0.2427851414  0.5496138474
#>  [421,] -0.4115040633  0.2720427274
#>  [422,]  0.2900805118  0.3042789025
#>  [423,] -0.4749752179  0.1282955273
#>  [424,]  0.6061208886  1.0541828723
#>  [425,]  2.2682395454  1.3525327489
#>  [426,] -0.5889595921  0.3465053065
#>  [427,]  0.7524468291  0.3566347679
#>  [428,]  1.1409011822  1.5687430722
#>  [429,]  0.1461316635  0.1461415304
#>  [430,] -0.4273795719 -0.1554466082
#>  [431,] -0.4117364757  0.4678350642
#>  [432,]  0.6678012152  0.7285936444
#>  [433,] -0.3149121934  0.1382439416
#>  [434,] -0.5517877749 -0.3802605072
#>  [435,]  0.6636918108  0.5897922012
#>  [436,]  0.2495105984  0.8491745313
#>  [437,]  1.0113175371  0.6580238914
#>  [438,]  0.4256056187  0.5494902587
#>  [439,] -0.2919573917  0.7948943280
#>  [440,]  1.2682102415  1.1687394618
#>  [441,]  1.1627271556  0.7848015055
#>  [442,] -0.3397150805 -0.0416428904
#>  [443,]  0.7261211230  1.2345722881
#>  [444,]  0.1867199748  0.7646782422
#>  [445,] -0.2406077472  0.1782433647
#>  [446,]  0.2523746851  0.5884382411
#>  [447,]  0.6706363948  1.3957074337
#>  [448,]  0.8993276101  0.6299263150
#>  [449,] -1.0043742999 -0.2048321267
#>  [450,] -0.1602561239  0.2254701273
#>  [451,] -1.3115635209 -0.1161246137
#>  [452,] -0.0421841712 -0.0267083925
#>  [453,]  0.4206059166  1.2331719369
#>  [454,]  1.1934649079  1.0069763782
#>  [455,]  0.2202523955  0.2729774726
#>  [456,]  1.3800562938  1.4756161850
#>  [457,] -0.0164324664  0.5210977144
#>  [458,] -0.0344005797  1.1031363086
#>  [459,] -0.9398593676  0.2566581342
#>  [460,]  0.2844095757  0.3801896586
#>  [461,] -0.6151805334 -0.0714333930
#>  [462,]  0.3458916789  0.3461577737
#>  [463,] -1.0063279836 -0.4239612919
#>  [464,]  0.6001621485  0.7201143010
#>  [465,]  0.5284179752  1.0667959573
#>  [466,]  0.2787793780  0.4729277613
#>  [467,]  0.3207466903  0.8941691787
#>  [468,]  0.0563236793  0.5970450095
#>  [469,] -0.9364804407  0.1997559070
#>  [470,] -0.5397001831  0.3867981540
#>  [471,]  0.7008993108  0.3347581861
#>  [472,]  0.6925159040  0.6224620412
#>  [473,]  0.6471212655  0.7770936419
#>  [474,]  1.4270732163  1.5290621388
#>  [475,]  1.0930535865  0.8485968467
#>  [476,]  0.5328048334  0.3967565156
#>  [477,]  0.3925644745  0.7598877710
#>  [478,]  0.2429294800  0.3642712818
#>  [479,] -0.1449167541 -0.0133862917
#>  [480,]  0.8734704684  1.1385802389
#>  [481,]  0.1293180245  0.6938826866
#>  [482,] -0.4954786349  0.2977819804
#>  [483,] -0.6458820045  0.3926897284
#>  [484,] -0.6323661486  0.1435255066
#>  [485,] -1.0369526727 -0.1139730286
#>  [486,]  0.5698715628  0.5926460614
#>  [487,] -0.3494917001  0.2386956522
#>  [488,] -0.3032986004 -0.1584643081
#>  [489,] -1.4860886267 -0.0135900276
#>  [490,]  1.2809213034  1.5975486716
#>  [491,] -0.2888059168  0.8679749858
#>  [492,]  0.1363291509  0.6874346983
#>  [493,]  0.3148229603  0.6789981993
#>  [494,]  0.9693911894  1.1023376166
#>  [495,] -0.1305948067  1.4906053647
#>  [496,]  0.1889550574  0.6093850791
#>  [497,]  0.7308001234  0.9874222740
#>  [498,] -0.1535280626  0.1911080325
#>  [499,]  0.1618389892  0.7546853284
#>  [500,] -0.8354513244 -0.2468906861
#>  [501,]  0.3956774108  0.9547035610
#>  [502,] -0.4831788921  0.2628114796
#>  [503,] -0.5121514115  0.8806750439
#>  [504,] -0.3854895905  0.8191092204
#>  [505,]  0.6471793758  1.3049395689
#>  [506,]  0.2869819652  0.0969255562
#>  [507,] -0.5767177579  0.3759916230
#>  [508,] -0.2308340692 -0.2117969579
#>  [509,] -0.9177308187  0.4592477888
#>  [510,] -0.9160494113  0.2213809720
#>  [511,]  0.7706489663  1.1595506081
#>  [512,] -1.7617963282 -0.7348221054
#>  [513,]  0.2658019136  0.3113941939
#>  [514,]  0.1972701041  0.8324945473
#>  [515,]  0.3971429852  0.1626954435
#>  [516,]  0.6525065219  0.8211815886
#>  [517,] -0.7059534433  0.2507295752
#>  [518,] -0.1886463785  0.6826687213
#>  [519,]  0.9206425660  0.9475049468
#>  [520,] -0.7210557135  0.0883255695
#>  [521,]  0.5523131073  0.7308912575
#>  [522,]  0.0534430778  0.7061311327
#>  [523,] -0.4031633661  0.3384929492
#>  [524,] -0.2447576846  0.0493393006
#>  [525,] -0.6310174745  0.6751305046
#>  [526,]  0.4690813022  0.7559418592
#>  [527,] -0.0991316464  0.5720354531
#>  [528,]  1.9114237143  1.7283353202
#>  [529,]  0.6940593138  0.7808118520
#>  [530,]  0.4435547597  0.9862710573
#>  [531,]  0.4256031659  0.8333120082
#>  [532,] -0.4748112328  0.3044690793
#>  [533,]  0.3610501352  0.6963833361
#>  [534,] -0.1280236980  0.2911033800
#>  [535,] -0.2322908352  0.1564815413
#>  [536,] -0.4920080355  0.1819720433
#>  [537,] -0.6839283896  0.6620214602
#>  [538,]  0.2546762537  1.0387281012
#>  [539,]  0.1814722454 -0.1332889597
#>  [540,] -0.7475650562 -0.0261148915
#>  [541,] -1.0646982389 -0.5679013396
#>  [542,]  0.1283583958  0.4213062180
#>  [543,] -0.8520799581 -0.2259898115
#>  [544,]  0.5369743130  0.5966799110
#>  [545,]  0.2917972621  0.3612200964
#>  [546,]  0.5275162414  0.6985881816
#>  [547,]  0.3141438533  0.6760317097
#>  [548,] -0.2816050742  0.0223547352
#>  [549,] -0.4753261825  0.3085749326
#>  [550,]  0.3499732535  1.0038897643
#>  [551,]  0.5395619938  0.8328668996
#>  [552,]  0.4630052598  1.2175485576
#>  [553,] -1.1748802359 -0.3577717691
#>  [554,] -0.3294237862  0.4959989662
#>  [555,] -0.0657363453  0.2076439407
#>  [556,]  0.5447641164  0.8587855862
#>  [557,] -0.2666199704  0.7687989185
#>  [558,]  0.2101000403  0.6734081182
#>  [559,] -1.2449440414 -0.0121824615
#>  [560,]  0.1804872078  1.1988347941
#>  [561,] -0.3234268818  0.1199839284
#>  [562,]  0.3146633904  0.2223104926
#>  [563,]  0.9148393703  1.6524429226
#>  [564,] -1.1879582565  0.2200241640
#>  [565,]  2.0774441449  1.2091878887
#>  [566,]  0.9140135436  0.6098836001
#>  [567,]  0.4528536084  0.8559463629
#>  [568,]  0.3783017529  0.9659459166
#>  [569,] -0.3053621861  0.1663658160
#>  [570,]  0.4357137543  0.2817854998
#>  [571,]  0.5717110935  0.8669985831
#>  [572,]  1.0207738079  0.9333731785
#>  [573,]  0.8453587101  0.3902890798
#>  [574,]  0.3838772849  0.8551147287
#>  [575,] -0.6342316467  0.2705912932
#>  [576,] -0.1407845640  0.2618395292
#>  [577,] -1.4541467703  0.0720385214
#>  [578,]  0.7261615997  0.7008115643
#>  [579,]  0.1126799399  0.3891256981
#>  [580,]  0.5581482746  0.3219672583
#>  [581,]  0.5586123950  0.4066427772
#>  [582,]  0.1042200444  0.4777659953
#>  [583,]  1.6687383479  1.6398962310
#>  [584,]  0.1100936753  0.0650485942
#>  [585,] -0.3549387971  0.5349507180
#>  [586,] -0.6843334240 -0.5989642570
#>  [587,] -0.2340247630 -0.2999493578
#>  [588,] -1.0237797907  0.3598809258
#>  [589,]  0.0431088183  0.6409827346
#>  [590,]  0.1887612652  0.9869306284
#>  [591,] -1.1453403910  0.0145534159
#>  [592,]  0.1791029380  0.7157020729
#>  [593,]  0.2817605220  0.1266938541
#>  [594,]  0.4475527325  0.6430197127
#>  [595,]  0.5086965499  0.8216698762
#>  [596,] -0.0841564938  0.8039659152
#>  [597,]  0.1583655995  1.1067865170
#>  [598,]  0.1053759363  0.0401935174
#>  [599,] -0.5841512088  0.2928134020
#>  [600,]  1.1997394963  1.5080299298
#>  [601,] -0.2775793529  0.5858186772
#>  [602,] -0.1931967579 -0.1552608650
#>  [603,] -0.1503334206  0.1001230981
#>  [604,] -1.1501651754 -0.6809121638
#>  [605,]  0.3968608676  0.6369601300
#>  [606,]  0.6979634148  1.0787379042
#>  [607,]  0.6303400444  0.9399963762
#>  [608,] -0.3643219442  0.3556981986
#>  [609,]  0.8310922909  0.3789538012
#>  [610,] -0.1745353978  0.4218121500
#>  [611,]  0.8614478177  0.9873069600
#>  [612,]  1.7523005525  1.4404931811
#>  [613,]  1.0863738804  1.0339468817
#>  [614,]  0.3243970387  0.6401215490
#>  [615,]  0.8360362834  0.5977063514
#>  [616,]  0.0859773492  0.5477792897
#>  [617,] -0.9162473287 -0.0740608341
#>  [618,]  0.4578402988  1.0831508635
#>  [619,] -0.4510048463  0.2040449331
#>  [620,]  0.9306691825  1.1816626727
#>  [621,] -0.6090303368 -0.3150995172
#>  [622,]  0.9719107898  1.0890487190
#>  [623,] -0.7288781942 -0.0160047436
#>  [624,] -0.4678123402  0.1828458520
#>  [625,]  0.6855545427  0.7208470283
#>  [626,]  0.7240328938  1.0080903342
#>  [627,] -0.0115392137  0.5756871814
#>  [628,] -0.5180482065  0.0916749955
#>  [629,]  0.1928106783  0.9564547342
#>  [630,] -0.4062340300 -0.3015701374
#>  [631,] -0.0119377291  0.7874526085
#>  [632,]  0.9743266370  0.6258284746
#>  [633,]  0.7835151870  0.6879670354
#>  [634,] -0.1794483545  0.3454903988
#>  [635,] -0.4602282544 -0.1047656165
#>  [636,]  0.4984417719  1.4614856422
#>  [637,]  0.6274124424  1.2354313691
#>  [638,]  0.9120847216  1.2359112275
#>  [639,]  0.0378148161  1.1992229635
#>  [640,] -0.5374177530  0.0756992491
#>  [641,] -0.5795582983  0.2196992543
#>  [642,] -0.2688589988  1.1948548121
#>  [643,] -0.4211405062  0.2385372870
#>  [644,]  0.6395113639  0.6519404400
#>  [645,] -0.5746971602  0.0624201276
#>  [646,] -0.1052489489  0.6499613876
#>  [647,] -0.2209401941  0.5107268860
#>  [648,]  0.7320757732  0.4387005287
#>  [649,]  0.8805590454  0.9564397003
#>  [650,] -0.5132661950  0.7044733597
#>  [651,]  0.4657376319  0.9021096595
#>  [652,]  0.0334982255  0.6952704362
#>  [653,] -0.1063924488  0.1304587129
#>  [654,]  0.1904226518  0.6367046627
#>  [655,]  0.0826272927  0.2052476256
#>  [656,] -0.2528275379 -0.1991695649
#>  [657,]  0.3645953174  0.8503756181
#>  [658,]  0.8666808324  0.8010465633
#>  [659,] -0.6173415409  0.6759018997
#>  [660,] -0.4091326160  0.0292983452
#>  [661,]  0.7593764632  1.1525395568
#>  [662,] -0.7690352587  0.1725286454
#>  [663,] -0.3950059206  0.4233174302
#>  [664,] -0.1841624830 -0.4184312577
#>  [665,]  0.3312196244  0.1645323061
#>  [666,]  0.5825246505  0.9890172443
#>  [667,]  0.3934350891  0.4248566182
#>  [668,] -0.1407732435  0.4909904730
#>  [669,]  0.0439691844 -0.2069235253
#>  [670,] -2.0059888881 -0.0718364043
#>  [671,]  1.5283218690  1.3817317585
#>  [672,] -0.4136387529  0.5216221820
#>  [673,]  0.0643375722  0.9460791411
#>  [674,]  0.6397874826  0.4602196232
#>  [675,]  0.1939117482  0.2634044985
#>  [676,] -0.0933249034  0.0093937976
#>  [677,]  0.5381510734  0.6027054042
#>  [678,] -0.7049577212 -0.1292481877
#>  [679,]  0.0288155116  0.8157356847
#>  [680,]  1.1048876248  1.2407365455
#>  [681,]  1.0317657368  1.1440352099
#>  [682,]  1.3477267077  1.4748579923
#>  [683,]  0.4529094294  0.9262300447
#>  [684,]  0.2997289025  0.2214516969
#>  [685,] -0.5587788207 -0.1436175302
#>  [686,] -1.4232401329 -0.5321066567
#>  [687,] -1.3202861932 -0.9377790774
#>  [688,]  0.3422947987  1.5164255166
#>  [689,] -0.0636289370  0.5278360904
#>  [690,]  0.6402446671  0.6518768293
#>  [691,]  0.4112133128  1.0429954047
#>  [692,]  0.2455876275  1.2207579645
#>  [693,] -0.2720795021 -0.1090685149
#>  [694,]  0.4024848288  0.2024583176
#>  [695,]  0.8495978399  1.4498988474
#>  [696,]  0.5013652790  0.5870384842
#>  [697,]  1.6332334507  0.9703241774
#>  [698,] -0.3497836842  0.4450320403
#>  [699,] -0.3348806803  0.1954614226
#>  [700,] -0.2190510344  0.8107948955
#>  [701,] -0.5369263827 -0.0056745023
#>  [702,]  0.5304272209  0.2212100532
#>  [703,] -0.4228870103  0.2518847062
#>  [704,] -0.0682796974  0.6179803027
#>  [705,]  0.4777193126  0.5284571265
#>  [706,]  0.3463021353  0.9230198841
#>  [707,]  0.4497844422  0.7507739439
#>  [708,] -0.0637180251  0.4390774407
#>  [709,] -0.6721327917 -0.2235511011
#>  [710,]  0.1105079909 -0.3672698790
#>  [711,] -1.5489134637 -0.4915231023
#>  [712,] -0.1781049494  0.5788078114
#>  [713,]  1.5015342170  1.1466715172
#>  [714,]  0.4889325842  0.3559747634
#>  [715,]  0.5659396575  0.1665995038
#>  [716,] -0.8604729253  0.0136952048
#>  [717,]  0.9372006958  1.3549791994
#>  [718,]  0.1308112349  0.7658611530
#>  [719,] -0.0115119103  0.5495699759
#>  [720,]  1.6664535881  1.4817615667
#>  [721,] -0.7706468382  0.3775707945
#>  [722,]  0.8172296335  1.4595097554
#>  [723,] -0.3142891066  0.3225912024
#>  [724,] -0.7578875170 -0.3815805385
#>  [725,]  0.2192618758  0.1761729817
#>  [726,] -0.0139859979  0.6906984054
#>  [727,] -0.0712766122  0.2080554336
#>  [728,] -0.3601564679  0.6151710066
#>  [729,] -0.0322162127  0.8343327839
#>  [730,] -0.3895512817  0.2690949229
#>  [731,]  0.2009509144  0.4619314345
#>  [732,]  0.5055330421  0.9319936250
#>  [733,]  1.1473994128  1.1680643537
#>  [734,] -0.1787650899  0.0076356675
#>  [735,] -1.1541981389 -0.3321151731
#>  [736,] -0.2798224802  0.0411521094
#>  [737,] -0.6279571855 -0.0485048049
#>  [738,] -0.2492942442  0.7162117231
#>  [739,]  0.7161340401  1.0002457188
#>  [740,]  0.5209308874  1.4168828128
#>  [741,]  0.8257627516 -0.0210743856
#>  [742,]  0.6251030219  0.5369336521
#>  [743,]  0.2219223655  0.4524568291
#>  [744,] -1.1170926290 -0.2846785315
#>  [745,]  0.5857862795  1.1429913252
#>  [746,] -0.1859839918  1.0409518399
#>  [747,]  0.0609126119  0.4728445624
#>  [748,]  0.5708609673  1.0438112682
#>  [749,] -0.2214527100  0.2177585093
#>  [750,] -0.0983897068  0.4837243644
#>  [751,] -0.0555790111  0.4915678821
#>  [752,]  0.9731488224  0.3401189766
#>  [753,] -0.4055054580 -0.2308907782
#>  [754,]  0.7440320350  0.5237085027
#>  [755,] -0.5868870150  0.4213164389
#>  [756,] -0.1785392724  0.5500323699
#>  [757,] -0.1285316740  0.5099380063
#>  [758,]  1.1144683299  1.0558975162
#>  [759,] -0.1448080090  0.4229032237
#>  [760,]  0.5419286083  0.7273195414
#>  [761,] -0.2848284022  0.4042710294
#>  [762,] -0.9322873940 -0.1443735505
#>  [763,] -0.3276555707  0.4566331196
#>  [764,]  1.0630466065  1.1892332939
#>  [765,]  0.1386158776  0.3552645141
#>  [766,] -0.2131714457  0.0538276319
#>  [767,] -0.4483899688  0.3793018552
#>  [768,] -0.5951067576  0.1274700475
#>  [769,] -0.4626444142  0.1954525098
#>  [770,]  0.1994619619  0.4276025097
#>  [771,]  1.3124096068  1.1223316170
#>  [772,]  0.3201284052  1.5780242768
#>  [773,] -0.7022899508  0.3588630198
#>  [774,] -0.4967117151  0.3690855908
#>  [775,] -0.0220112872  0.7236404867
#>  [776,] -0.6898931807  0.1803386750
#>  [777,]  0.6165423297  0.9816662609
#>  [778,] -0.1533740959  0.3556341156
#>  [779,]  0.5487188367  0.4514409931
#>  [780,] -0.6399688548  0.1038133986
#>  [781,]  0.3927656196  0.6918510130
#>  [782,]  0.2267791706  0.8292757244
#>  [783,]  0.6866654751  0.0649512190
#>  [784,]  0.1484701989  1.2080400342
#>  [785,] -0.1930361558  0.1327333666
#>  [786,] -0.1109200526 -0.0631687026
#>  [787,] -0.0634575805 -0.2315978984
#>  [788,]  1.4671880429  1.6511949861
#>  [789,] -0.3862617697  0.6474022771
#>  [790,] -0.7008665429 -0.5266294908
#>  [791,] -0.8598566924 -0.2278887895
#>  [792,] -0.4930737771 -0.3339709899
#>  [793,] -0.7492179635  0.7394460797
#>  [794,] -0.8896674463  0.2699795901
#>  [795,] -1.0631022451  0.1327591232
#>  [796,] -0.3933776517  0.5787851600
#>  [797,] -1.5408209275 -0.3451772416
#>  [798,] -0.2397936603  0.6738944490
#>  [799,] -0.9310781923 -0.1114325834
#>  [800,]  0.4552713781  0.5196192043
#>  [801,] -0.8937790759 -0.4158275635
#>  [802,]  0.3722029163  0.5382270024
#>  [803,]  0.5472727772  0.4155285800
#>  [804,] -0.5572363158  0.0927133153
#>  [805,] -0.1363091696  0.2058155691
#>  [806,]  0.5519056372  0.8832390719
#>  [807,]  0.2974529391  0.3552494868
#>  [808,]  1.0606179733  1.1615234315
#>  [809,] -0.3103429253  0.6432416694
#>  [810,] -0.1936078956  0.7654528927
#>  [811,] -0.5366280841  0.8175838769
#>  [812,] -1.2803124888  0.0961709786
#>  [813,] -1.1675646535  0.1888398312
#>  [814,]  0.7155542361  0.8015535769
#>  [815,] -0.4926325781 -0.2659138616
#>  [816,] -0.2933966673  0.9166581917
#>  [817,]  1.1802890892  1.2743605799
#>  [818,]  0.5097434348  1.1680078969
#>  [819,]  0.3312170997  0.3165163430
#>  [820,]  1.0471780738  0.3057635325
#>  [821,] -2.1262657169 -0.2027649377
#>  [822,] -0.9361585650 -0.1279374506
#>  [823,] -0.8550159887 -0.2094970597
#>  [824,]  0.4681100932  0.3048991701
#>  [825,]  0.5942655284  0.8756150300
#>  [826,]  0.3174929259  0.7657963077
#>  [827,] -0.9075328148  0.0991766035
#>  [828,] -0.6642849018  0.7947063575
#>  [829,]  1.6717358150  1.0189039848
#>  [830,]  0.0734379730  0.3287323061
#>  [831,] -0.6929138392  0.3089436012
#>  [832,] -0.5365252057  0.2244055172
#>  [833,] -0.4555026830  0.3796475798
#>  [834,]  0.7750715077  1.1672799750
#>  [835,]  0.1765413432  0.5342357767
#>  [836,] -0.0254265559 -0.1112808812
#>  [837,]  0.0944980335  1.1268261539
#>  [838,]  0.5039576368  0.6256170324
#>  [839,] -1.9098153874 -0.1210160886
#>  [840,]  0.1724624122  0.4747944293
#>  [841,]  0.4282971057  0.4847159479
#>  [842,] -0.5686277893  0.6168203542
#>  [843,]  1.2043065821  1.5033854210
#>  [844,]  1.3478085922  1.3701844624
#>  [845,] -0.4638961424  0.4485551556
#>  [846,] -0.5959021291  0.2673046229
#>  [847,]  0.5724334403  0.8235056750
#>  [848,] -0.4777579444 -0.1284689629
#>  [849,] -0.6949934087  0.1645407447
#>  [850,]  1.0377049433  0.8953929313
#>  [851,] -0.3119901255  0.4597900234
#>  [852,] -0.1947940108 -0.0008089577
#>  [853,] -0.0372464207  1.0388071909
#>  [854,] -0.0019082938  0.8788044154
#>  [855,] -0.1683496499  1.2215498095
#>  [856,]  0.7248954744  1.0630405913
#>  [857,]  0.5329352489  1.3656297391
#>  [858,] -1.4298109959 -0.3472037802
#>  [859,]  0.5046911468  1.0829975138
#>  [860,] -0.2823681121  0.2515082121
#>  [861,]  0.1128981072  0.3890654460
#>  [862,]  0.3038459483  1.0562483148
#>  [863,]  1.2107389879  0.8320560366
#>  [864,]  0.2352894639  0.7233090618
#>  [865,]  1.4762192404  0.8039108641
#>  [866,]  0.0455145666  0.4858592512
#>  [867,]  0.7803532851  0.4640545693
#>  [868,]  0.5913919159  0.9525645585
#>  [869,] -0.2300171305  0.7594285019
#>  [870,] -0.7996403923 -0.1115044073
#>  [871,] -0.4908436779  0.1887419363
#>  [872,] -0.3196546852  0.1870585700
#>  [873,]  0.5727941903  1.2062707291
#>  [874,]  0.6051743184  0.3529173977
#>  [875,]  0.3315563338  1.0641174703
#>  [876,]  0.1875220809  0.4719126003
#>  [877,] -0.0290426044  0.2931287686
#>  [878,]  0.6169937511  0.6210492092
#>  [879,]  0.3602057857  0.7448210344
#>  [880,]  0.7821788393  0.1571974537
#>  [881,] -0.9593321103  0.2331859508
#>  [882,] -0.0298082612  0.2570621039
#>  [883,] -0.1587235721  0.3737268399
#>  [884,] -0.6043753219 -0.1327187850
#>  [885,]  0.9727732175  0.9100500035
#>  [886,] -0.3345975576  0.4641467132
#>  [887,]  0.4821763961  0.5768314843
#>  [888,] -0.0825060462  0.1681108568
#>  [889,] -0.2861379410  0.6209479923
#>  [890,]  0.6412910668  0.9429919683
#>  [891,] -0.0951097539  0.5738212318
#>  [892,] -0.6315499323  0.1368495229
#>  [893,]  0.0935507313  0.9563217057
#>  [894,] -0.4995825476 -0.1599381694
#>  [895,]  0.5575041858  0.5031080854
#>  [896,]  0.7562078553  1.5486599319
#>  [897,]  0.9820516521  1.0311859507
#>  [898,]  1.1082393073  0.7776078958
#>  [899,]  0.4010245132  0.9846979718
#>  [900,]  0.1464232714  0.9865093756
#>  [901,] -1.0658238582  0.0999999139
#>  [902,] -1.3741043745  0.1457708028
#>  [903,] -0.8912785914  0.1310459689
#>  [904,]  0.3361645709  0.1852052765
#>  [905,]  0.0392558497  0.7101247215
#>  [906,] -0.5530208750  0.1589291120
#>  [907,]  0.5758343598  0.2081236839
#>  [908,] -0.6668663753 -0.2868171577
#>  [909,] -0.9479457572  0.1250906287
#>  [910,] -0.7306882866  0.6271317990
#>  [911,] -0.0610071691  0.6399505983
#>  [912,] -0.3353340569  0.5661864838
#>  [913,] -0.0638570327  0.2188211846
#>  [914,]  1.0530199516  0.8835565701
#>  [915,] -0.2937621874  0.0529665821
#>  [916,] -0.0918484779  0.6353063778
#>  [917,] -0.2593828818  0.7800918098
#>  [918,]  0.3514722681  0.9424133117
#>  [919,]  0.4584226787  0.4996431745
#>  [920,]  0.0034126521  0.4671516476
#>  [921,]  0.3378680103  1.2760075983
#>  [922,] -0.8209025224 -0.2610449454
#>  [923,] -0.8299888415 -0.0268367344
#>  [924,] -0.9624467912 -0.1800703042
#>  [925,]  0.2250487179  0.2680684709
#>  [926,]  0.6646535823  0.3749085007
#>  [927,]  0.1423419362  0.6055538971
#>  [928,]  0.8159344343  1.2203025572
#>  [929,]  1.5486438522  1.6293073270
#>  [930,]  0.9173076965  1.1854850103
#>  [931,] -0.3676962250  0.7034680509
#>  [932,] -0.3623853449  0.6310542302
#>  [933,] -0.1499898952  0.6552982415
#>  [934,]  0.0653617009  0.5522121276
#>  [935,] -1.1026145877 -0.7196328113
#>  [936,] -0.6000309108  0.1188964185
#>  [937,]  0.7874221234  1.2366009449
#>  [938,] -0.2075319916  0.6967682773
#>  [939,] -0.7076134592  0.2563563859
#>  [940,] -0.2832532465  0.1128408977
#>  [941,]  0.3934514314  1.2019271042
#>  [942,]  0.1720958310  0.2818882341
#>  [943,]  0.2110374652  0.9563493124
#>  [944,]  0.7176289639  0.7420442224
#>  [945,] -0.2925482860  0.7420441625
#>  [946,] -0.3000176792  0.3078014449
#>  [947,] -0.8548418399 -0.1477726054
#>  [948,] -0.2175980716  0.4726625160
#>  [949,] -0.5439223286 -0.0692372944
#>  [950,]  0.3860133895  0.4019339761
#>  [951,]  0.4779407350  0.1516218924
#>  [952,]  0.0059941959 -0.1000858904
#>  [953,] -0.9912132113 -0.1227492233
#>  [954,] -0.9308014158 -0.6621298621
#>  [955,] -0.5330288571  0.6600862196
#>  [956,] -0.4554409210  0.1481829685
#>  [957,] -1.3518034008 -0.4885812105
#>  [958,] -0.0977066124  0.6212980108
#>  [959,] -0.4282904029  0.8145111260
#>  [960,]  0.0429893217  0.0117858411
#>  [961,]  0.8536755075  0.6459607253
#>  [962,] -0.0759678222  0.6653566217
#>  [963,] -1.6153119661 -0.3069773191
#>  [964,] -0.1268755864  0.4659931156
#>  [965,]  0.5028631899  1.0237827597
#>  [966,]  0.7680046073  1.0646958639
#>  [967,]  1.3411661307  1.0378125553
#>  [968,]  0.3700648714 -0.1885867852
#>  [969,] -0.7575425172 -0.0662822670
#>  [970,] -0.3763498917  0.5893910143
#>  [971,]  0.6006935233  1.1507949151
#>  [972,] -0.3809692916  0.0996021613
#>  [973,]  0.4518592043  1.0359726178
#>  [974,] -0.4550012565  0.5636818484
#>  [975,] -0.4295214113  0.6219338623
#>  [976,]  0.9892651384  1.4334217233
#>  [977,]  1.4323906885  1.1423857590
#>  [978,] -0.5979820331  0.8815134877
#>  [979,] -0.6715172169 -0.5304551711
#>  [980,] -0.5991716919  0.1986296477
#>  [981,]  0.3549517694  0.5660111204
#>  [982,] -0.7836361475  0.2565771158
#>  [983,] -0.2743393621  0.5712125945
#>  [984,]  0.6638782174  0.8691434989
#>  [985,] -0.8831265189 -0.3883492340
#>  [986,]  0.5359785462  1.2729448006
#>  [987,] -0.1193514075  0.4434059697
#>  [988,] -0.2104529642  0.4396973302
#>  [989,]  1.1268659188  0.9668000296
#>  [990,] -0.9985952014 -0.0619830711
#>  [991,] -0.3846431351  0.2567556096
#>  [992,] -0.2974759365 -0.0066960159
#>  [993,]  0.6145355305  0.8760494534
#>  [994,]  0.1362377180  0.4588955405
#>  [995,]  0.4999036808  1.1406889454
#>  [996,] -0.9180802622  0.2684024616
#>  [997,] -0.3839903529  0.3978458617
#>  [998,] -0.6406460827  0.2122403483
#>  [999,]  0.3042380485  0.1160617955
#> [1000,]  0.1002070827  1.2814488586
#> attr(,"n")
#> [1] 1000
#> attr(,"A")
#>      [,1] [,2] [,3]
#> [1,]    0    1    1
#> [2,]    0    0    0
#> [3,]    0    0    0
#> attr(,"S")
#>      [,1] [,2] [,3]
#> [1,]    0 0.00 0.00
#> [2,]    0 0.25 0.00
#> [3,]    0 0.00 0.25
#> attr(,"u")
#> [1] -0.5  0.5  0.0
#> attr(,"filter")
#>   y x e
#> y 1 0 0
#> x 0 1 0
#> attr(,"v")
#>        v
#> [1,] 0.0
#> [2,] 0.5
#> [3,] 0.0
#> attr(,"g")
#>     v
#> y 0.0
#> x 0.5
#> attr(,"C")
#>      [,1] [,2] [,3]
#> [1,] 0.50 0.25 0.25
#> [2,] 0.25 0.25 0.00
#> [3,] 0.25 0.00 0.25
#> attr(,"M")
#>      y    x
#> y 0.50 0.25
#> x 0.25 0.25